This page docuements the Korg API for users of the package. Low-level functions that only people digging into the guts of Korg will be interested in can be found on the developer documentation.

Top-level functions

If you are synthesize a spectrum Korg, these are the functions you will call. These functions are exported, so if you do using Korg, you can call them unquallified (i.e. synthesize instead of Korg.synthesize).

Korg.synthesize — Function

synthesize(atm, linelist, λs; metallicity=0, abundances=Dict(), vmic=0, ... )

Solve the transfer equation in the model atmosphere atm with the transitions in linelist at the wavelengths λs [Å] to get the resultant astrophysical flux at each wavelength.

For efficiency reasons, λs must be an AbstractRange, such as 6000:0.01:6500. It can't be an arbitrary list of wavelengths.

Optional arguments:

metallicity, i.e. [metals/H] is log_10 solar relative

abundances is a Dict mapping atomic symbols to $A(X)$ format abundances, i.e. $A(x) = \log_{10}(n_X/n_\mathrm{H}) + 12$, where $n_X$ is the number density of $X$. These override metallicity.

vmic (default: 0) is the microturbulent velocity, $\xi$, in km/s.

line_buffer (default: 10): the farthest (in Å) any line can be from the provide wavelenth range before it is discarded. If the edge of your window is near a strong line, you may have to turn this up.

cntm_step (default 1): the distance (in Å) between point at which the continuum opacity is calculated.

hydrogen_lines (default: true): whether or not to include H lines in the synthesis.

mu_grid: the range of (surface) μ values at which to calculate the surface flux when doing transfer in spherical geometry (when atm is a ShellAtmosphere).

ionization_energies, a Dict mapping Species to their first three ionization energies, defaults to Korg.ionization_energies.

partition_funcs, a Dict mapping Species to partition functions. Defaults to data from Barklem & Collet 2016, Korg.partition_funcs.

equilibrium_constants, a Dict mapping Species representing diatomic molecules to their molecular equilbrium constants in partial pressure form. Defaults to data from Barklem and Collet 2016, Korg.equilibrium_constants.

Korg.read_linelist — Function

read_linelist(filename; format="vald")

Parse a linelist file.

Pass format="kurucz" for a Kurucz linelist (format=kurucz_vac if it uses vacuum wavelengths; Be warned that Korg will not assume that wavelengths are vacuum below 2000 Å),format="vald" for a VALD linelist, and format="moog" for a MOOG linelist (doesn't support broadening parameters or dissociation energies).

VALD linelists (the default and preferred format) can be either "short" or "long" format, "extract all" or "extract stellar". Air wavelengths will automatically be converted into vacuum wavelengths, and energy levels will be automatically converted from cm$^{-1}$ to eV.

Korg.read_model_atmosphere — Function

read_model_atmosphere(filename; truncate_at_10000K=true)

Parse the provided model atmosphere file in "Kurucz" format. Returns either a PlanarAtmosphere or a ShellAtmosphere.

When truncate_at_10000K is true, layers with temperatures greater than 10,000 Kelvin will be elided. These are typically at the deepest level, where optical depth is very large, and have minimal impact on the surface spectrum. They are ignored by default because Korg's default partition and molecular equillibrium functions are tabulated only up to that tempurature.

Secondary functions

You don't need use these to synthesize spectra, but they might be relevant depending on what you are doing.

Postprocessing

These are used to transform observational or synthetic spectra.

Korg.rectify — Function

rectify(flux, wls; bandwidth=50, q=0.95, wl_step=1.0)

Rectify the spectrum with flux vector flux and wavelengths wls by dividing out a moving q-quantile with window size bandwidth. wl_step controls the size of the grid that the moving quantile is calculated on and interpolated from. Setting wl_step to 0 results in the exact calculation with no interpolation, but note that this is very slow.

Experiments on real spectra show an agreement between the interpolated rectified spectrum and the "exact" one (with default values) at the 3 × 10^-4 level.

Korg.constant_R_LSF — Function

constant_R_LSF(flux, wls, R)

Applies a gaussian line spread function the the spectrum with flux vector flux and wavelength vector wls with constant spectral resolution, $R = \lambda/\Delta\lambda.$

Warning

This is a naive, slow implementation. Do not use it when performance matters.

constant_R_LSF will have weird behavior if your wavelength grid is not locally linearly-spaced. It is intended to be run on a fine wavelength grid, then downsampled to the observational (or otherwise desired) grid.

Line absorption

These functions can be used to directly compute line opacities.

Korg.line_absorption — Function

line_absorption(linelist, λs, temp, nₑ, n_densities, partition_fns, ξ
               ; α_cntm=nothing, cutoff_threshold=1e-3, window_size=20.0*1e-8)

Calculate the opacity coefficient, α, in units of cm^-1 from all lines in linelist, at wavelengths λs [cm^-1].

other arguments:

temp the temerature in K

n_densities, a Dict mapping species to absolute number density [cm^-3].

partition_fns, a Dict containing the partition function of each species

ξ is the microturbulent velocity in cm/s (n.b. NOT km/s)

The three keyword arguments specify how the size of the window over which each line is calculated is chosen.

If α_cntm is passed as a callable returning the continuum opacity as a function of wavelength, the line window will extend to 4 doppler widths or the wavelength at which the Lorentz wings of the line are at cutoff_threshold * α_cntm[line.wl], whichever is greater. cutoff_threshold defaults to 1e-3.

if α_cntm is not passed, defaults to window_size, which is 2e-7 (in cm, i.e. 20 Å) unless otherwise specified

Korg.line_profile — Function

line_profile(λ₀, invΔλ_D, Δλ_L, line_amplitude, λ)

A voigt profile centered on λ₀ with doppler width invΔλ_D = 1/ΔλD and lorentz width `ΔλLevaluated atλ` (cm). Note that this returns values in units of cm^-1.

Korg.hydrogen_line_absorption — Function

hydrogen_line_absorption(λs, T, nₑ, nH_I, UH_I, hline_stark_profiles, ξ; 
                              stark_window_size=3e-7, self_window_size=1e-6)

Calculate contribution to the the absorption coefficient, α, from hydrogen lines in units of cm^-1, at wavelengths λs.

Arguments:

T: temperature [K]

nₑ: electron number density [cm^-3]

nH_I: neutral hydrogen number density [cm^-3]

hline_stark_profiles: (returned by setuphlinestark_profiles`)

ξ: microturbulent velocity [cm/s]. This is only applied to Hα-Hγ. Other hydrogen lines profiles are dominated by stark broadening, and the stark broadened profiles are pre-convolved with a doppler profile.

stark_window_size: the max distance from each line center [cm] at which to calculate the line absorption for lines besides Hα-Hγ (those dominated by stark broadening).

self_window_size: the max distance from each line center [cm] at which to calculate the line absorption for Hα-Hγ (those dominated by self-broadening).

Korg.setup_hydrogen_stark_profiles — Function

load_hydrogen_stark_profiles()

Load the (doppler-convolved) Stark-broadening profiles from disk.

Korg.voigt — Function

voigt(α, v)

The voigt function, $H$. Approximation from Hunger 1965.

Continuum absorption

These function can be used to directly compute continuum opacities.

Korg.total_continuum_opacity — Function

total_continuum_opacity(νs, T, nₑ, ρ, number_densities, partition_funcs)

The total continuum opacity, κ, at many frequencies, ν.

νs are frequencies in Hz

T is temperature in K

nₑ is the electron number density in cm^-3

ρ is the density in g cm^-3

number_densities is a Dict mapping each species to its number density

partition_funcs is a Dict mapping each species to its partition function (e.g. Korg.partition_funcs)

Missing docstring.

Missing docstring for Korg.ContinuumOpacity.H_I_bf. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Korg.ContinuumOpacity.H_I_ff. Check Documenter's build log for details.

Korg.ContinuumOpacity.Hminus_bf — Function

Hminus_bf(nH_I_div_partition, ne, ν, ρ, T, [ion_energy_H⁻])

Compute the H⁻ bound-free opacity κ

Arguments

nH_I_div_partition: the total number density of H I divided by its partition function.

ne: the electron number density

ν: frequency in Hz

ρ: mass density in g/cm³

T: temperature in K

ion_energy_H⁻: Specifies the ionization energy of the single state of H⁻ in eV. This is roughly 0.7552 eV.

Notes

This function assumes that n(H⁻) ≪ n(H I) + n(H II). The number density of n(H⁻) should not be pre-computed (instead it's computed internally by this function).

This function is adapted from equation 8.12 from Grey (2005). This equation gives the absorption coefficient per H I atom (uncorrected by stimulated emission) is given by:

$8.316 \times 10^{-10} \alpha_{bf}(H^-) P_e \theta^{2.5} 10^{\chi_{\mathrm{H}^-} \theta} U(H⁻,T)/ U(H I,T)$

where:

$\alpha_{bf}(\mathrm{H}^-)$ is the photo dissociation cross-section. We estimate this by linearly interpolating data from Wishart (1979).

$\chi_{\mathrm{H}^-}$ is the ionization enrgy of H⁻.

θ = log10(e)/(k*T) or θ = 5040/T in units of eV⁻¹

U(H⁻,T) is the partition function of H⁻ at temperature T. This is always 1

U(H I,T) is the partition function of H I at temperature T. This is 2 at low to intermediate T.

Eqn 8.12 of Grey (2005) implicitly assumes that (U(H⁻,T)/ U(H I,T)) is always equal to 0.5.

This expression is simplicitly a rewritten form of: αff_H⁻ * n(H⁻)/n(H I) where n(H⁻)/n(H I) has been replaced with the expanded form of the saha equation, in which n₀ = n(H⁻) and n₁ = n(H I).

Combining 8.18, and 8.19 of Gray (2005), indicate that the opacity contribution of H⁻ bound-free absorption (with stimulated emission correction) is given by:

$\kappa_\nu = \alpha_{bf}(H^-) \frac{n(H^-)}{n(H I)} (1 - \exp \left( \frac{-h\nu}{k T}\right) \frac{n(H I)}{n(H I) + n(H II)} \frac{n(H I) + n(H II)}{\rho} )$

This can be rewritten as: $\kappa_\nu = \alpha_{bf}(H^-) n(H⁻) (1 - \exp \left( \frac{-h\nu}{k T}\right) / \rho$

Wishart (1979) expects the tabulated data to have better than 1% percent accuracy. Mathisen (1984) suggests that this data has better than 3% accuracy.

Korg.ContinuumOpacity.Hminus_ff — Function

Hminus_ff(nH_I_div_partition, ne, ν, ρ, T)

Compute the H⁻ free-free opacity κ

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + H I -> e⁻ + H I.

Arguments

nH_I_div_partition::Flt: the total number density of H I divided by its partition function.

ne: the number density of free electrons.

ν: frequency in Hz

ρ: mass density in g/cm³

T: temperature in K

Notes

This is taken from equation 8.13 of Gray (2005). This uses a polynomial to Bell & Berrington (1987) tabulated values for α_ff(H⁻), which implicitly includes the correction for stimulated emission. The polynomial fits α_ff(H⁻) in the range 2520 K ≤ T ≤ 10080 K and 2604 Å ≤ λ ≤ 113918 Å. According to the book, the polynomial fit to the tabulated data typically has 1% precision. We find that at worst, the discrepancy never exceeds 2.25%.

From equations 8.13, 8.18, and 8.19 of Gray (2005), the free-free opacity from H⁻ is given by:

                              n(H I)          n(H I) + n(H II)
   κ_ν = α_ff(H⁻) * Pₑ * ------------------ * -----------------
                          n(H I) + n(H II)            ρ

This can be rewritten as: κ_ν = α_ff(H⁻) * Pₑ * n(H I) / ρ

Based on Section 5.3 from Kurucz (1970), I'm fairly confident that the "more correct" version of the opacity equation should actually read κ_ν = α_ff(H⁻) * Pₑ * n(H I, n = 1) / ρ, and that the form provided by Gray (2005) implicitly assumes that n(H I, n = 1) ≈ n(H I). This seems to be a prettygood assumption given the tabulated values of the partition function in Table D.2 of Gray (2005). From the boltzmann equation: n(H I, n = 1) = n(H I)*gₙ₌₁/u(T)*exp(-Eₙ₌₁/(k*T)) = n(H I) * 2/u(T)*exp(0) = n(H I) * 2/u(T). According to Table D.2 of Grey, this approximation only introduces a ≤0.5% overestimate in n(H I, n = 1) over the temperature range where the polynomial is valid.

We also considered the polynomial fit in Section 5.3 from Kurucz (1970). Unfortunately, it seems wrong (it gives lots of negative numbers).

Korg.ContinuumOpacity.H2plus_bf_and_ff — Function

H2plus_bf_and_ff(nH_I_div_partition, n_HII, ν, ρ, T)

Compute the combined H₂⁺ bound-free and free-free opacity κ.

This uses polynomial fits from Gray (2005) that were derived from data tabulated in Bates (1952).

Arguments

nH_I_div_partition: the total number density of H I divided by its partition function.

nH_II: the number density of H II (not of H₂⁺).

ν: frequency in Hz

ρ: mass density in g/cm³

T: temperature in K

While the formal type signature requires only that these be Real, Float32s (or Float32-derived types) may introduce numerical instability.

Notes

This follows equation 8.15 of Gray (2005), which involves 2 polynomials that were fit to data provided in Bates (1952).

The combined H₂⁺ bound-free and free-free opacity opacity calculation can be cast as:

κ = const * σ₁ * exp(-U₁ / (kboltz*T)) * n(H I, n=1) * n(H II) * (1 - exp(-hν/(kboltz*T))) / ρ

where σ₁ and U₁ are just functions of ν. Note that Bates (1952) and Gray (2005) both use the number density of all energy states of H I instead of n(H I, n=1). However, Kurucz (1970) makes a note in section 5.2 about how the photodisociation of H₂⁺ produces a ground state H atom and that we should therefore use n(H I, n=1) instead. This difference will only cause Bates/Gray to be wrong by a fraction of a percent (at most) for T ≤ 1.2e4 K. In this function we use n(H I, n=1).

In this function, we use polynomial approximations that Gray (2005) fit for σ₁ and U₁ using data from Table 1 of Bates (1952). Gray (2005) notes that the approximations match Bates' (1952) tabulated absorption data for 3800 Å ≤ λ ≤ 25000 Å at an accuracy of 0.3%. After reviewing the data in Bates (1952) I've concluded that Gray (2005) actually rounded down the lower end of the wavelength from 10⁵/26 Å or ∼3847 Å. (In other words, the approximation is good for 3847 Å ≤ λ ≤ 25000.0 Å).

While this may fit to σ₁ and U₁ to better than 0.3%, comparisons against values from table 2 of Bates (1952) indicate that the function reproduces the full absorption coeffient at better than 1.5%. Gray (2005) did not provide a temperature range, but Bates (1952) only computed the absorption coefficient for 2500 K ≤ T ≤ 12000 K. In the text they have a comment that the proton's De Broglie wavelength is large at 2500 K and their semi-classical treatment may start to break down. However, as long as use n(H I, n=1), there doesn't seem to be any reason for us to enforce an upper temperature limit.

Bates (1952) states that his classical treatment of the interaction is probably most accurate at higher temperatures and longer wavelengths (note that the longest λ considered in the paper is 8 times larger than the max λ the polynomials are fit against). He suggests that treatment is probably correct "to well within one part in ten even at the lower temperatures and [lower wavelengths]."

Missing docstring.

Missing docstring for Korg.ContinuumOpacity.He_II_bf. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Korg.ContinuumOpacity.He_II_ff. Check Documenter's build log for details.

Korg.ContinuumOpacity.Heminus_ff — Function

Heminus_ff(nHe_I_div_partition, ne, ν, ρ, T)

Compute the He⁻ free-free opacity κ.

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + He I -> e⁻ + He I.

Arguments

nHe_I_div_partition: the total number density of H I divided by its partition function.

ne: the number density of free electrons.

ν: frequency in Hz

ρ: mass density in g/cm³

T: temperature in K

Notes

This follows equation 8.16 of Grey (2005) which provides a polynomial fit absorption to data tabulated in John 1994. According to that equation the "continuous absorption coefficient per Hydrogen" is given by:

    α(He⁻_ff)*Pₑ*A(He) / (1 + Φ(He)/Pₑ)

in which

log_10(α(He⁻_ff)) is the polynomial term.

A(He) is the number abundance of Helium particles relative to Hydrogen particles. For reference,
A(He) = [n(He I) + n(He II) + n(He III)] / [n(H I) + n(H II)]

Pₑ is the partial pressure contributed by electrons. For reference, Pₑ = nₑ*kb * T.

1/(1 + Φ(He)/Pₑ) comes from the Saha equation and expresses n(He I) / [n(He I) + n(He II)].

In the above expression, α(He⁻_ff)*Pₑ specifies the free-free atomic absorption coefficient per ground state He I atom. The expression seems to implicitly assume that

approximately all He I is in the ground state

n(He I) / [n(He I) + n(He II)] is roughly n(He I) / [n(He I) + n(He II) + n(HeIII)]

(at least for the range of temperatures where α can be accurately computed).

To convert the expression to opacity, they divided it by the product of the mean molecular weight and the Hydrogen mass. With that in mind, we can write an the equation for opacity (removing the apparent assumptions) as κ_ν = α(He⁻_ff)*Pₑ*n(He I, n=1)/ρ.

For 5063 Å ≤ λ ≤ 151878 Å and 2520 K ≤ T ≤ 10080 K, Gray (2005) claims that the polynomial fit the tabulated data at a precision of 1% or better. In practice, we found that it only fits the data to better than 3.1% (it's possible that for smaller λ the fit may be better). For reference, the tabulated data in these ranges of values consist of an irregularly spaced rectangular grid with 15 λ values and 9 Temperature values. According to John (1994), improved calculations are unlikely to alter the tabulated data for λ > 1e4Å, "by more than about 2%." The errors introduced by the approximations for 5.06e3 Å ≤ λ ≤ 1e4 Å "are expected to be well below 10%."

An alternative approach using a fit to older data is provided in section 5.7 of Kurucz (1970).

Korg.ContinuumOpacity.electron_scattering — Function

electron_scattering(nₑ, ρ)

Compute the opacity from scattering off of free electrons. This has no wavelength dependence. It assumes isotropic scattering

Arguments

nₑ::F: number density of free electrons (in cgs)

ρ::F: the mass density

Notes

I adopted the formula described in section 5.12 of Kurucz (1970) and the equation in the electron scattering subsection of Gray (2005); the actual coefficient value comes from the latter. It turns out that the coefficient in Kurucz (1970) has a typo (it's a factor of 10 too large).

Statistical mechanics

These functions can be used to calculate the number densities of all species in a given atospheric layer, or other statmech calculations.

Korg.saha_ion_weights — Function

saha_ion_weights(T, nₑ, atom, ionization_energies, partition_functions)

Returns (wII, wIII), where wII is the ratio of singly ionized to neutral atoms of a given element, and wIII is the ration of doubly ionized to neutral atoms.

arguments:

temperature T [K]

electron number density nₑ [cm^-3]

atom, the atomic number of the element

ionization_energies is a collection indexed by integers (e.g. a Vector) mappping elements' atomic numbers to their first three ionization energies

partition_funcs is a Dict mapping species to their partition functions

Korg.molecular_equilibrium_equations — Function

molecular_equilibrium_equations(absolute_abundances, ironization_energies, partiation_fns, equilibrium_constants)

Returns a NamedTuple representing the system of equations specifying molecular equilibrium.

arguments:

A Dict of absolute_abundances, NX/Ntotal

a Dict of ionization energies, ionization_energies. The keys of act as a list of all atoms.

a Dict of partition functions, partition_fns

a Dict of log molecular equilibrium constants, equilibrium_constants, in partial pressure form. The keys of equilibrium_constants act as a list of all molecules.

The system of equations is specified with the number densities of the neutral atoms as free parameters. Each equation specifies the conservation of a particular species, e.g. (simplified)

n(O) = n(CO) + n(OH) + n(O I) + n(O II) + n(O III).

In this equation:

n(O), the number density of oxygen atoms in any form comes absolute_abundances and the total

number density (supplied later)

n(O I) is a free parameter. The numerical solver is varying this to satisfy the system of equations.

n(O II), and n(O III) come from the Saha (ionization) equation given n(O I)

n(CO) and n(OH) come from the molecular equilibrium constants K, which are precomputed over a range of temperatures.

Equilibrium constants are defined in terms of partial pressures, so e.g.

K(OH)  ==  (p(O) p(H)) / p(OH)  ==  (n(O) n(H)) / n(OH)) kT

This could also be computed via a Saha equation involving the molecular partition function and dissolution energy.

Korg.molecular_equilibrium — Function

molecular_equilibrium(MEQS, T, nₜ, nₑ; x0)

Iteratively solve for the number density of each species. Returns a Dict mapping species to number densities.

arguments:

the system of molecular equilibrium equations MEQs (the thing returned by

molecular_equilibrium_equations)

the temperature T,

the number density of non-electron particles nₜ

the electron number density nₑ

optionally, x0, a starting point for the solver

Misc

Korg.blackbody — Function

blackbody(T, λ)

The value of the Planck blackbody function for temperature T at wavelength λ.

Korg.get_absolute_abundances — Function

get_absolute_abundances(metallicity, A_X)

Calculate $n_X/n_\mathrm{total}$ for each element X given some specified abundances, $A(X)$. Use the metallicity [$X$/H] to calculate those remaining from the solar values (except He).

Korg.air_to_vacuum — Function

air_to_vacuum(λ; cgs=λ<1)

Convert λ from an air wavelength to a vacuum wavelength. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

Korg.vacuum_to_air — Function

vacuum_to_air(λ; cgs=λ<1)

convert λ from a vacuum wavelength to an air wavelength λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.