This page docuements the Korg API for users of the package. Low-level functions that only people digging into the guts of Korg will be interested in can be found on the developer documentation.

Top-level functions

If you are synthesize a spectrum Korg, these are the functions you will call. These functions are exported, so if you do using Korg, you can call them unquallified (i.e. synthesize instead of Korg.synthesize).

Korg.synthesize — Function

synthesize(atm, linelist, λ_start, λ_stop, [λ_step=0.01]; metallicity=0, abundances=Dict(), vmic=0, ... )

Solve the transfer equation in the model atmosphere atm with the transitions in linelist at the from λ_start to λ_stop in steps of λ_step (all Å) to get the resultant astrophysical flux at each wavelength. An AbstractRange can also be probvided directly in place of λ_start, λ_stop, and λ_step.

Returns a named tuple with keys: - flux: the output spectrum - alpha: the linear absorption coefficient at each wavelenth and atmospheric layer a Matrix of size (layers x wavelengths) - number_densities: A dictionary mapping Species to vectors of number densities at each atmospheric layer - wavelengths: The vacuum wavelenths (in Å) over which the synthesis was performed. If air_wavelengths=true this will not be the same as the input wavelenths.

Optional arguments:

metallicity, i.e. [metals/H] is log_10 solar relative

abundances is a Dict mapping atomic symbols to $A(X)$ format abundances, i.e. $A(x) = \log_{10}(n_X/n_\mathrm{H}) + 12$, where $n_X$ is the number density of $X$. These override metallicity.

vmic (default: 0) is the microturbulent velocity, $\xi$, in km/s.

air_wavelengths (default: false): Whether or not the input wavelengths are air wavelenths to be converted to vacuum wavelengths by Korg. The conversion will not be exact, so that the wavelenth range can internally be represented by an evenly-spaced range. If the approximation error is greater than wavelength_conversion_warn_threshold, an error will be thrown.

wavelength_conversion_warn_threshold (default: 1e-4): see air_wavelengths.

line_buffer (default: 10): the farthest (in Å) any line can be from the provided wavelenth range before it is discarded. If the edge of your window is near a strong line, you may have to turn this up.

cntm_step (default 1): the distance (in Å) between point at which the continuum opacity is calculated.

hydrogen_lines (default: true): whether or not to include H lines in the synthesis.

mu_grid: the range of (surface) μ values at which to calculate the surface flux when doing transfer in spherical geometry (when atm is a ShellAtmosphere).

line_cutoff_threshold (default: 1e-3): the fraction of the continuum absorption coefficient at which line profiles are truncated. This has major performance impacts, since line absorption calculations dominate more syntheses. Turn it down for more precision at the expense of runtime. The default value should effect final spectra below the 10^-3 level.

ionization_energies, a Dict mapping Species to their first three ionization energies, defaults to Korg.ionization_energies.

partition_funcs, a Dict mapping Species to partition functions. Defaults to data from Barklem & Collet 2016, Korg.partition_funcs.

equilibrium_constants, a Dict mapping Species representing diatomic molecules to their molecular equilbrium constants in partial pressure form. Defaults to data from Barklem and Collet 2016, Korg.equilibrium_constants.

Korg.read_linelist — Function

read_linelist(filename; format="vald", isotopic_abundances=Korg.isotopic_abundances)

Parse a linelist file, returning a vector of Lines.

Pass format="kurucz" for a Kurucz linelist (format=kurucz_vac if it uses vacuum wavelengths; Be warned that Korg will not assume that wavelengths are vacuum below 2000 Å),format="vald" for a VALD linelist, and format="moog" for a MOOG linelist (doesn't support broadening parameters or dissociation energies).

VALD linelists (the default and preferred format) can be either "short" or "long" format, "extract all" or "extract stellar". Air wavelengths will automatically be converted into vacuum wavelengths, and energy levels will be automatically converted from cm$^{-1}$ to eV.

When they are not pre-scaled by isotopic abundace (which VALD does by default), Korg will automatically adjust the loggf of each line according to the `isotopicabundances, which defaults to the values from [NIST](https://www.nist.gov/pml/atomic-weights-and-isotopic-compositions-relative-atomic-masses). To use custom isotopic abundances, just passisotopic_abundancesas a dictionary mapping(atomic number, atomic weight)` pairs to abundances between 0 and 1.

Korg.read_model_atmosphere — Function

read_model_atmosphere(filename; truncate_at_10000K=true)

Parse the provided model atmosphere file in "Kurucz" format. Returns either a PlanarAtmosphere or a ShellAtmosphere.

When truncate_at_10000K is true, layers with temperatures greater than 10,000 Kelvin will be elided. These are typically at the deepest level, where optical depth is very large, and have minimal impact on the surface spectrum. They are ignored by default because Korg's default partition and molecular equillibrium functions are tabulated only up to that tempurature.

Secondary functions

You don't need use these to synthesize spectra, but they might be relevant depending on what you are doing.

Postprocessing

These are used to transform observational or synthetic spectra.

Korg.rectify — Function

rectify(flux, wls; bandwidth=50, q=0.95, wl_step=1.0)

Rectify the spectrum with flux vector flux and wavelengths wls by dividing out a moving q-quantile with window size bandwidth. wl_step controls the size of the grid that the moving quantile is calculated on and interpolated from. Setting wl_step to 0 results in the exact calculation with no interpolation, but note that this is very slow.

Experiments on real spectra show an agreement between the interpolated rectified spectrum and the "exact" one (with default values) at the 3 × 10^-4 level.

Korg.constant_R_LSF — Function

constant_R_LSF(flux, wls, R)

Applies a gaussian line spread function the the spectrum with flux vector flux and wavelength vector wls with constant spectral resolution, $R = \lambda/\Delta\lambda.$

Warning

This is a naive, slow implementation. Do not use it when performance matters.

constant_R_LSF will have weird behavior if your wavelength grid is not locally linearly-spaced. It is intended to be run on a fine wavelength grid, then downsampled to the observational (or otherwise desired) grid.

Line absorption

These functions can be used to directly compute line opacities.

Korg.line_absorption! — Function

line_absorption(linelist, λs, temp, nₑ, n_densities, partition_fns, ξ
               ; α_cntm=nothing, cutoff_threshold=1e-3, window_size=20.0*1e-8)

Calculate the opacity coefficient, α, in units of cm^-1 from all lines in linelist, at wavelengths λs [cm^-1].

other arguments:

temp the temerature in K

n_densities, a Dict mapping species to absolute number density [cm^-3].

partition_fns, a Dict containing the partition function of each species

ξ is the microturbulent velocity in cm/s (n.b. NOT km/s)

The three keyword arguments specify how the size of the window over which each line is calculated is chosen.

If α_cntm is passed as a callable returning the continuum opacity as a function of wavelength, the line window will extend to 4 doppler widths or the wavelength at which the Lorentz wings of the line are at cutoff_threshold * α_cntm[line.wl], whichever is greater. cutoff_threshold defaults to 1e-3.

if α_cntm is not passed, defaults to window_size, which is 2e-7 (in cm, i.e. 20 Å) unless otherwise specified

Korg.line_profile — Function

line_profile(λ₀, invΔλ_D, Δλ_L, line_amplitude, λ)

A voigt profile centered on λ₀ with doppler width invΔλ_D = 1/ΔλD and lorentz width `ΔλLevaluated atλ` (cm). Note that this returns values in units of cm^-1.

Korg.hydrogen_line_absorption — Function

hydrogen_line_absorption(λs, T, nₑ, nH_I, UH_I, hline_stark_profiles, ξ; 
                              stark_window_size=3e-7, self_window_size=1e-6)

Calculate contribution to the the absorption coefficient, α, from hydrogen lines in units of cm^-1, at wavelengths λs.

Arguments:

T: temperature [K]

nₑ: electron number density [cm^-3]

nH_I: neutral hydrogen number density [cm^-3]

hline_stark_profiles: (returned by setuphlinestark_profiles`)

ξ: microturbulent velocity [cm/s]. This is only applied to Hα-Hγ. Other hydrogen lines profiles are dominated by stark broadening, and the stark broadened profiles are pre-convolved with a doppler profile.

stark_window_size: the max distance from each line center [cm] at which to calculate the line absorption for lines besides Hα-Hγ (those dominated by stark broadening).

self_window_size: the max distance from each line center [cm] at which to calculate the line absorption for Hα-Hγ (those dominated by self-broadening).

Korg.setup_hydrogen_stark_profiles — Function

load_hydrogen_stark_profiles()

Load the (doppler-convolved) Stark-broadening profiles from disk.

Korg.voigt — Function

voigt(α, v)

The voigt function, $H$. Approximation from Hunger 1965.

Continuum absorption

These function can be used to directly compute continuum absorption coefficients.

Continuum Absorption Kwargs

All bound-free and free-free absorption functions (other than absorption_coef_bf_TOPBase) support the following keyword arguments:

error_oobounds::Bool: specifies the function's behavior when it encounters invalid ν or T values. When false (the default), the function assumes that the absorption coefficient for these invalid ν or T values is 0.0. Otherwise, a DomainError is thrown when invalid ν or T values are encountered.

out_α::Union{Nothing,AbstractVector}: When this is nothing (the default case), the function will simply allocate a new vector to store the output continuum absorption coefficients. Alternatively, this can be a vector (with the same length as the function's ν argument). In this case, the function will directly add the computed continuum absorption to the elements of this vector, in-place (the vector is also returned by the function).

Korg.ContinuumAbsorption.total_continuum_absorption — Function

total_continuum_absorption(νs, T, nₑ, number_densities, partition_funcs; error_oobounds)

The total continuum linear absoprtion coefficient, α, at many frequencies, ν.

Arguments

νs are frequencies in Hz

T is temperature in K

nₑ is the electron number density in cm^-3

number_densities is a Dict mapping each Species to its number density

partition_funcs is a Dict mapping each Species to its partition function (e.g. Korg.partition_funcs)

error_oobounds::Bool specifies the behavior of most continnum absorption sources when passed frequencies or temperature values that are out of bounds for their implementation. When false (the default), those absorption sources are ignored at those values. Otherwise, an error is thrown.

Note

For efficiency reasons, νs must be sorted. While this function technically supports any sorted AbstractVector, it is most effient when passed an AbstractRange.

Korg.ContinuumAbsorption.H_I_bf — Function

H_I_bf(ν, T, nH_I_div_partition, [ion_energy], [nmax_explicit_sum], [integrate_high_n];
       kwargs...)

Compute the bound-free linear absorption coefficient contributed by all energy states of a neutral Hydrogen atom.

Required Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I_div_partition is the total number density of neutral Hydrogen divided by the its partition function.

Optional Arguments

ion_energy: The ionization energy of Hydrogen. By default, this is set to the values loaded into the global ionization_energies list.

nmax_explicit_sum: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in an integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function wraps hydrogenic_ff_absorption. See that function for implementation details.

Korg.ContinuumAbsorption.H_I_ff — Function

H_I_ff(ν, T, nH_II, ne; kwargs...)

Compute the H I free-free linear absorption coefficient α.

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + H II -> e⁻ + H II.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_II: the number density of ionized Hydrogen in cm⁻³.

ne: the number density of free electrons.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function wraps hydrogenic_ff_absorption. See that function for implementation details.

Korg.ContinuumAbsorption.Hminus_bf — Function

Hminus_bf(ν, T, nH_I_div_partition, ne, [ion_energy_H⁻]; kwargs...)

Compute the H⁻ bound-free linear absorption coefficient α

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I_div_partition: the total number density of H I divided by its partition function.

ne: the electron number density

ion_energy_H⁻: Specifies the ionization energy of the single state of H⁻ in eV. This is roughly 0.7552 eV.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function assumes that n(H⁻) ≪ n(H I) + n(H II). The number density of n(H⁻) should not be pre-computed (instead it's computed internally by this function).

This function is adapted from equation 8.12 from Grey (2005). That equation specifies the absorption coefficient per H I atom (uncorrected by stimulated emission) as:

$8.316 \times 10^{-10} \sigma_{bf}(H^-) P_e \theta^{2.5} 10^{(\theta \chi_{\mathrm{H}^-})} U(H⁻,T)/ U(H I,T)$

where:

$\sigma_{bf}(\mathrm{H}^-)$ is the photo dissociation cross-section (Grey uses the variable $\alpha_{bf}$ in place of $\sigma_{bf}$). We estimate this by linearly interpolating data from [Wishart (1979)].

$\chi_{\mathrm{H}^-}$ is the ionization enrgy of H⁻.

θ = log10(e)/(k*T) or θ = 5040/T in units of eV⁻¹

U(H⁻,T) is the partition function of H⁻ at temperature T. This is always 1

U(H I,T) is the partition function of H I at temperature T. This is 2 at low to intermediate T.

Eqn 8.12 of Grey (2005) implicitly assumes that (U(H⁻,T)/ U(H I,T)) is always equal to 0.5.

This expression is simply a rewritten form of $\sigma_{bf}(H^-) n(H^-)/n(H I),$ where $n(H^-)/n(H I)$ has been replaced with the expanded form of the saha equation, in which n₀ = n(H⁻) and n₁ = n(H I).

Equations 8.18 and 8.19 of Gray (2005), indicate that the H⁻ bound-free opacity (with the stimulated emission correction) is given by:

$\kappa_\nu = \sigma_{bf}(H^-) \frac{n(H^-)}{n(H I)} (1 - \exp \left( \frac{-h\nu}{k T}\right)) \frac{n(H I)}{n(H I) + n(H II)} \frac{n(H I) + n(H II)}{\rho}$

In other words, the linear absorption coefficient is: $\alpha_\nu = \sigma_{bf}(H^-) n(H⁻) (1 - \exp \left( \frac{-h\nu}{k T}\right))$

Wishart (1979) expects the tabulated data to have better than 1% percent accuracy. Mathisen (1984) suggests that this data has better than 3% accuracy.

Korg.ContinuumAbsorption.Hminus_ff — Function

Hminus_ff(ν, T, nH_I_div_partition, ne; kwargs...)

Compute the H⁻ free-free linear absorption coefficient α

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + H I -> e⁻ + H I.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I_div_partition::Flt: the total number density of H I divided by its partition function.

ne: the number density of free electrons.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This is taken from equation 8.13 of Gray (2005). The equation uses a polynomial fig against Table 1 of Bell & Berrington's (1987), which tabulates values for "the H⁻ absorption coefficient" (the values include the correction for stimulated emission). We denote this quantity as f(H⁻) (Grey instead denotes it with the variable, $\alpha_{ff}(H^-)$). The polynomial fits f(H⁻) in the range 2520 K ≤ T ≤ 10080 K and 2604 Å ≤ λ ≤ 113918 Å. According to Grey, the polynomial fit to the tabulated data typically has 1% precision. We find that at worst, the discrepancy never exceeds 2.25%.

Based on equations 8.13, 8.18, and 8.19 of Gray (2005), the free-free linear absorption coefficient of H⁻ is given by: α = f(H⁻) * Pₑ * n(H I).

Based on Section 5.3 from Kurucz (1970), I'm fairly confident that the "more correct" version of this equation should actually read α = f(H⁻) * Pₑ * n(H I, n = 1), and that the form inferred from Gray (2005) implicitly assumes that n(H I, n = 1) ≈ n(H I). In detail, the Boltzmann equation relates n(H I, n = 1) and n(H I) as: n(H I, n = 1) = n(H I)*gₙ₌₁/U(T)*exp(-Eₙ₌₁/(k*T)) = n(H I) * 2/U(T)*exp(0) = n(H I) * 2/U(T). Based on the values of the partition function in Table D.2 of Grey, this approximation seems fairly good. It only introduces a ≤0.5% overestimate in n(H I, n = 1) over the temperature range where the polynomial is valid.

We also considered the polynomial fit in Section 5.3 from Kurucz (1970). Unfortunately, it seems to be wrong (it gives lots of negative numbers).

Korg.ContinuumAbsorption.H2plus_bf_and_ff — Function

H2plus_bf_and_ff(ν, T, nH_I_div_partition, n_HII; kwargs...)

Compute the combined H₂⁺ bound-free and free-free linear absorption coefficient α using the tables from Stancil 1994.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I: the total number density of H I divided by its partition function.

nH_II: the number density of H II (not of H₂⁺).

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This computes the H₂⁺ number density from those of H I and H II, since ionized molecules are not included in the molecular equlibrium calculation. Stancil provides approximate equilibrium constants, K, for the molecule, but the Barklem and Collet values used elsewhere by Korg may be more reliable. Once ionized molecules are fully supported, those values should be used instead. While cross sections are tabulated down to only 3150 K, the cross sections could be linearly interpolated 1000 K or so lower, if reliable K values are availble.

Because n(H₂⁺) is computed on the fly, the "cross-sections" used (internally) by this function have units of cm^-5, since they must be multiplied by n(H I) and n(H II) to obtain absorption coefficients.

Stancil, Bates (1952), and Gray (2005) all use n(H I), but Kurucz (1970) notes in section 5.2 that the photodisociation of H₂⁺ produces a ground state H atom and that we should therefore use n(H I, n=1) instead. This difference causes a descrepancy of a fraction of a percent (at most) for T ≤ 1.2e4 K. Here we use n(H I).

Korg.ContinuumAbsorption.He_II_bf — Function

He_II_bf(ν, T, nHe_I_div_partition, [ion_energy], [nmax_explicit_sum], [integrate_high_n];
         kwargs...)

Compute the bound-free linear absorption coefficient contributed by all energy states of a singly-ionized Helium atom.

Required Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nHe_II_div_partition is the total number density of singly-ionized Helium divided by the its partition function.

Optional Arguments

ion_energy: The ionization energy of Helium. By default, this is set to the values loaded into the global ionization_energies list.

nmax_explicit_sum: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in an integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function simply wraps hydrogenic_ff_absorption. See that function for additional details.

Korg.ContinuumAbsorption.He_II_ff — Function

He_II_ff(ν, T, nHe_III, ne; kwargs...)

Compute the He II free-free linear absorption coefficient α.

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + He III -> e⁻ + He III.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_III: the number density of doubly-ionized Helium in cm⁻³.

ne: the number density of free electrons.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function wraps hydrogenic_ff_absorption. See that function for implementation details.

Korg.ContinuumAbsorption.Heminus_ff — Function

Heminus_ff(ν, T, nHe_I_div_partition, ne; kwargs...)

Compute the He⁻ free-free opacity κ.

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + He I -> e⁻ + He I.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nHe_I_div_partition: the total number density of H I divided by its partition function.

ne: the number density of free electrons.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This follows equation 8.16 of Grey (2005) which provides a polynomial fit absorption to data tabulated in John (1994). According to that equation, the "continuous absorption coefficient per Hydrogen" is given by:

    f(He⁻_ff)*Pₑ*A(He) / (1 + Φ(He)/Pₑ)

in which

log_10(f(He⁻_ff)) is the polynomial term (Grey actually uses the variable α in place of f)

A(He) is the number abundance of Helium particles relative to Hydrogen particles. For reference,
A(He) = [n(He I) + n(He II) + n(He III)] / [n(H I) + n(H II)]

Pₑ is the partial pressure contributed by electrons. For reference, Pₑ = nₑ*kb * T.

1/(1 + Φ(He)/Pₑ) comes from the Saha equation and expresses n(He I) / [n(He I) + n(He II)].

In the above expression, f(He⁻_ff)*Pₑ specifies the free-free atomic absorption coefficient per ground state He I atom. The expression seems to implicitly assume that

approximately all He I is in the ground state

n(He I) / [n(He I) + n(He II)] is roughly n(He I) / [n(He I) + n(He II) + n(HeIII)]

(at least for the range of temperatures where f can be accurately computed).

To convert the expression to opacity, Grey divided it by the product of the mean molecular weight and the Hydrogen mass. With that in mind, we can write the equation for linear absorption coefficient, αν, (removing the apparent assumptions) as `αν = f(He⁻_ff)Pₑn(He I, n=1).`

For 5063 Å ≤ λ ≤ 151878 Å and 2520 K ≤ T ≤ 10080 K, Gray (2005) claims that the polynomial fit the tabulated data at a precision of 1% or better. In practice, we found that it only fits the data to better than 3.1% (it's possible that for smaller λ the fit may be better). For reference, the tabulated data in these ranges of values consist of an irregularly spaced rectangular grid with 15 λ values and 9 Temperature values. According to John (1994), improved calculations are unlikely to alter the tabulated data for λ > 1e4Å, "by more than about 2%." The errors introduced by the approximations for 5.06e3 Å ≤ λ ≤ 1e4 Å "are expected to be well below 10%."

An alternative approach using a fit to older data is provided in section 5.7 of Kurucz (1970).

Korg.ContinuumAbsorption.electron_scattering — Function

electron_scattering(nₑ)

Compute the linear absorption coefficient, α, from scattering off of free electrons. This has no wavelength dependence. It assumes isotropic scattering. (See, e.g. Gray p 160.)

Arguments

nₑ::F: number density of free electrons (in cgs)

Korg.ContinuumAbsorption.rayleigh — Function

rayleigh(λs, nH_I, nHe_II, nH2)

Absorption coefficient from Rayleigh scattering by neutral H, He, and H2. Formulations for H and He are via Colgan+ 2016. Formulation for H2 from Dalgarno and Williams 1962.

The Dalgarno and Williams H2 is applicable redward of 1300 Å. Since Rayleigh scattering breaks down when the particle size to wavelength ratio gets large, we that all frequencies passed to this function be equivalent to 1300 Å or greater.

The formulations for H is adapted from Lee 2005, which states that it is applicable redward of Lyman alpha. See Colgan 2016 for details on He.

Korg.ContinuumAbsorption.hydrogenic_bf_absorption — Function

hydrogenic_bf_absorption(ν, T, Z, nmax_explicit_sum, nsdens_div_partition,
                         ion_energy, [integrate_high_n])

Compute the bound-free linear absorption coefficient contributed by all energy states of a Hydrogenic species

The calculation is broken into 2 parts: (i) the contributions of the lowest energy states are explicitly summed and (ii) the contributions of the higher energy states are estimated with an integral.

Arguments

ν: frequency in Hz

T: temperature in K

Z::Integer: Z is the atomic number of the species (e.g. 1 for H I or 2 for He II)

nsdens_div_partition is the total number density of the species divided by the species's partition function.

ion_energy: the ionization energy from the ground state (in eV). This can be estimated as Z²*RydbergH (RydbergH is the ionization energy of Hydrogen)

nmax_explicit_sum::Integer: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in the integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

Notes

This follows the approach described in section 5.1 of Kurucz (1970).

Korg.ContinuumAbsorption.hydrogenic_ff_absorption — Function

hydrogenic_ff_absorption(ν, T, Z, ni, ne)

computes the free-free linear absorption coefficient for a hydrogenic species

The naming convention for free-free absorption is counter-intuitive. A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). In practice, this means that ni should refer to:

the number density of H II if computing the H I free-free absorption

the number density of He III if computing the He II free-free absorption

the number density of Li IV if computing the Li III free-free absorption

Arguments

Z::Integer: the charge of the ion. For example, this is 1 for ionized H.

ni: the number density of the ion species in cm⁻³.

ne: the number density of free electrons.

ν: frequency in Hz

T: temperature in K

Note

This approach was adopted from equation 5.8 from section 5.1 of Kurucz (1970). Comparison against equation 5.18b of Rybicki & Lightman (2004), reveals that the equation in Kurucz (1970) omits the dependence on ρ. According to Rybicki & Lightman (2004) the free-free absorption coefficient (corrected for stimulated emission) is:

    α = coef * Z² *ne * ni * (1 - exp(-hplanck*ν/(kboltz*T))) * g_ff / (sqrt(T) * ν³)

Note that the g_ff is the free-free gaunt factor and coef is ∼3.7e8 (a more exact coefficient can be computed from eqn 5.18a).

With this in mind, equation 5.8 of Kurucz (1970) should actually read

    κ = ne * n(H II) * F_ν(T) * (1 - exp(-hplanck*ν/(kboltz*T))) / ρ

where Fν(T) = coef * Z² * gff / (sqrt(T) * ν³).

See gaunt_ff_vanHoof for details about where our gaunt factor data comes from. For simplicity, we enforce temperature and ν bounds constraints that don't include all of the available gaunt factor data. In practice, this should never be a concern for stellar spectroscopy.

Statistical mechanics

These functions can be used to calculate the number densities of all species in a given atospheric layer, or other statmech calculations.

Korg.saha_ion_weights — Function

saha_ion_weights(T, nₑ, atom, ionization_energies, partition_functions)

Returns (wII, wIII), where wII is the ratio of singly ionized to neutral atoms of a given element, and wIII is the ration of doubly ionized to neutral atoms.

arguments:

temperature T [K]

electron number density nₑ [cm^-3]

atom, the atomic number of the element

ionization_energies is a collection indexed by integers (e.g. a Vector) mappping elements' atomic numbers to their first three ionization energies

partition_funcs is a Dict mapping species to their partition functions

Korg.molecular_equilibrium_equations — Function

molecular_equilibrium_equations(absolute_abundances, ironization_energies, partiation_fns, equilibrium_constants)

Returns a NamedTuple representing the system of equations specifying molecular equilibrium.

arguments:

A Dict of absolute_abundances, NX/Ntotal

a Dict of ionization energies, ionization_energies. The keys of act as a list of all atoms.

a Dict of partition functions, partition_fns

a Dict of log molecular equilibrium constants, equilibrium_constants, in partial pressure form. The keys of equilibrium_constants act as a list of all molecules.

The system of equations is specified with the number densities of the neutral atoms as free parameters. Each equation specifies the conservation of a particular species, e.g. (simplified)

n(O) = n(CO) + n(OH) + n(O I) + n(O II) + n(O III).

In this equation:

n(O), the number density of oxygen atoms in any form comes absolute_abundances and the total

number density (supplied later)

n(O I) is a free parameter. The numerical solver is varying this to satisfy the system of equations.

n(O II), and n(O III) come from the Saha (ionization) equation given n(O I)

n(CO) and n(OH) come from the molecular equilibrium constants K, which are precomputed over a range of temperatures.

Equilibrium constants are defined in terms of partial pressures, so e.g.

K(OH)  ==  (p(O) p(H)) / p(OH)  ==  (n(O) n(H)) / n(OH)) kT

This could also be computed via a Saha equation involving the molecular partition function and dissolution energy.

Korg.molecular_equilibrium — Function

molecular_equilibrium(MEQS, T, nₜ, nₑ; x0)

Iteratively solve for the number density of each species. Returns a Dict mapping species to number densities.

arguments:

the system of molecular equilibrium equations MEQs (the thing returned by

molecular_equilibrium_equations)

the temperature T,

the number density of non-electron particles nₜ

the electron number density nₑ

optionally, x0, a starting point for the solver

Misc

Korg.blackbody — Function

blackbody(T, λ)

The value of the Planck blackbody function for temperature T at wavelength λ.

Korg.get_absolute_abundances — Function

get_absolute_abundances(metallicity, A_X)

Calculate $n_X/n_\mathrm{total}$ for each element X given some specified abundances, $A(X)$. Use the metallicity [$X$/H] to calculate those remaining from the solar values (except He).

Korg.air_to_vacuum — Function

air_to_vacuum(λ; cgs=λ<1)

Convert λ from an air wavelength to a vacuum wavelength. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

Korg.vacuum_to_air — Function

vacuum_to_air(λ; cgs=λ<1)

convert λ from a vacuum wavelength to an air wavelength λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.