This is an automatically generated list of all functions in no particular order. In future, we should organize this and add context.

SSSynth.Line — Type

This type represents an individual line.

SSSynth.Line — Method

Construct a Line without explicit broadening parameters.

SSSynth.blackbody — Method

blackbody(T, λ)

The value of the Planck blackbody function for temperature T at wavelength λ.

SSSynth.constant_R_LSF — Method

Applies a gaussian line spread function with constant spectral resolution, R = λ/Δλ.

This will have weird behavior if your wavelength grid is not locally linearly-spaced. It is intended to be run on a fine wavelength grid, then downsampled to the observational (or otherwise desired) grid.

SSSynth.get_absolute_abundances — Method

Calculate NX/Ntotal for each X in elements given some specified_abundances, A(X). Use the metallicity [X/H] to calculate those remaining from the solar values (except He).

SSSynth.get_atoms — Method

Get the atoms that make up a diatomic molecule

SSSynth.get_mass — Method

Returns the mass of atom_or_molecule in g.

SSSynth.ismolecule — Method

true if the string passed represents a molecule (with or without its ionization state)

SSSynth.line_absorption — Method

line_absorption(linelist, λs, temp, n_densities, atomic_masses, partition_fns, 
             ionization_energies; window_size)

Calculate the opacity coefficient, α, in units of cm^-1 from all lines in linelist, at wavelengths λs.

other arguments:

temp the temerature in K

n_densities, a Dict mapping species to absolute number density [cm^-3].

window_size (optional, default: 20), the maximum distance from the line center at which line

opacities should be calculated in included.

ξ is the microturbulent velocity in cm/s

SSSynth.line_profile — Method

line_profile(temp, atomic_mass, ξ, line, λs)

The line profile, ϕ, at wavelengths λs in cm. temp should be in K, atomic_mass should be in g, microturbulent velocity, ξ, should be in cm/s, line should be one of the entries returned by read_line_list. Note that this returns values in units of cm^-1, not Å^-1

SSSynth.molecular_equilibrium — Method

Iteratively solve for the number density of each species. Returns a Dict mapping species to number densities.

arguments:

the system of molecular equilibrium equations MEQs (the thing returned by

molecular_equilibrium_equations)

the temperature T,

the number density of non-electron particles nₜ

the electron number density nₑ

a Dict of NX/Ntotal, absolute_abundances

a Dict of ionization energies, ionization_energies

a Dict of partition functions, partition_fns

a Dict of log molecular equilibrium constants, equilibrium_constants, in partial pressure form.

SSSynth.molecular_equilibrium_equations — Method

Returns a NamedTuple representing the system of equations specifying molecular equilibrium.

arguments:

A Dict of absolute_abundances, NX/Ntotal

a Dict of ionization energies, ionization_energies. The keys of ionization_energies act as a list of all atoms.

a Dict of partition functions, partition_fns

a Dict of log moleculare equilibrium constants, equilibrium_constants, in partial pressure form. The keys of equilibrium_constants act as a list of all molecules.

The system of equations is specified with the number densities of the neutral atoms as free parameters. Each equation specifies the conservation of a particular species, e.g. (simplified)

n(O) = n(CO) + n(OH) + n(O I) + n(O II) + n(O III).

In this equation:

n(O), the number density of oxygen atoms in any form comes absolute_abundances and the total

number density (supplied later)

n(O I) is a free parameter. The numerical solver is varying this to satisfy the system of equations.

n(O II), and n(O III) come from the Saha (ionization) equation given n(O I)

n(CO) and n(OH) come from the molecular equilibrium constants K, which are precomputed over a range of temperatures.

K is defined in terms of partial pressures, so e.g. K(OH) == (p(O) p(H)) / p(OH) == (n(O) + n(H)) / n(OH) kT Alternatively, this could be computed via a Saha equation involving the molecular partition function and dissolution energy.

SSSynth.parse_species_code — Method

parse_species_code(code)

Parse the "species code" as it is often specified in line lists and return a the "astronomy" notation. 01.00 → "HI", 02.01 → "HeII", 02.1000 → "HeII", 0608 → "COI", etc.

SSSynth.read_line_list — Method

read_line_list(fname; format="kurucz")

Parse the provided line list. in "Kurucz" format. Pass format="kurucz" for a Kurucz line list, format="vald" for a Vald line list, and format="moog" for a MOOG line list.

Note that dissociation energies in a MOOG line list will be ignored.

SSSynth.read_model_atmosphere — Method

read_model_atmosphere(fname)

Parse the provided model atmosphere file in "Kurucz" format.

SSSynth.read_partition_funcs — Method

functon setup_atomic_partition_funcs()

Constructs a Dict holding functions for various atomic species that computes the partition function at an input temperature.

In more detail, the partition function is linearly interpolated from values tabulated from Barklem & Collet (2016). The partition functions can only be evaluated at temperatures between 1e-5 K and 1e4 K. The tabulated values were NOT computed with a cutoff principle quantum number. This can lead to a couple percent error for some neutral species in solar type stars (I think the problem becomes worse in hotter stars). See the paper for more details.

Note that the paper seems to suggest that we could actually use cubic interpolation. That should be revisited in the future.

SSSynth.saha_ion_weights — Method

Returns (wII, wIII), where wII is the ratio of singly ionized to neutral atoms of a given element, and wIII is the ration of doubly ionized to neutral atoms.

arguments:

temperature T [K]

electron number density nₑ [cm^-3]

atom, the atomic symbol of the element

ionization_energies is a Dict mapping elements to their first three ionization energies

partition_funcs is a Dict mapping species to their partition functions

SSSynth.saha_ion_weights — Method

convieience method for easier testing

SSSynth.setup_equilibrium_constants — Method

Returns a Dict holding the default (Barklem & Collet 2016) log equilibrium constants.

There are two confusing things about equation 7 in this paper. The first is that it defines K as the reciprocal of what is actually supplied in the data table. The second is that in order for it to be correct, m must be interpreted as a reduced mass, m₁m₂/(m₁ + m₂). This can be verified by re-deriving the equilibrium constants from the partition functions and dissolution constants provided by the paper.

SSSynth.setup_ionization_energies — Function

setupionizationenergies([filename])

Parses the table of ionization energies and returns it as a dictionary mapping elements to their ionization energies, [χ₁, χ₂, χ₃].

SSSynth.setup_partition_funcs — Function

Returns a Dict holding the default (Barklem & Collet 2016) partition functions.

SSSynth.sigma_line — Method

sigma_line(wl, log_gf)

The cross-section at wavelength wl in Ångstroms of a transition for which the product of the degeneracy and oscillator strength is 10^log_gf.

SSSynth.strip_ionization — Method

get the chemical symbol for the element of the species

SSSynth.synthesize — Method

synthesize(atm, linelist, λs, [metallicity, [alpha]]; abundances=Dict())

Solve the transfer equation in the model atmosphere atm with the transitions in linelist at the wavelengths λs [Å] to get the resultant astrophysical flux at each wavelength.

optional arguments:

metallicity, i.e. [metals/H] is log_10 solar relative

vmic (default: 0) is the microturbulent velocity, ξ, in km/s.

abundances are A(X) format, i.e. A(x) = log10(nX/nH), where nX is the number density of X.

line_window (default: 10): the farthest any line can be from the provide wavelenth range range before it is discarded (in Å).

ionization_energies, a Dict containing the first three ionization energies of each element, defaults to SSSynth.ionization_energies.

partition_funcs, a Dict mapping species to partition functions. Defaults to data from Barklem & Collet 2016, SSSynth.partition_funcs.

equilibrium_constants, a Dict mapping diatomic molecules to theirmolecular equilbrium constants in partial pressure form. Defaults to data from Barklem and Collet 2016, SSSynth.equilibrium_constants.

Uses solar abundances scaled by metallicity and for those not provided.

SSSynth.total_continuum_opacity — Method

The total continuum opacity, κ, at many frequencies, ν.

νs are frequencies in Hz

T is temperature in K

nₑ is the electron number density in cm^-3

ρ is the density in g cm^-3

number_densities is a Dict mapping each species to its number density

partition_funcs is a `Dict mapping each species to its partition function

SSSynth.translational_U — Method

The contribution to the partition function from the free movement of a particle. Used in the Saha equation.

arguments

m is the particle mass

T is the temperature in K

SSSynth.trapezoid_rule — Method

trapezoid_rule(xs, fs)

Approximate the integral from x₁ to x₂ of f(x) with the trapezoid rule given x-values xs and f(x) values fs.

This should be good enough to numerically solve the transport equation, since model atmospheres usually have carefully chosen knots. We probably want to add higher-order aproximations later.

SSSynth.ContinuumOpacity.H2plus_bf_and_ff — Method

H2plus_bf_and_ff(nH_I_div_partition, n_HII, ν, ρ, T)

Compute the combined H₂⁺ bound-free and free-free opacity κ.

This uses polynomial fits from Gray (2005) that were derived from data tabulated in Bates (1952).

Arguments

nH_I_div_partition::Float64: the total number density of H I divided by its partition function.

nH_II::Float64: the number density of H II (not of H₂⁺).

ν::Float64: frequency in Hz

ρ::Float64: mass density in g/cm³

T::Float64: temperature in K

Notes

This follows equation 8.15 of Gray (2005), which involves 2 polynomials that were fit to data provided in Bates (1952).

The combined H₂⁺ bound-free and free-free opacity opacity calculation can be cast as: κ = const * σ₁ * exp(-U₁ / (kboltzT)) * n(H I, n=1) * n(H II) * (1 - exp(-hν/(kboltzT))) / ρ where σ₁ and U₁ are just functions of ν. Note that Bates (1952) and Gray (2005) both use the number density of all energy states of H I instead of n(H I, n=1). However, Kurucz (1970) makes a note in section 5.2 about how the photodisociation of H₂⁺ produces a ground state H atom and that we should therefore use n(H I, n=1) instead. This difference will only cause Bates/Gray to be wrong by a fraction of a percent (at most) for T ≤ 1.2e4 K. In this function we use n(H I, n=1).

In this function, we use polynomial approximations that Gray (2005) fit for σ₁ and U₁ using data from Table 1 of Bates (1952). Gray (2005) notes that the approximations match Bates' (1952) tabulated absorption data for 3800 Å ≤ λ ≤ 25000 Å at an accuracy of 0.3%. After reviewing the data in Bates (1952) I've concluded that Gray (2005) actually rounded down the lower end of the wavelength from 10⁵/26 Å or ∼3847 Å. (In other words, the approximation is good for 3847 Å ≤ λ ≤ 25000.0 Å).

While this may fit to σ₁ and U₁ to better than 0.3%, comparisons against values from table 2 of Bates (1952) indicate that the function reproduces the full absorption coeffient at better than 1.5%. Gray (2005) did not provide a temperature range, but Bates (1952) only computed the absorption coefficient for 2500 K ≤ T ≤ 12000 K. In the text they have a comment that the proton's De Broglie wavelength is large at 2500 K and their semi-classical treatment may start to break down. However, as long as use n(H I, n=1), there doesn't seem to be any reason for us to enforce an upper Temperature limit.

Bates (1952) states that his classical treatment of the interaction is probably most accurate at higher temperatures and longer wavelengths (note that the longest λ considered in the paper is 8 times larger than the max λ the polynomials are fit against). He suggests that treatment is probably correct "to well within one part in ten even at the lower temperatures and [lower wavelengths]."

SSSynth.ContinuumOpacity.Heminus_ff — Method

Heminus_ff(nHe_I_div_partition, ne, ν, ρ, T)

Compute the He⁻ free-free opacity κ.

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + He I -> e⁻ + He I.

Arguments

nHe_I_div_partition::Flt: the total number density of H I divided by its partition function.

ne::Flt: the number density of free electrons.

ν::Flt: frequency in Hz

ρ::Flt: mass density in g/cm³

T::Flt: temperature in K

Notes

This follows equation 8.16 of Grey (2005) which provides a polynomial fit absorption to data tabulated in John 1994. According to that equation the "continuous absorption coefficient per Hydrogen" is given by: α(He⁻_ff)PₑA(He) / (1 + Φ(He)/Pₑ) in which

log10(α(He⁻ff)) is the polynomial term.

A(He) is the number abundance of Helium particles relative to Hydrogen particles. For reference, A(He) = [n(He I) + n(He II) + n(He III)] / [n(H I) + n(H II)]

Pₑ is the partial pressure contributed by electrons. For reference, Pₑ = nₑ*kb * T.

1/(1 + Φ(He)/Pₑ) comes from the Saha equation and expresses n(He I) / [n(He I) + n(He II)].

In the above expression, α(He⁻_ff)*Pₑ specifies the free-free atomic absorption coefficient per ground state He I atom. The expression seems to implicitly assume that

approximately all He I is in the ground state

n(He I) / [n(He I) + n(He II)] is roughly n(He I) / [n(He I) + n(He II) + n(HeIII)]

(at least for the range of temperatures where α can be accurately computed).

To convert the expression to opacity, they divided it by the product of the mean molecular weight and the Hydrogen mass. With that in mind, we can write an the equation for opacity (removing the apparent assumptions) as κν = α(He⁻ff)Pₑn(He I, n=1)/ρ.

For 5063 Å ≤ λ ≤ 151878 Å and 2520 K ≤ T ≤ 10080 K, Gray (2005) claims that the polynomial fit the tabulated data at a precision of 1% or better. In practice, we found that it only fits the data to better than 3.1% (it's possible that for smaller λ the fit may be better). For reference, the tabulated data in these ranges of values consist of an irregularly spaced rectangular grid with 15 λ values and 9 Temperature values. According to John (1994), improved calculations are unlikely to alter the tabulated data for λ > 1e4Å, "by more than about 2%." The errors introduced by the approximations for 5.06e3 Å ≤ λ ≤ 1e4 Å "are expected to be well below 10%."

An alternative approach using a fit to older data is provided in section 5.7 of Kurucz (1970).

SSSynth.ContinuumOpacity.Hminus_bf — Method

Hminus_bf(nH_I_div_partition, ne, ν, ρ, T, [ion_energy_H⁻])

Compute the H⁻ bound-free opacity κ

Arguments

nH_I_div_partition::Flt: the total number density of H I divided by its partition function.

ne the electron number density

ν::Flt: frequency in Hz

ρ::Flt: mass density in g/cm³

T::Flt: temperature in K

ion_energy_H⁻::Flt: Specifies the ionization energy of the single state of H⁻ in eV. This is roughly 0.7552 eV.

Notes

This function assumes that n(H⁻) ≪ n(H I) + n(H II). The number density of n(H⁻) should not be pre-computed (instead it's computed internally by this function).

This function is adapted from equation 8.12 from Grey (2005). This equation gives the absorption coefficient per H I atom (uncorrected by stimulated emission) is given by: αffH⁻ * 8.316e-10 * Pₑ * θ^2.5 * 10^(ionenergy_H⁻ * θ) * (U(H⁻,T)/ U(H I,T)) where:

αff_H⁻ is the photo dissociation cross-section. This can estimated with equation 8.11.

θ = log10(e)/(k*T) or θ = 5040/T in units of eV⁻¹

U(H⁻,T) is the partition function of H⁻ at temperature T. This is always 1

U(H I,T) is the partition function of H I at temperature T. This is 2 at low to intermediate T.

Eqn 8.12 of Grey (2005) implicitly assumes that (U(H⁻,T)/ U(H I,T)) is always equal to 0.5.

This expression is simplicitly a rewritten form of: αff_H⁻ * n(H⁻)/n(H I) where n(H⁻)/n(H I) has been replaced with the expanded form of the saha equation, in which n₀ = n(H⁻) and n₁ = n(H I).

Combining 8.18, and 8.19 of Gray (2005), indicate that the version opacity contribution of H⁻ bound-free absorption (with stimulated emission correction) is given by: n(H⁻) n(H I) n(H I) + n(H II) κν = αbf(H⁻) * ––– * (1 - exp(-hν/(kT))) * ––––––––– * ––––––––- n(H I) n(H I) + n(H II) ρ This can be rewritten as: κν = αbf(H⁻) * n(H⁻) * (1 - exp(-hν/(kT))) / ρ

This function uses the polynomial provided in equation 8.11 of Gray (2005), that fits the tabulated data from Wishart (1979). While Gray (2005) claims that the polynomial fits the data with 0.2% precision for 2250 Å ≤ λ ≤ 15000 Å, in practice we find that it fits the data to better than 0.25% precision. Wishart (1979) expects the tabulated data to have better than 1% percent accuracy.

SSSynth.ContinuumOpacity.Hminus_ff — Method

Hminus_ff(nH_I_div_partition, ne, ν, ρ, T)

Compute the H⁻ free-free opacity κ

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + H I -> e⁻ + H I.

Arguments

nH_I_div_partition::Flt: the total number density of H I divided by its partition function.

ne::Flt: the number density of free electrons.

ν::Flt: frequency in Hz

ρ::Flt: mass density in g/cm³

T::Flt: temperature in K

Notes

This is taken from equation 8.13 of Gray (2005). This uses a polynomial to Bell & Berrington (1987) tabulated values for αff(H⁻), which implicitly includes the correction for stimulated emission. The polynomial fits αff(H⁻) in the range 2520 K ≤ T ≤ 10080 K and 2604 Å ≤ λ ≤ 113918 Å. According to the book, the polynomial fit to the tabulated data typically has 1% precision. We find that at worst, the discrepancy never exceeds 2.25%.

From equations 8.13, 8.18, and 8.19 of Gray (2005), the free-free opacity from H⁻ is given by: n(H I) n(H I) + n(H II) κν = αff(H⁻) * Pₑ * ––––––––– * ––––––––- n(H I) + n(H II) ρ This can be rewritten as: κν = αff(H⁻) * Pₑ * n(H I) / ρ

Based on Section 5.3 from Kurucz (1970), I'm fairly confident that the "more correct" version of the opacity equation should actually read κν = αff(H⁻) * Pₑ * n(H I, n = 1) / ρ, and that the form provided by Gray (2005) implicitly assumes that n(H I, n = 1) ~ n(H I). This seems to be a prettygood assumption given the tabulated values of the partition function in Table D.2 of Gray (2005). From the boltzmann equation: n(H I, n = 1) = n(H I)gₙ₌₁/u(T)exp(-Eₙ₌₁/(kT)) = n(H I) * 2/u(T)exp(0) = n(H I) * 2/u(T). According to Table D.2 of Grey, this approximation only introduces a ≤0.5% overestimate in n(H I, n = 1) over the temperature range where the polynomial is valid.

We also considered the polynomial fit in Section 5.3 from Kurucz (1970). Unfortunately, it seems wrong (it gives lots of negative numbers).

SSSynth.ContinuumOpacity._Hminus_bf_cross_section — Method

_Hminus_bf_cross_section(ν)

Compute the H⁻ bound-free cross-section, which has units of cm^2 per H⁻ particle.

The cross-section does not include a correction for stimulated emission.

SSSynth.ContinuumOpacity._hydrogenic_bf_high_n_opacity — Method

_hydrogenic_bf_high_n_opacity(Z, nmin, nsdens_div_partition, ν, T, ion_energy)

Approximates the sum of opacity contributions by all bound-free transitions in which the electron originates in an energy level of nmin or larger using an integral.

In full detail, the calculated quantity is described as ∑{n'=nmin}^∞ ndens(n=n') * αν(n=n') * (1 - exp(-hν/k/T))/ρ, where α(n=n') is the absorption coefficient for the bound-free atomic absorption coefficient (uncorrected for stimulated emission). This function approximates this sum with an integral.

Arguments

Z::Integer: Z is the atomic number of the ion (1 for HI)

nmin::Integer: The lowest energy level (principle quantum number) included in the calculation

nsdens_div_partition::AbstractFloat is the number density of the current species divided by the partition function.

ν::Flt: frequency in Hz

ρ::Flt: mass density in g/cm³

T::Flt: temperature in K

ion_energy::AbstractFloat: the ionization energy from the ground state (in eV).

Notes

This implements equation (5.6) from Kurucz (1970). I think ρ was simply omitted from that equation.

SSSynth.ContinuumOpacity._ndens_Hminus — Function

_ndens_Hminus(nH_I_div_partition, ne, T, ion_energy = _H⁻_ion_energy)

Compute the number density of H⁻ (implements eqn 5.10 of Kurucz 1970). This is an application of the saha equation where the "ground state" is H⁻ and the "first ionization state" is H I. The partition function of H⁻ is 1 at all temperatures.

SSSynth.ContinuumOpacity.electron_scattering — Method

electron_scattering(nₑ, ρ)

Compute the opacity from scattering off of free electrons. This has no wavelength dependence. It assumes isotropic scattering

Arguments

nₑ::F: number density of free electrons (in cgs)

ρ::F: the mass density

Notes

I adopted the formula described in section 5.12 of Kurucz (1970) and the equation in the electron scattering subsection of Gray (2005); the actual coefficient value comes from the latter. It turns out that the coefficient in Kurucz (1970) has a typo (it's a factor of 10 too large).

SSSynth.ContinuumOpacity.gaunt_ff_kurucz — Method

gaunt_ff_kurucz(log_u, log_γ2)

computes the thermally averaged free-free gaunt factor by interpolating the table provided in section 5.1 of Kurucz (1970). The table was derived from a figure in Karsas and Latter (1961).

Arguments

log_u::F: Equal to log₁₀(u) = log₁₀(hν/(kT))

log_γ2::F: Equal to log₁₀(γ²) = log₁₀(RydbergHZ²/(kT))

Note

There is some ambiguity over whether we should replace RydbergH*Z² in the definition of γ² with the ionization energy, but it's probably a negligible difference (given the log scale).

In the future, we may want to the interpolate the table reported in van Hoof, Ferland, Williams, Volk, Chatzikos, Lykins, & Porter (2013) because it's more accurate and applies over a larger range of values.

SSSynth.ContinuumOpacity.hydrogenic_bf_opacity — Method

hydrogenic_bf_opacity(Z, nmax_explicit_sum, nsdens_div_partition, ν, ρ, T,
                      ion_energy, [integrate_high_n])

Compute the bound-free opacity contributed by all energy states of a Hydrogenic species

The calculation is broken into 2 parts: (i) the contributions of the lowest energy states are explicitly summed and (ii) the contributions of the higher energy states are estimated with an integral.

Arguments

Z::Integer: Z is the atomic number of the species (e.g. 1 for H I or 2 for He II)

nsdens_div_partition::Flt is the total number density of the species divided by the species's partition function.

ν::Flt: frequency in Hz

ρ::Flt: mass density in g/cm³

T::Flt: temperature in K

ion_energy::AbstractFloat: the ionization energy from the ground state (in eV). This can be estimated as Z²*RydbergH (RydbergH is the ionization energy of Hydrogen)

nmax_explicit_sum::Integer: The highest energy level whose opacity contribution is included in the explicit sum. The contributions from higher levels are included in the integral.

integrate_high_n::bool: When this is false, bf opacity from higher energy states are not estimated at all. Default is true.

Notes

This follows the approach described in section 5.1 of Kurucz (1970).

SSSynth.ContinuumOpacity.hydrogenic_ff_opacity — Method

hydrogenic_ff_opacity(Z, ni, ne, ν, ρ, T)

computes the free-free opacity for a hydrogenic species

The naming convention for free-free absorption is counter-intuitive. A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). In practice, this means that ni should refer to:

the number density of H II if computing the H I free-free opacity

the number density of He III if computing the He II free-free opacity

the number density of Li IV if computing the Li III free-free opacity

Arguments

Z::Integer: the charge of the ion. For example, this is 1 for ionized H.

ni::Flt: the number density of the ion species in cm⁻³.

ne::Flt: the number density of free electrons.

ν::Flt: frequency in Hz

ρ::Flt: mass density in g/cm³

T::Flt: temperature in K

Note

This approach was adopted from equation 5.8 from section 5.1 of Kurucz (1970). Comparison against equation 5.18b of Rybicki & Lightman (2004), reveals that the equation in Kurucz (1970) omits the dependence on ρ. According to Rybicki & Lightman (2004) the free-free opacity (corrected for stimulated emission) is: coef * Z² ne * ni * (1 - exp(-hplanckν/(kboltz*T))) * gff / (sqrt(T) * ν³ * ρ) Note that the gff is the free-free gaunt factor and coef is ∼3.7e8 (a more exact coefficient can be computed from eqn 5.18a).

With this in mind, equation 5.8 of Kurucz (1970) should actually read ne * n(H II) * Fν(T) * (1 - exp(-hplanckν/(kboltzT))) / ρ where Fν(T) = coef * Z² * g_ff / (sqrt(T) * ν³).

SSSynth.ContinuumOpacity.ndens_state_hydrogenic — Method

ndens_state_hydrogenic(n, nsdens_div_partition, ion_energy, T)

Calculates the LTE number density (in cm^-3) of a hydrogenic species at a given energy level.

Arguments

n::Integer: The quantum number of the state

nsdens_div_partition::Flt: The total (including all states) number density of the current

ionization species (in cm^-3) divided by the species's partition function.

ion_energy::Flt: The minimum energy (in eV) required to ionize the species (from the ground

state). This can be estimated as Z²ion_energy of hydrogen or Z²Rydberg_H.

T::Flt: Temperature

Note

This assumes that all hydrogenic species have a statistical weight of gₙ = 2*n².

This was taken from equation (5.4) of Kurucz (1970) (although this comes directly from the boltzmann equation).