API

This page docuements the Korg API for users of the package. Low-level functions that only people digging into the internals of Korg will be interested in can be found on the Developer documentation.

Top-level functions

If you are synthesize a spectrum Korg, these are the functions you will call. These functions are exported, so if you do using Korg, you can call them unquallified (i.e. synthesize instead of Korg.synthesize).

Korg.synthesize — Function

synthesize(atm, linelist, A_X, λ_start, λ_stop, [λ_step=0.01]; kwargs... )

Compute a synthetic spectrum.

Arguments

atm: the model atmosphere (see read_model_atmosphere)

linelist: A vector of [Line](@ref)s (see read_linelist)

A_X: a vector containing the A(X) abundances (log(X/H) + 12) for elements from hydrogen to uranium. (see format_A_X)

λ_start: the lower bound (in Å) of the region you wish to synthesize.

λ_stop: the upper bound (in Å) of the region you wish to synthesize.

λ_step (default: 0.01): the (approximate) step size to take (in Å).

Returns

A named tuple with keys:

flux: the output spectrum

alpha: the linear absorption coefficient at each wavelenth and atmospheric layer a Matrix of size (layers x wavelengths)

number_densities: A dictionary mapping Species to vectors of number densities at each atmospheric layer

wavelengths: The vacuum wavelenths (in Å) over which the synthesis was performed. If air_wavelengths=true this will not be the same as the input wavelenths.

Example

to synthesize a spectrum between 5000 Å and 5100 Å, with all metal abundances set to 0.5 dex less than the solar value except carbon, except carbon, which we set to [C/H]=-0.25:

atm = read_model_atmosphere("path/to/atmosphere.mod")
linelist = read_linelist("path/to/linelist.vald")
A_X = format_A_X(-0.5, Dict("C" => -0.25))
solution = synthesize(atm, linelist, A_X, 5000, 5100)

Optional arguments:

vmic (default: 0) is the microturbulent velocity, $\xi$, in km/s.

air_wavelengths (default: false): Whether or not the input wavelengths are air wavelenths to be converted to vacuum wavelengths by Korg. The conversion will not be exact, so that the wavelenth range can internally be represented by an evenly-spaced range. If the approximation error is greater than wavelength_conversion_warn_threshold, an error will be thrown. (To do wavelength conversions yourself, see air_to_vacuum and vacuum_to_air.)

wavelength_conversion_warn_threshold (default: 1e-4): see air_wavelengths. (In Å.)

line_buffer (default: 10): the farthest (in Å) any line can be from the provided wavelenth range before it is discarded. If the edge of your window is near a strong line, you may have to turn this up.

cntm_step (default 1): the distance (in Å) between point at which the continuum opacity is calculated.

hydrogen_lines (default: true): whether or not to include H lines in the synthesis.

hydrogen_line_window_size (default: 150): the mamximum distance (in Å) from each hydrogen line center at which to calculate its contribution to the total absorption coefficient.

n_mu_points (default: 20): the number of μ values at which to calculate the surface flux when doing transfer in spherical geometry (when atm is a ShellAtmosphere). 20 points is sufficient for accuracy at the 10^-3 level.

line_cutoff_threshold (default: 3e-4): the fraction of the continuum absorption coefficient at which line profiles are truncated. This has major performance impacts, since line absorption calculations dominate more syntheses. Turn it down for more precision at the expense of runtime. The default value should effect final spectra below the 10^-3 level.

ionization_energies, a Dict mapping Species to their first three ionization energies, defaults to Korg.ionization_energies.

partition_funcs, a Dict mapping Species to partition functions (in terms of ln(T)). Defaults to data from Barklem & Collet 2016, Korg.partition_funcs.

equilibrium_constants, a Dict mapping Species representing diatomic molecules to their molecular equilbrium constants in partial pressure form. Defaults to data from Barklem and Collet 2016, Korg.equilibrium_constants.

bezier_radiative_transfer (default: false): Use the radiative transfer scheme. This is for testing purposes only.

Korg.read_linelist — Function

read_linelist(filename; format="vald", isotopic_abundances=Korg.isotopic_abundances)

Parse a linelist file, returning a vector of Lines.

Pass format="kurucz" for a Kurucz linelist (format=kurucz_vac if it uses vacuum wavelengths; Be warned that Korg will not assume that wavelengths are vacuum below 2000 Å),format="vald" for a VALD linelist, and format="moog" for a MOOG linelist (doesn't support broadening parameters or dissociation energies).

VALD linelists (the default and preferred format) can be either "short" or "long" format, "extract all" or "extract stellar". Air wavelengths will automatically be converted into vacuum wavelengths, and energy levels will be automatically converted from cm$^{-1}$ to eV.

When they are not pre-scaled by isotopic abundace (which VALD does by default), Korg will automatically adjust the log_gf of each line according to isotopic_abundances, which defaults to the values from NIST. To use custom isotopic abundances, just pass isotopic_abundances as a dictionary mapping (atomic number, atomic weight) pairs to abundances between 0 and 1.

Korg.read_model_atmosphere — Function

read_model_atmosphere(filename)

Parse the provided model atmosphere file in MARCS ".mod" format. Returns either a PlanarAtmosphere or a ShellAtmosphere.

Warning

Korg does not yet support cool ($\lesssim$ 3500 K) stars. While it will happily parse their model atmospheres, it is very likely to crash when you feed them into synthesize.

Korg.format_A_X — Function

format_A_X(metallicity, abundances; kwargs... )

Returns a 92 element vector containing abundances in $A(X)$ ($\log_{10}(X/H) + 12$) format for elements from hydrogen to uranium.

Arguments

You can provide either or both of:

metallicity (default: 0), i.e. [metals/H] is the $\log_{10}$ solar-relative abundance of elements heavier than He. It is overriden by abundances, on a per-element basis.

abundances is a Dict mapping atomic numbers or symbols to [$X$/H] abundances. (Set solar_relative=false to use $A(X)$ abundances instead.) These override metallicity. This is the only way to specify an abundance of He that is non-solar.

Keyword arguments

solar_relative (default: true): When true, interpret abundances as being in [$X$/H] ($\log_{10}$ solar-relative) format. When false, interpret them as $A(X)$ abundances, i.e. $A(x) = \log_{10}(n_X/n_\mathrm{H}) + 12$, where $n_X$ is the number density of $X$. Note that abundances not specified default to the solar value, adjusted with metallicity, in either case.

solar_abundances (default: Korg.asplund_2020_solar_abundances) is the set of solar abundances to use, as a vector indexed by atomic number. Korg.asplund_2009_solar_abundances and Korg.grevesse_2007_solar_abundances are also provided for convienience.

Secondary functions

You don't need use these to synthesize spectra, but they might be relevant depending on what you are doing.

Postprocessing

These are used to transform observational or synthetic spectra.

Korg.rectify — Function

rectify(flux, wls; bandwidth=50, q=0.95, wl_step=1.0)

Rectify the spectrum with flux vector flux and wavelengths wls by dividing out a moving q-quantile with window size bandwidth. wl_step controls the size of the grid that the moving quantile is calculated on and interpolated from. Setting wl_step to 0 results in the exact calculation with no interpolation, but note that this is very slow.

Experiments on real spectra show an agreement between the interpolated rectified spectrum and the "exact" one (with default values) at the 3 × 10^-4 level.

Korg.constant_R_LSF — Function

constant_R_LSF(flux, wls, R; window_size=3)

Applies a gaussian line spread function the the spectrum with flux vector flux and wavelength vector wls with constant spectral resolution, $R = \lambda/\Delta\lambda$, where $\Delta\lambda$ is the LSF FWHM. The window_size argument specifies how far out to extend the convolution kernel in standard deviations.

For the best match to data, your wavelength range should extend a couple $\Delta\lambda$ outside the region you are going to compare.

Warning

This is a naive, slow implementation. Do not use it when performance matters.

constant_R_LSF will have weird behavior if your wavelength grid is not locally linearly-spaced. It is intended to be run on a fine wavelength grid, then downsampled to the observational (or otherwise desired) grid.

Line absorption

These functions can be used to directly compute line opacities.

Korg.line_absorption! — Function

line_absorption(linelist, λs, temp, nₑ, n_densities, partition_fns, ξ
               ; α_cntm=nothing, cutoff_threshold=1e-3, window_size=20.0*1e-8)

Calculate the opacity coefficient, α, in units of cm^-1 from all lines in linelist, at wavelengths λs [cm^-1].

other arguments:

temp the temerature in K

n_densities, a Dict mapping species to absolute number density [cm^-3].

partition_fns, a Dict containing the partition function of each species

ξ is the microturbulent velocity in cm/s (n.b. NOT km/s)

α_cntm is as a callable returning the continuum opacity as a function of wavelength. The window within which a line is calculated will extend to the wavelength at which the Lorentz wings or Doppler core of the line are at cutoff_threshold * α_cntm[line.wl], whichever is greater.

cuttoff_threshold (optional, default: 1e-3): see α_cntm

Korg.line_profile — Function

line_profile(λ₀, σ, γ, amplitude, λ)

A voigt profile centered on λ₀ with Doppler width σ (NOT √[2] σ, as the "Doppler width" is often defined) and Lorentz HWHM γ evaluated at λ (cm). Returns values in units of cm^-1.

Korg.hydrogen_line_absorption! — Function

hydrogen_line_absorption!(αs, λs, T, nₑ, nH_I, UH_I, ξ, window_size)

Calculate contribution to the the absorption coefficient, αs, from hydrogen lines in units of cm^-1, at wavelengths λs.

Uses profiles from Stehlé & Hutcheon (1999), which include Stark and Doppler broadening. For Halpha, Hbeta, and Hgamma, the p-d approximated profiles from Barklem, Piskunovet, and O'Mara 2000 are added to the absortion coefficient. This "convolution by summation" is inexact, but true convolution is expensive.

Arguments:

T: temperature [K]

nₑ: electron number density [cm^-3]

nH_I: neutral hydrogen number density [cm^-3]

UH_I: the value of the neutral hydrogen partition function

ξ: microturbulent velocity [cm/s]. This is only applied to Hα-Hγ. Other hydrogen lines profiles are dominated by stark broadening, and the stark broadened profiles are pre-convolved with a doppler profile.

window_size: the max distance from each line center [cm] at which to calculate the stark and self broadening profiles absorption for Hα-Hγ (those dominated by self-broadening).

Keyword arguments:

stark_profiles (default: Korg._hline_stark_profiles): tables from which to interpolate Stark profiles

Missing docstring.

Missing docstring for Korg.setup_hydrogen_stark_profiles. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Korg.voigt. Check Documenter's build log for details.

Continuum absorption

These function can be used to directly compute continuum absorption coefficients.

Continuum Absorption Kwargs

All bound-free and free-free absorption functions (other than absorption_coef_bf_TOPBase) support the following keyword arguments:

error_oobounds::Bool: specifies the function's behavior when it encounters invalid ν or T values. When false (the default), the function assumes that the absorption coefficient for these invalid ν or T values is 0.0. Otherwise, a DomainError is thrown when invalid ν or T values are encountered.

out_α::Union{Nothing,AbstractVector}: When this is nothing (the default case), the function will simply allocate a new vector to store the output continuum absorption coefficients. Alternatively, this can be a vector (with the same length as the function's ν argument). In this case, the function will directly add the computed continuum absorption to the elements of this vector, in-place (the vector is also returned by the function).

Korg.ContinuumAbsorption.total_continuum_absorption — Function

total_continuum_absorption(νs, T, nₑ, number_densities, partition_funcs; error_oobounds)

The total continuum linear absoprtion coefficient, α, at many frequencies, ν.

Arguments

νs are frequencies in Hz

T is temperature in K

nₑ is the electron number density in cm^-3

number_densities is a Dict mapping each Species to its number density

partition_funcs is a Dict mapping each Species to its partition function (e.g. Korg.partition_funcs)

error_oobounds::Bool specifies the behavior of most continnum absorption sources when passed frequencies or temperature values that are out of bounds for their implementation. When false (the default), those absorption sources are ignored at those values. Otherwise, an error is thrown.

Note

For efficiency reasons, νs must be sorted. While this function technically supports any sorted AbstractVector, it is most effient when passed an AbstractRange.

Korg.ContinuumAbsorption.H_I_bf — Function

H_I_bf(ν, T, nH_I_div_partition, [ion_energy], [nmax_explicit_sum], [integrate_high_n];
       kwargs...)

Compute the bound-free linear absorption coefficient contributed by all energy states of a neutral Hydrogen atom.

Required Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I_div_partition is the total number density of neutral Hydrogen divided by the its partition function.

Optional Arguments

ion_energy: The ionization energy of Hydrogen. By default, this is set to the values loaded into the global ionization_energies list.

nmax_explicit_sum: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in an integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function wraps hydrogenic_ff_absorption. See that function for implementation details.

Missing docstring.

Missing docstring for Korg.ContinuumAbsorption.H_I_ff. Check Documenter's build log for details.

Korg.ContinuumAbsorption.Hminus_bf — Function

Hminus_bf(ν, T, nH_I_div_partition, ne; kwargs...)

Compute the H⁻ bound-free linear absorption coefficient α, $\alpha_\nu = \sigma_{bf}(H^-) n(H⁻) (1 - \exp \left( \frac{-h\nu}{k T}\right))$

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I_div_partition: the total number density of H I divided by its partition function.

ne: the electron number density

This uses cross-sections from McLaughlin 2017. For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function assumes that n(H⁻) ≪ n(H I) + n(H II). The number density of n(H⁻) isn't precomputed as part of Korg's molecular equlibrium, it's computed here instead.

Note

The McLaughlin tabulated values (downloaded as a text file) contained a stray line out of monotonic order. A version of the data file with the line deleted is saved at data/McLaughlin2017Hminusbf.dat for archival purposes. (Korg doesn't read this file, it reads data/McLaughlin2017Hminusbf.h5.)

Korg.ContinuumAbsorption.Hminus_ff — Function

Hminus_ff(ν, T, nH_I_div_partition, ne; kwargs...)

Compute the H⁻ free-free linear absorption coefficient α

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + H I -> e⁻ + H I.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I_div_partition::Flt: the total number density of H I divided by its partition function.

ne: the number density of free electrons.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This is based on Table 1, in Bell & Berrington (1987), which tabulates values for "the H⁻ absorption coefficient", K (including the correction for stimulated emission). This quantity is in units of cm^4/dyn, and must be multiplied by the electron partial pressure and the ground-state neutral hydrogen number density to obtain a linear absorption coefficent, α.

The stipulation that the hydrogen should be ground-state only is based on the beginning of Section 2 in Bell and Berrington (1987), or alternately, Section 5.3 from Kurucz (1970). When Gray (2005) refers to this, it implicitly assumes that n(H I, n = 1) ≈ n(H I). Note that n(H I, n = 1) = n(H I)*gₙ₌₁/U(T)*exp(-Eₙ₌₁/(k*T)) = n(H I) * 2/U(T)*exp(0) = n(H I) * 2/U(T). Since U(T) ≈ 2 up to fairly large temperatures, this is not unreasonable.

Korg.ContinuumAbsorption.H2plus_bf_and_ff — Function

H2plus_bf_and_ff(ν, T, nH_I_div_partition, n_HII; kwargs...)

Compute the combined H₂⁺ bound-free and free-free linear absorption coefficient α using the tables from Stancil 1994. Note that this refers to interactions between free-protons and neutral hydrogen, not electrons and doubly ionized H2, i.e. the b-f interaction is photodissociation, not photoionization.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nH_I: the total number density of H I divided by its partition function.

nH_II: the number density of H II (not of H₂⁺).

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This computes the H₂⁺ number density from those of H I and H II, since ionized molecules are not included in the molecular equlibrium calculation. Stancil provides approximate equilibrium constants, K, for the molecule, but the Barklem and Collet values used elsewhere by Korg may be more reliable. Once ionized molecules are fully supported, those values should be used instead. While cross sections are tabulated down to only 3150 K, the cross sections could be linearly interpolated 1000 K or so lower, if reliable K values are availble.

Because n(H₂⁺) is computed on the fly, the "cross-sections" used (internally) by this function have units of cm^-5, since they must be multiplied by n(H I) and n(H II) to obtain absorption coefficients.

Stancil, Bates (1952), and Gray (2005) all use n(H I), but Kurucz (1970) notes in section 5.2 that the photodisociation of H₂⁺ produces a ground state H atom and that we should therefore use n(H I, n=1) instead. This difference causes a descrepancy of a fraction of a percent (at most) for T ≤ 1.2e4 K. Here we use n(H I).

Korg.ContinuumAbsorption.He_II_bf — Function

He_II_bf(ν, T, nHe_I_div_partition, [ion_energy], [nmax_explicit_sum], [integrate_high_n];
         kwargs...)

Compute the bound-free linear absorption coefficient contributed by all energy states of a singly-ionized Helium atom.

Required Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nHe_II_div_partition is the total number density of singly-ionized Helium divided by the its partition function.

Optional Arguments

ion_energy: The ionization energy of Helium. By default, this is set to the values loaded into the global ionization_energies list.

nmax_explicit_sum: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in an integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This function simply wraps hydrogenic_ff_absorption. See that function for additional details.

Missing docstring.

Missing docstring for Korg.ContinuumAbsorption.He_II_ff. Check Documenter's build log for details.

Korg.ContinuumAbsorption.Heminus_ff — Function

Heminus_ff(ν, T, nHe_I_div_partition, ne; kwargs...)

Compute the He⁻ free-free opacity κ.

The naming scheme for free-free absorption is counter-inutitive. This actually refers to the reaction: photon + e⁻ + He I -> e⁻ + He I.

Arguments

ν::AbstractVector{<:Real}: sorted frequency vector in Hz

T: temperature in K

nHe_I_div_partition: the total number density of H I divided by its partition function.

ne: the number density of free electrons.

For a description of the kwargs, see Continuum Absorption Kwargs.

Notes

This uses the tabulated values from John (1994). The quantity K is the same used by Bell and Berrington (1987). See Hminus_ff for an explanation.

According to John (1994), improved calculations are unlikely to alter the tabulated data for λ > 1e4Å, "by more than about 2%." The errors introduced by the approximations for 5.06e3 Å ≤ λ ≤ 1e4 Å "are expected to be well below 10%."

Korg.ContinuumAbsorption.electron_scattering — Function

electron_scattering(nₑ)

Compute the linear absorption coefficient, α, from scattering off of free electrons. This has no wavelength dependence. It assumes isotropic scattering. (See, e.g. Gray p 160.)

Arguments

nₑ::F: number density of free electrons (in cgs)

Korg.ContinuumAbsorption.rayleigh — Function

rayleigh(λs, nH_I, nHe_II, nH2)

Absorption coefficient from Rayleigh scattering by neutral H, He, and H2. Formulations for H and He are via Colgan+ 2016. Formulation for H2 from Dalgarno and Williams 1962.

The Dalgarno and Williams H2 is applicable redward of 1300 Å. Since Rayleigh scattering breaks down when the particle size to wavelength ratio gets large, we that all frequencies passed to this function be equivalent to 1300 Å or greater.

The formulations for H is adapted from Lee 2005, which states that it is applicable redward of Lyman alpha. See Colgan 2016 for details on He.

Korg.ContinuumAbsorption.hydrogenic_bf_absorption — Function

hydrogenic_bf_absorption(ν, T, Z, nmax_explicit_sum, nsdens_div_partition,
                         ion_energy, [integrate_high_n])

Compute the bound-free linear absorption coefficient contributed by all energy states of a Hydrogenic species

The calculation is broken into 2 parts: (i) the contributions of the lowest energy states are explicitly summed and (ii) the contributions of the higher energy states are estimated with an integral.

Arguments

ν: frequency in Hz

T: temperature in K

Z::Integer: Z is the atomic number of the species (e.g. 1 for H I or 2 for He II)

nsdens_div_partition is the total number density of the species divided by the species's partition function.

ion_energy: the ionization energy from the ground state (in eV). This can be estimated as Z²*RydbergH (RydbergH is the ionization energy of Hydrogen)

nmax_explicit_sum::Integer: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in the integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

Notes

This follows the approach described in section 5.1 of Kurucz (1970).

Korg.ContinuumAbsorption.hydrogenic_ff_absorption — Function

hydrogenic_ff_absorption(ν, T, Z, ni, ne)

computes the free-free linear absorption coefficient for a hydrogenic species

The naming convention for free-free absorption is counter-intuitive. A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). In practice, this means that ni should refer to:

the number density of H II if computing the H I free-free absorption

the number density of He III if computing the He II free-free absorption

the number density of Li IV if computing the Li III free-free absorption

Arguments

Z::Integer: the charge of the ion. For example, this is 1 for ionized H.

ni: the number density of the ion species in cm⁻³.

ne: the number density of free electrons.

ν: frequency in Hz

T: temperature in K

Note

This approach was adopted from equation 5.8 from section 5.1 of Kurucz (1970). Comparison against equation 5.18b of Rybicki & Lightman (2004), reveals that the equation in Kurucz (1970) omits the dependence on ρ. According to Rybicki & Lightman (2004) the free-free absorption coefficient (corrected for stimulated emission) is:

    α = coef * Z² * ne * ni * (1 - exp(-hplanck*ν/(kboltz*T))) * g_ff / (sqrt(T) * ν³)

Note that the g_ff is the free-free gaunt factor and coef is ∼3.7e8 (a more exact coefficient can be computed from eqn 5.18a).

With this in mind, equation 5.8 of Kurucz (1970) should actually read

    κ = ne * n(H II) * F_ν(T) * (1 - exp(-hplanck*ν/(kboltz*T))) / ρ

where Fν(T) = coef * Z² * gff / (sqrt(T) * ν³).

See gaunt_ff_vanHoof for details about where our gaunt factor data comes from. For simplicity, we enforce temperature and ν bounds constraints that don't include all of the available gaunt factor data. In practice, this should never be a concern for stellar spectroscopy.

Statistical mechanics

These functions can be used to calculate the number densities of all species in a given atospheric layer, or other statmech calculations.

Korg.saha_ion_weights — Function

saha_ion_weights(T, nₑ, atom, ionization_energies, partition_functions)

Returns (wII, wIII), where wII is the ratio of singly ionized to neutral atoms of a given element, and wIII is the ration of doubly ionized to neutral atoms.

arguments:

temperature T [K]

electron number density nₑ [cm^-3]

atom, the atomic number of the element

ionization_energies is a collection indexed by integers (e.g. a Vector) mappping elements' atomic numbers to their first three ionization energies

partition_funcs is a Dict mapping species to their partition functions

Korg.molecular_equilibrium_equations — Function

molecular_equilibrium_equations(absolute_abundances, ironization_energies, partiation_fns, equilibrium_constants)

Returns a NamedTuple representing the system of equations specifying molecular equilibrium.

arguments:

A Dict of absolute_abundances, NX/Ntotal

a Dict of ionization energies, ionization_energies. The keys of act as a list of all atoms.

a Dict of partition functions, partition_fns

a Dict of log molecular equilibrium constants, equilibrium_constants, in partial pressure form. The keys of equilibrium_constants act as a list of all molecules.

The system of equations is specified with the number densities of the neutral atoms as free parameters. Each equation specifies the conservation of a particular species, e.g. (simplified)

n(O) = n(CO) + n(OH) + n(O I) + n(O II) + n(O III).

In this equation:

n(O), the number density of oxygen atoms in any form comes absolute_abundances and the total

number density (supplied later)

n(O I) is a free parameter. The numerical solver is varying this to satisfy the system of equations.

n(O II), and n(O III) come from the Saha (ionization) equation given n(O I)

n(CO) and n(OH) come from the molecular equilibrium constants K, which are precomputed over a range of temperatures.

Equilibrium constants are defined in terms of partial pressures, so e.g.

K(OH)  ==  (p(O) p(H)) / p(OH)  ==  (n(O) n(H)) / n(OH)) kT

This could also be computed via a Saha equation involving the molecular partition function and dissolution energy.

Korg.molecular_equilibrium — Function

molecular_equilibrium(MEQS, T, nₜ, nₑ; x0)

Iteratively solve for the number density of each species. Returns a Dict mapping species to number densities.

arguments:

the system of molecular equilibrium equations MEQs (the thing returned by

molecular_equilibrium_equations)

the temperature T,

the number density of non-electron particles nₜ

the electron number density nₑ

optionally, x0, a starting point for the solver

Misc

Korg.blackbody — Function

blackbody(T, λ)

The value of the Planck blackbody function for temperature T at wavelength λ [cm].

Missing docstring.

Missing docstring for Korg.get_absolute_abundances. Check Documenter's build log for details.

Korg.air_to_vacuum — Function

air_to_vacuum(λ; cgs=λ<1)

Convert λ from an air to vacuum. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

See also: vacuum_to_air.

Korg.vacuum_to_air — Function

vacuum_to_air(λ; cgs=λ<1)

convert λ from a vacuum to air. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

See also: air_to_vacuum.