Developer documentation

This page contains info for people interested in contributing code to Korg. If you have questions, do not hesitate to ask.

Julia development

If you've never developed a package in Julia, here are some tips.

Once you have a local copy of Korg, you can make it importable locally by deving it from the Julia environment in which you would like to run it: Pkg.dev("/path/to/Korg"). If you run Pkg.dev("Korg") instead, Julia will automatically clone the repo to ~/.julia/dev/Korg. (See the Pkg documentation for details.)

When working on your local copy of Korg, Revise is easiest way to make and test changes without constantly restarting your Julia session.

To run the test suite locally, start a Julia session in the Korg root directory and run ]activate . then test. This will automatically run test/runtests.jl in the test environment.

Documentation is generated with Documentor. Continuous integration on github will ensure that the documentation is generated without errors, but it won't catch all formatting problems. If you wish to generate documentation locally in order to check that everything is as expected, start a Julia session in Korg/docs, activate the test environment (]activate .), and run instantiate. This downloads and installs the docs dependences, and will only have to be run once. To generate documentation, run julia --project make.jl on the command line (the --project flag activates the local environment). The generated docs can be served from docs/build.

Code guidelines

Try to be explicit about units throughout the code, particularly when not using CGS.

Whenever possible, calculations should be precise up to a factor of $10^{-3}$. When it's easy and inexpensive, they should be precise to $10^{-5}$ or better. There are exceptions to this (e.g. the Voigt function), but ideally they will eventually be vanquished.

Ensure types are generic enough to support dual numbers and autodifferentiation.

Limit lines to 100 characters.

Use ASCII characters in the names of functions that are part of the public API.

Unless they will never be called elsewhere, provide docstrings describing the inputs, assumptions and outputs of any functions you write.

Continuum absorption

Steps for implementing new continuum sources of absorption:

Define a helper function that computes a single absorption coefficient (in units of cm⁻²). The function should accept ν (in Hz) and T (in K) as the first and second arguments, respectively. The convention is for it should share a name with the corresponding public function, but have an underscore pre-appended (e.g. we define _H_I_bf to help implement H_I_bf).

The public function is the function constructed and returned by Korg.ContinuumAbsorption.bounds_checked_absorption that wraps the above helper function. This wrapper function implements bounds-checking for ν and T and supports the keyword arguments described in Continuum Absorption Kwargs.

Add a docstring describing the new function. At the very least, please describe any non-standard arguments and include a reference in the docstring to the source where the function was taken from.

Add a line to doc/src/API.md under the Continuum absorption heading to render the docstring of your new function.

Add a line to total_continuum_absorption that calls the new public function for absorption.

The first two steps may not apply for sources that don't directly depend on ν and T (e.g. absorption from scattering).

Complete API

Here are all the documented methods in Korg.

Korg.Formula — Type

Represents an atom or molecule, irespective of its charge.

Korg.Line — Method

Line(wl::F, log_gf::F, species::Species, E_lower::F, 
     gamma_rad::Union{F, Missing}=missing, gamma_stark::Union{F, Missing}=missing, 
     vdw::Union{F, Tuple{F, F}, Missing}, missing) where F <: Real

Arguments:

wl: wavelength, in cm

log_gf: (log base 10) oscillator strength (unitless)

species: the Species associated with the line

E_lower: The energy (excitiation potential) of the lower energy level (eV)

Optional Arguments (these override default recipes):

gamma_rad: Fundemental width

gamma_stark: Stark broadening width at 10,000 K (s⁻¹)

vdW: Either the van der Waals broadening width at 10,000 K (s⁻¹) or a Tuple, (σ, α) from ABO theory.

Note the the "gamma" values here are FWHM, not HWHM, of the Lorenztian component of the line profile.

Construct a Line. If any of gamma_rad, gamma_stark, or vdW are missing, guess them. vdW may be log(ΓvdW) (assumed if negative), ΓvdW (assumed if 0 < vdW < 1), or packed ABO parameters (assumed if vdW > 1). It may also be passed as a Tuple, (σ, α).

Korg.PlanarAtmosphere — Method

PlanarAtmosphere(atm::ShellAtmosphere)

Construct a planar atmosphere with the data from a shell atmosphere. Mostly useful for testing.

Korg.ShellAtmosphere — Method

ShellAtmosphere(atm::PlanarAtmosphere, R)

Construct a shell atmosphere with the data from a planar atmosphere and an outer radius. Mostly useful for testing.

Korg.Species — Type

Represents an atom or molecule (a Formula) with a particular number of electrons (regardless of their configuration).

Korg.Species — Method

Species(code::AbstractString)

Parse the "species code" in many of the forms in which it is often specifieds and return an object representing the sepcies.

Examples

"H I" -> H I

"H 1" -> H I

"H 1" -> H I

"H_1" -> H I

"H.I" -> H I

"H 2" -> H II

"H2" -> H₂

"H" -> H I

"01.00" → H I

"02.01" → He II

"02.1000" → He II

"0608" → CO I

Note

To parse at compile time, use the species string macro, i.e. species"H I". This is important in hot inner loops.

Korg.air_to_vacuum — Method

air_to_vacuum(λ; cgs=λ<1)

Convert λ from an air to vacuum. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

See also: vacuum_to_air.

Korg.all_atomic_species — Method

all_atomic_species()

Returns an iterator that runs over all atomic species supported by Korg.

Korg.approximate_gammas — Method

approximate_gammas(wl, species, E_lower; ionization_energies=Korg.ionization_energies)

A simplified form of the Unsoeld (1955) approximation for van der Waals broadening and the Cowley 1971 approximation for Stark broadening, evaluated at 10,000 K. Used for atomic lines with no vdW and stark broadening info in the linelist.

Returns (γ_stark, log10(γ_vdW)) in Hz, where these are the per-perturber quantities. For autoionizing lines (those for which Eupper > χ), Returns 0.0 for γvdW. Note the the "gamma" values here are FWHM, not HWHM, of the Lorenztian component of the line profile.

In the calculation of n*², uses the approximation that $\overbar{r^2} = 5/2 {n^*}^4 / Z^2$ which neglects the dependence on the angular momentum quantum number, l, in the the form given by Warner 1967 (the earliest english work reporting the Unsoeld result).

Korg.approximate_radiative_gamma — Method

approximate_radiative_gamma(wl, log_gf)

Approximate radiate broadening parameter.

Korg.blackbody — Method

blackbody(T, λ)

The value of the Planck blackbody function for temperature T at wavelength λ [cm].

Korg.constant_R_LSF — Method

constant_R_LSF(flux, wls, R; window_size=3)

Applies a gaussian line spread function the the spectrum with flux vector flux and wavelength vector wls with constant spectral resolution, $R = \lambda/\Delta\lambda$, where $\Delta\lambda$ is the LSF FWHM. The window_size argument specifies how far out to extend the convolution kernel in standard deviations.

For the best match to data, your wavelength range should extend a couple $\Delta\lambda$ outside the region you are going to compare.

Warning

This is a naive, slow implementation. Do not use it when performance matters.

constant_R_LSF will have weird behavior if your wavelength grid is not locally linearly-spaced. It is intended to be run on a fine wavelength grid, then downsampled to the observational (or otherwise desired) grid.

Korg.contained — Method

contained(value, interval)

Returns whether value is contained by interval.

Examples

julia> contained(0.5, Interval(1.0,10.0))
false
julia> contained(5.0, Interval(1.0,10.0))
true

Korg.contained_slice — Method

contained_slice(vals, interval)

Returns a range of indices denoting the elements of vals (which are assumed to be sorted in increasing order) that are contained by interval. When no entries are contained by interval, this returns (1,0) (which is a valid empty slice).

Korg.doppler_width — Method

doppler_width(λ₀ T, m, ξ)

The standard deviation of of the doppler-broadening profile. In standard spectroscopy texts, the Doppler width often refers to σ√2, but this is σ

Korg.format_A_X — Function

format_A_X(metallicity, abundances; kwargs... )

Returns a 92 element vector containing abundances in $A(X)$ ($\log_{10}(X/H) + 12$) format for elements from hydrogen to uranium.

Arguments

You can provide either or both of:

metallicity (default: 0), i.e. [metals/H] is the $\log_{10}$ solar-relative abundance of elements heavier than He. It is overriden by abundances, on a per-element basis.

abundances is a Dict mapping atomic numbers or symbols to [$X$/H] abundances. (Set solar_relative=false to use $A(X)$ abundances instead.) These override metallicity. This is the only way to specify an abundance of He that is non-solar.

Keyword arguments

solar_relative (default: true): When true, interpret abundances as being in [$X$/H] ($\log_{10}$ solar-relative) format. When false, interpret them as $A(X)$ abundances, i.e. $A(x) = \log_{10}(n_X/n_\mathrm{H}) + 12$, where $n_X$ is the number density of $X$. Note that abundances not specified default to the solar value, adjusted with metallicity, in either case.

solar_abundances (default: Korg.asplund_2020_solar_abundances) is the set of solar abundances to use, as a vector indexed by atomic number. Korg.asplund_2009_solar_abundances and Korg.grevesse_2007_solar_abundances are also provided for convienience.

Korg.get_mass — Method

get_mass(f::Formula)

Returns the mass [g] of f.

Korg.hummer_mihalas_U_H — Method

hummer_mihalas_U_H(T, nH, nHe, ne)

!!!note This is experimental, and not used by Korg for spectal synthesis.

Calculate the partition function of neutral hydrogen using the occupation probability formalism from Hummer and Mihalas 1988. See hummer_mihalas_w for details.

Korg.hummer_mihalas_w — Method

hummer_mihalas_w(T, n, E_level, nH, nHe, ne)

!!!note This is experimental, and not used by Korg for spectal synthesis.

Calculate the corection, w, to the occupation fraction of a hydrogen energy level using the occupation probability formalism from Hummer and Mihalas 1988, optionally with the generalization by Hubeny+ 1994. (Sometimes Daepen+ 1987 is cited instead, but H&M seems to be where the theory originated. Presumably it was delayed in publication.)

The expression for w is in equation 4.71 of H&M. K, the QM correction used in defined in equation 4.24. Note that H&M's "N"s are numbers (not number densities), and their "V" is volume. These quantities apear only in the form N/V, so we use the number densities instead.

This is based partially on Paul Barklem and Kjell Eriksson's WCALC fortran routine (part of HBOP.f), which is used by (at least) Turbospectrum and SME. As in that routine, we do consider hydrogen and helium as the relevant neutral species, and assume them to be in the ground state. All ions are assumed to have charge 1. Unlike that routine, the generalization to the formalism from Hubeny+ 1994 is turned off by default because I haven't closely checked it. The difference effects the chargedterm only, and temerature is only used when `usehubeny_generalizationis set totrue`.

Korg.hydrogen_line_absorption! — Method

hydrogen_line_absorption!(αs, λs, T, nₑ, nH_I, UH_I, ξ, window_size)

Calculate contribution to the the absorption coefficient, αs, from hydrogen lines in units of cm^-1, at wavelengths λs.

Uses profiles from Stehlé & Hutcheon (1999), which include Stark and Doppler broadening. For Halpha, Hbeta, and Hgamma, the p-d approximated profiles from Barklem, Piskunovet, and O'Mara 2000 are added to the absortion coefficient. This "convolution by summation" is inexact, but true convolution is expensive.

Arguments:

T: temperature [K]

nₑ: electron number density [cm^-3]

nH_I: neutral hydrogen number density [cm^-3]

UH_I: the value of the neutral hydrogen partition function

ξ: microturbulent velocity [cm/s]. This is only applied to Hα-Hγ. Other hydrogen lines profiles are dominated by stark broadening, and the stark broadened profiles are pre-convolved with a doppler profile.

window_size: the max distance from each line center [cm] at which to calculate the stark and self broadening profiles absorption for Hα-Hγ (those dominated by self-broadening).

Keyword arguments:

stark_profiles (default: Korg._hline_stark_profiles): tables from which to interpolate Stark profiles

Korg.inverse_gaussian_density — Method

inverse_gaussian_density(ρ, σ)

Calculate the inverse of a (0-centered) Gaussian PDF with standard deviation σ, i.e. the value of x for which ρ = exp(-0.5 x^2/σ^2}) / √[2π], which is given by σ √[-2 log (√[2π]σρ)]. Returns 0 when ρ is larger than any value taken on by the PDF.

See also: inverse_lorentz_density.

Korg.inverse_lorentz_density — Method

inverse_lorentz_density(ρ, γ)

Calculate the inverse of a (0-centered) Lorentz PDF with width γ, i.e. the value of x for which ρ = 1 / (π γ (1 + x^2/γ^2)), which is given by √[γ/(πρ) - γ^2]. Returns 0 when ρ is larger than any value taken on by the PDF.

See also: inverse_gaussian_density.

Korg.ismolecule — Method

ismolecule(f::Formula)

true when f is composed of more than one atom

Korg.line_absorption! — Method

line_absorption(linelist, λs, temp, nₑ, n_densities, partition_fns, ξ
               ; α_cntm=nothing, cutoff_threshold=1e-3, window_size=20.0*1e-8)

Calculate the opacity coefficient, α, in units of cm^-1 from all lines in linelist, at wavelengths λs [cm^-1].

other arguments:

temp the temerature in K

n_densities, a Dict mapping species to absolute number density [cm^-3].

partition_fns, a Dict containing the partition function of each species

ξ is the microturbulent velocity in cm/s (n.b. NOT km/s)

α_cntm is as a callable returning the continuum opacity as a function of wavelength. The window within which a line is calculated will extend to the wavelength at which the Lorentz wings or Doppler core of the line are at cutoff_threshold * α_cntm[line.wl], whichever is greater.

cuttoff_threshold (optional, default: 1e-3): see α_cntm

Korg.line_profile — Method

line_profile(λ₀, σ, γ, amplitude, λ)

A voigt profile centered on λ₀ with Doppler width σ (NOT √[2] σ, as the "Doppler width" is often defined) and Lorentz HWHM γ evaluated at λ (cm). Returns values in units of cm^-1.

Korg.load_atomic_partition_functions — Function

load_atomic_partition_functions()

Loads saved tabulated values for atomic partition functions from disk. Returns a dictionary mapping species to interpolators over log(T).

Korg.molecular_equilibrium — Method

molecular_equilibrium(MEQS, T, nₜ, nₑ; x0)

Iteratively solve for the number density of each species. Returns a Dict mapping species to number densities.

arguments:

the system of molecular equilibrium equations MEQs (the thing returned by

molecular_equilibrium_equations)

the temperature T,

the number density of non-electron particles nₜ

the electron number density nₑ

optionally, x0, a starting point for the solver

Korg.molecular_equilibrium_equations — Method

molecular_equilibrium_equations(absolute_abundances, ironization_energies, partiation_fns, equilibrium_constants)

Returns a NamedTuple representing the system of equations specifying molecular equilibrium.

arguments:

A Dict of absolute_abundances, NX/Ntotal

a Dict of ionization energies, ionization_energies. The keys of act as a list of all atoms.

a Dict of partition functions, partition_fns

a Dict of log molecular equilibrium constants, equilibrium_constants, in partial pressure form. The keys of equilibrium_constants act as a list of all molecules.

The system of equations is specified with the number densities of the neutral atoms as free parameters. Each equation specifies the conservation of a particular species, e.g. (simplified)

n(O) = n(CO) + n(OH) + n(O I) + n(O II) + n(O III).

In this equation:

n(O), the number density of oxygen atoms in any form comes absolute_abundances and the total

number density (supplied later)

n(O I) is a free parameter. The numerical solver is varying this to satisfy the system of equations.

n(O II), and n(O III) come from the Saha (ionization) equation given n(O I)

n(CO) and n(OH) come from the molecular equilibrium constants K, which are precomputed over a range of temperatures.

Equilibrium constants are defined in terms of partial pressures, so e.g.

K(OH)  ==  (p(O) p(H)) / p(OH)  ==  (n(O) n(H)) / n(OH)) kT

This could also be computed via a Saha equation involving the molecular partition function and dissolution energy.

Korg.move_bounds — Method

move_bounds(λs, lb, ub, λ₀, window_size)

Using lb and ub as initial guesses, return the indices of λs corresponding to λ₀±window_size. Ifλsis anAbstractRange, then compute them directly. Assumesλs` is sorted.

Korg.parse_fwf — Method

parse_fwf(source, rowspec; datarow, lastrow)

Parse fixed-width files, returning a vector of NamedTuples. source can be either an AbstractVector of Strings or a filename. rowspec should be a vector of tuples each containing, in order

a range object corresponding to the column postion

the type. Strings will be pulled directly & other types will be parsed with parse.

the column name, a Symbol.

optionally, a function to apply to the parsed value.

Korg.read_Barklem_Collet_table — Method

functon read_Barklem_Collet_table([transform=identity])

Constructs a Dict holding tables containing partition function or equilibrium constant values across ln(temperature).

Korg.read_linelist — Method

read_linelist(filename; format="vald", isotopic_abundances=Korg.isotopic_abundances)

Parse a linelist file, returning a vector of Lines.

Pass format="kurucz" for a Kurucz linelist (format=kurucz_vac if it uses vacuum wavelengths; Be warned that Korg will not assume that wavelengths are vacuum below 2000 Å),format="vald" for a VALD linelist, and format="moog" for a MOOG linelist (doesn't support broadening parameters or dissociation energies).

VALD linelists (the default and preferred format) can be either "short" or "long" format, "extract all" or "extract stellar". Air wavelengths will automatically be converted into vacuum wavelengths, and energy levels will be automatically converted from cm$^{-1}$ to eV.

When they are not pre-scaled by isotopic abundace (which VALD does by default), Korg will automatically adjust the log_gf of each line according to isotopic_abundances, which defaults to the values from NIST. To use custom isotopic abundances, just pass isotopic_abundances as a dictionary mapping (atomic number, atomic weight) pairs to abundances between 0 and 1.

Korg.read_model_atmosphere — Method

read_model_atmosphere(filename)

Parse the provided model atmosphere file in MARCS ".mod" format. Returns either a PlanarAtmosphere or a ShellAtmosphere.

Warning

Korg does not yet support cool ($\lesssim$ 3500 K) stars. While it will happily parse their model atmospheres, it is very likely to crash when you feed them into synthesize.

Korg.rectify — Method

rectify(flux, wls; bandwidth=50, q=0.95, wl_step=1.0)

Rectify the spectrum with flux vector flux and wavelengths wls by dividing out a moving q-quantile with window size bandwidth. wl_step controls the size of the grid that the moving quantile is calculated on and interpolated from. Setting wl_step to 0 results in the exact calculation with no interpolation, but note that this is very slow.

Experiments on real spectra show an agreement between the interpolated rectified spectrum and the "exact" one (with default values) at the 3 × 10^-4 level.

Korg.saha_ion_weights — Method

saha_ion_weights(T, nₑ, atom, ionization_energies, partition_functions)

Returns (wII, wIII), where wII is the ratio of singly ionized to neutral atoms of a given element, and wIII is the ration of doubly ionized to neutral atoms.

arguments:

temperature T [K]

electron number density nₑ [cm^-3]

atom, the atomic number of the element

ionization_energies is a collection indexed by integers (e.g. a Vector) mappping elements' atomic numbers to their first three ionization energies

partition_funcs is a Dict mapping species to their partition functions

Korg.scaled_stark — Method

the stark broadening gamma scaled acording to its temperature dependence

Korg.scaled_vdW — Function

scaled_vdW(vdW, m, T)

The vdW broadening gamma scaled acording to its temperature dependence, using either simple scaling or ABO. See Anstee & O'Mara (1995) or https://www.astro.uu.se/~barklem/howto.html for the definition of the ABO γ.

vdW should be either γ_vdW evaluated at 10,000 K, or tuple containing the ABO params (σ, α). The species mass, m, is ignored in the former case.

Korg.setup_equilibrium_constants — Method

setup_equilibrium_constants()

Returns a Dict holding the default (Barklem & Collet 2016) log equilibrium constants, which are in pressure units.

In equation 7 in this paper, m is the reduced mass, m₁m₂/(m₁ + m₂). This can be verified by re-deriving the equilibrium constants from the partition functions and dissolution constants provided by the paper.

Korg.setup_ionization_energies — Function

setup_ionization_energies([filename])

Parses the table of ionization energies and returns it as a dictionary mapping elements to their ionization energies, [χ₁, χ₂, χ₃] in eV.

Korg.setup_partition_funcs — Method

setup_partition_funcs()

Returns a Dict holding the default partition functions. These are custom for atoms and via Barklem & Collet 2016 for molecules.

The partition function is linearly interpolated from values tabulated for the paper. The partition functions can only be evaluated at temperatures between 1e-5 K and 1e4 K. The tabulated values were NOT computed with a cutoff principle quantum number. This can lead to a couple percent error for some neutral species in solar type stars (I think the problem becomes worse in hotter stars). See the paper for more details.

Korg.sigma_line — Method

sigma_line(wl)

The cross-section (divided by gf) at wavelength wl in Ångstroms of a transition for which the product of the degeneracy and oscillator strength is 10^log_gf.

Korg.synthesize — Function

synthesize(atm, linelist, A_X, λ_start, λ_stop, [λ_step=0.01]; kwargs... )

Compute a synthetic spectrum.

Arguments

atm: the model atmosphere (see read_model_atmosphere)

linelist: A vector of [Line](@ref)s (see read_linelist)

A_X: a vector containing the A(X) abundances (log(X/H) + 12) for elements from hydrogen to uranium. (see format_A_X)

λ_start: the lower bound (in Å) of the region you wish to synthesize.

λ_stop: the upper bound (in Å) of the region you wish to synthesize.

λ_step (default: 0.01): the (approximate) step size to take (in Å).

Returns

A named tuple with keys:

flux: the output spectrum

alpha: the linear absorption coefficient at each wavelenth and atmospheric layer a Matrix of size (layers x wavelengths)

number_densities: A dictionary mapping Species to vectors of number densities at each atmospheric layer

wavelengths: The vacuum wavelenths (in Å) over which the synthesis was performed. If air_wavelengths=true this will not be the same as the input wavelenths.

Example

to synthesize a spectrum between 5000 Å and 5100 Å, with all metal abundances set to 0.5 dex less than the solar value except carbon, except carbon, which we set to [C/H]=-0.25:

atm = read_model_atmosphere("path/to/atmosphere.mod")
linelist = read_linelist("path/to/linelist.vald")
A_X = format_A_X(-0.5, Dict("C" => -0.25))
solution = synthesize(atm, linelist, A_X, 5000, 5100)

Optional arguments:

vmic (default: 0) is the microturbulent velocity, $\xi$, in km/s.

air_wavelengths (default: false): Whether or not the input wavelengths are air wavelenths to be converted to vacuum wavelengths by Korg. The conversion will not be exact, so that the wavelenth range can internally be represented by an evenly-spaced range. If the approximation error is greater than wavelength_conversion_warn_threshold, an error will be thrown. (To do wavelength conversions yourself, see air_to_vacuum and vacuum_to_air.)

wavelength_conversion_warn_threshold (default: 1e-4): see air_wavelengths. (In Å.)

line_buffer (default: 10): the farthest (in Å) any line can be from the provided wavelenth range before it is discarded. If the edge of your window is near a strong line, you may have to turn this up.

cntm_step (default 1): the distance (in Å) between point at which the continuum opacity is calculated.

hydrogen_lines (default: true): whether or not to include H lines in the synthesis.

hydrogen_line_window_size (default: 150): the mamximum distance (in Å) from each hydrogen line center at which to calculate its contribution to the total absorption coefficient.

n_mu_points (default: 20): the number of μ values at which to calculate the surface flux when doing transfer in spherical geometry (when atm is a ShellAtmosphere). 20 points is sufficient for accuracy at the 10^-3 level.

line_cutoff_threshold (default: 3e-4): the fraction of the continuum absorption coefficient at which line profiles are truncated. This has major performance impacts, since line absorption calculations dominate more syntheses. Turn it down for more precision at the expense of runtime. The default value should effect final spectra below the 10^-3 level.

ionization_energies, a Dict mapping Species to their first three ionization energies, defaults to Korg.ionization_energies.

partition_funcs, a Dict mapping Species to partition functions (in terms of ln(T)). Defaults to data from Barklem & Collet 2016, Korg.partition_funcs.

equilibrium_constants, a Dict mapping Species representing diatomic molecules to their molecular equilbrium constants in partial pressure form. Defaults to data from Barklem and Collet 2016, Korg.equilibrium_constants.

bezier_radiative_transfer (default: false): Use the radiative transfer scheme. This is for testing purposes only.

Korg.translational_U — Method

translational_U(m, T)

The (inverse) contribution to the partition function from the free movement of a particle. Used in the Saha equation.

arguments

m is the particle mass

T is the temperature in K

Korg.vacuum_to_air — Method

vacuum_to_air(λ; cgs=λ<1)

convert λ from a vacuum to air. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

See also: air_to_vacuum.

Korg.voigt_hjerting — Method

voigt_hjerting(α, v)

The Hjerting function, $H$, somtimes called the Voigt-Hjerting function. $H$ is defined as H(α, v) = ∫^∞_∞ exp(-y^2) / ((u-y)^2 + α^2) dy (see e.g. the unnumbered equation after Gray equation 11.47). It is equal to the ratio of the absorption coefficient to the value of the absorption coefficient obtained at the line center with only Doppler broadening.

If x = λ-λ₀, Δλ_D = σ√2 is the Doppler width, and Δλ_L = 4πγ is the Lorentz width,

voigt(x|Δλ_D, Δλ_L) = H(Δλ_L/(4πΔλ_D), x/Δλ_D) / (Δλ_D√π)
                    = H(γ/(σ√2), x/(σ√2)) / (σ√(2π))

Approximation from Hunger 1965.

Korg.λ_to_ν_bound — Method

λ_to_ν_bound(λ_bound)

Converts a λ Inverval (in cm) to an equivalent ν Interval (in Hz), correctly accounting for tricky floating point details at the bounds.

Korg.CubicSplines.CubicSpline — Method

CubicSpline(xs, ys; extrapolate=false)

Construct a interpolant using xs and ys as the knot coordinates. Assumes xs is sorted. Apply this object as a function to interpolate at any x value in the domain. If extrapolate is false, x values outside [xs[1], xs[end]] throw errors, if extrapolate is true, the interpolant uses flat extrapolation, i.e. it returns the extreme value.

Korg.CubicSplines.cumulative_integral! — Method

cumulative_integral!(out, A, t1, t2)

Given a curve described by the spine, A, Calculates the integral from t1 to t for all t = t1, t2, and all spline knots in between. So if t1 is A.t[1] and t2 is A.t[end], out should have the same length as A.t.

Korg.ContinuumAbsorption._hydrogenic_bf_high_n_absorption — Method

_hydrogenic_bf_high_n_absorption(ν, T, Z, nmin, nsdens_div_partition, ion_energy)

Approximates the sum of absorption coefficient contributions by all bound-free transitions in which the electron originates in an energy level of nmin or larger using an integral.

In full detail, the calculated quantity is described as ∑{n'=nmin}^∞ ndens(n=n') * αν(n=n') * (1 - exp(-hν/k/T))/ρ, where α(n=n') is the absorption coefficient for the bound-free atomic absorption coefficient (uncorrected for stimulated emission). This function approximates this sum with an integral.

Arguments

ν: frequency in Hz

T: temperature in K

Z::Integer: Z is the atomic number of the ion (1 for HI)

nmin::Integer: The lowest energy level (principle quantum number) included in the calculation

nsdens_div_partition is the number density of the current species divided by the partition function.

ion_energy: the ionization energy from the ground state (in eV).

Notes

This is based on equation (5.6) from Kurucz (1970) (which give the formula for opacity). I think ρ was simply omitted from that equation.

Korg.ContinuumAbsorption._load_gauntff_table — Function

_load_gauntff_table([fname])

Returns a table of thermally-averaged free-free Gaunt factors, and the values of log₁₀(γ²) and log₁₀(u) associated with each point.

This loads the non-relativistic free-free data published by van Hoof et al. (2014).

Note: This function code could trivially be adapted to load the relativistic free-free gaunt factors published by van Hoof et al (2015).

Korg.ContinuumAbsorption._ndens_Hminus — Function

_ndens_Hminus(nH_I_div_partition, ne, T, ion_energy = _H⁻_ion_energy)

Compute the number density of H⁻ (implements eqn 5.10 of Kurucz 1970). This is an application of the saha equation where the "ground state" is H⁻ and the "first ionization state" is H I. The partition function of H⁻ is 1 at all temperatures.

Korg.ContinuumAbsorption.bounds_checked_absorption — Method

bounds_checked_absorption(func; ν_bound, temp_bound)

Constructs a wrapped function that implements bounds checking and extrapolation

Parameters

func: a function that has a signature f(ν::Real, T::Real, args...)::Real, where ν is frequency (in Hz) and T is temperature (in K)

ν_bound::Interval: Interval of frequencies (in Hz) over which func is valid.

temp_bound::Interval: Interval of temperatures (in K) over which func is valid.

The resulting function will have this signature:

wrapped_func(ν::AbstractVector{<:Real}, T::Real, args...; kwargs...)

Wrapped Function Parameters

ν::AbstractVector{<:Real}: sorted vector of frequencies (in Hz)

T::Real: temperature (in K)

args...: function-specific arguments

For a description for the kwargs..., see Continuum Absorption Kwargs.

Korg.ContinuumAbsorption.electron_scattering — Method

electron_scattering(nₑ)

Compute the linear absorption coefficient, α, from scattering off of free electrons. This has no wavelength dependence. It assumes isotropic scattering. (See, e.g. Gray p 160.)

Arguments

nₑ::F: number density of free electrons (in cgs)

Korg.ContinuumAbsorption.gaunt_ff_vanHoof — Method

gaunt_ff_vanHoof(log_u, log_γ2)

computes the thermally averaged, non-relativistic free-free gaunt factor by interpolating the table provided by van Hoof et al. (2014).

Arguments

log_u: Equal to log₁₀(u) = log₁₀(hν/(kTₑ))

log_γ2: Equal to log₁₀(γ²) = log₁₀(RydbergZ²/(kTₑ))

Rydberg is the "infinite mass unit of energy" and Tₑ is the temperature of free electrons (for our purposes, we assume that free electrons are in thermal equilibrium with ions and neutral species).

Notes

van Hoof et al. (2014) computed the associated data table with a non-relativistic approach, which is invalid at very high temperatures. They conclude (from comparisons with a different paper) that their "results should be accurate up to electron temperatures of roughly 100 MK". This is more than adequate for stellar atmospheres. In van Hoof et al. (2015), they find that relativistic effects introduce a ∼0.75% at 100MK, for Z = 1 (when Z > 1, the change is smaller).

This function currently uses linear interpolation. However, van Hoof et al. (2014) provides an implementation of a third-order Lagrange scheme, which "reaches a relative precision better than 1.5e-4 everywhere." The C and Fortran implementations of this scheme can be found here, and are copyrighted by a BSD-style license.

Earlier variants of this function used less-accurate data from section 5.1 of Kurucz (1970) that extended over a smaller interval of data. That table was originally derived from a figure in Karsas and Latter (1961) and it's now used for testing purposes.

Korg.ContinuumAbsorption.hydrogenic_bf_absorption — Function

hydrogenic_bf_absorption(ν, T, Z, nmax_explicit_sum, nsdens_div_partition,
                         ion_energy, [integrate_high_n])

Compute the bound-free linear absorption coefficient contributed by all energy states of a Hydrogenic species

The calculation is broken into 2 parts: (i) the contributions of the lowest energy states are explicitly summed and (ii) the contributions of the higher energy states are estimated with an integral.

Arguments

ν: frequency in Hz

T: temperature in K

Z::Integer: Z is the atomic number of the species (e.g. 1 for H I or 2 for He II)

nsdens_div_partition is the total number density of the species divided by the species's partition function.

ion_energy: the ionization energy from the ground state (in eV). This can be estimated as Z²*RydbergH (RydbergH is the ionization energy of Hydrogen)

nmax_explicit_sum::Integer: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in the integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

Notes

This follows the approach described in section 5.1 of Kurucz (1970).

Korg.ContinuumAbsorption.hydrogenic_ff_absorption — Method

hydrogenic_ff_absorption(ν, T, Z, ni, ne)

computes the free-free linear absorption coefficient for a hydrogenic species

The naming convention for free-free absorption is counter-intuitive. A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). In practice, this means that ni should refer to:

the number density of H II if computing the H I free-free absorption

the number density of He III if computing the He II free-free absorption

the number density of Li IV if computing the Li III free-free absorption

Arguments

Z::Integer: the charge of the ion. For example, this is 1 for ionized H.

ni: the number density of the ion species in cm⁻³.

ne: the number density of free electrons.

ν: frequency in Hz

T: temperature in K

Note

This approach was adopted from equation 5.8 from section 5.1 of Kurucz (1970). Comparison against equation 5.18b of Rybicki & Lightman (2004), reveals that the equation in Kurucz (1970) omits the dependence on ρ. According to Rybicki & Lightman (2004) the free-free absorption coefficient (corrected for stimulated emission) is:

    α = coef * Z² * ne * ni * (1 - exp(-hplanck*ν/(kboltz*T))) * g_ff / (sqrt(T) * ν³)

Note that the g_ff is the free-free gaunt factor and coef is ∼3.7e8 (a more exact coefficient can be computed from eqn 5.18a).

With this in mind, equation 5.8 of Kurucz (1970) should actually read

    κ = ne * n(H II) * F_ν(T) * (1 - exp(-hplanck*ν/(kboltz*T))) / ρ

where Fν(T) = coef * Z² * gff / (sqrt(T) * ν³).

See gaunt_ff_vanHoof for details about where our gaunt factor data comes from. For simplicity, we enforce temperature and ν bounds constraints that don't include all of the available gaunt factor data. In practice, this should never be a concern for stellar spectroscopy.

Korg.ContinuumAbsorption.metal_bf_absorption! — Method

metal_bf_absorption!(α, νs, T, number_densities)

Adds to α the contributions of bf metal opacities. Uses precomputed tables from TOPBase for Li, Be, B, C, N, O, F, Ne, Na, Mg, Al, Si, S, Ar, and Ca. Uses tables from NORAD for Fe. For these elements, tables have been precomputed for the neutral and singly ionized species assuming and LTE distribution of energy levels. See Korg/data/metal_bf_cross-sections/ for the scripts which generate the tables.

Cross sections were computed for 100 K < T < 100,000 K and frequencies corresponding to 500 Å < λ < 30,000 Å. Outside of either of those ranges, flat extrapolation is used (i.e. the extreme value is returned).

Korg.ContinuumAbsorption.ndens_state_hydrogenic — Method

ndens_state_hydrogenic(n, nsdens_div_partition, ion_energy, T)

Calculates the LTE number density (in cm^-3) of a hydrogenic species at a given energy level.

Arguments

n::Integer: The quantum number of the state

nsdens_div_partition: The total (including all states) number density of the current

ionization species (in cm^-3) divided by the species's partition function.

ion_energy: The minimum energy (in eV) required to ionize the species (from the ground

state). This can be estimated as Z²ion_energy of hydrogen or Z²Rydberg_H.

T: Temperature

Note

This assumes that all hydrogenic species have a statistical weight of gₙ = 2*n².

This was taken from equation (5.4) of Kurucz (1970) (although this comes directly from the boltzmann equation).

Korg.ContinuumAbsorption.positive_ion_ff_absorption! — Method

positive_ion_ff_absorption!(α_out::Vector{Real}, ν::Real, T::Real, number_densities::Dict, 
                            ne::Real, departure_coefficients=Peach1970.departure_coefficients)

Computes the linear absorption coefficient (in cm⁻¹) for all free-free interactions involving positively charged atomic species. Uses the provided departure coefficients when they are available, and the uncorrected hydrogenic approximation when they are not.

Arguments

ν: frequency in Hz

T: temperature in K

number_densities is a Dict mapping each Species to its number density.

ne: the number density of free electrons.

departure_coefficients (optional, defaults to Peach1970): a dictionary mapping species to the departure coefficients for the ff process it participates in (e.g. species"C II" maps to the C III ff departure coefficients–see note below). Departure coefficients should be callables taking temperature and (photon energy / RydbergH / Zeff^2). See Peach1970 for details.

Note

A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). To make matters more confusing, some sources present the free-free linear absorption coefficient's formula in terms of the number density of the species in the interaction's name (by including the Saha equation in the formula).

To compute the absorption for a given free-free interaction, this function accesses elements from the dictionary arguments that are associated with the species that participates in the interaction. For example:

Si I ff absorption: uses the number density of Si II (number_densities[species"Si II"]) and checks departure_coefficients[species"Si II"] for departure coefficients.

Si II ff absorption: uses the number density of Si III (number_densities[species"Si III"]) and checks departure_coefficients[species"Si III"] for departure coefficients.

Korg.ContinuumAbsorption.rayleigh — Method

rayleigh(λs, nH_I, nHe_II, nH2)

Absorption coefficient from Rayleigh scattering by neutral H, He, and H2. Formulations for H and He are via Colgan+ 2016. Formulation for H2 from Dalgarno and Williams 1962.

The Dalgarno and Williams H2 is applicable redward of 1300 Å. Since Rayleigh scattering breaks down when the particle size to wavelength ratio gets large, we that all frequencies passed to this function be equivalent to 1300 Å or greater.

The formulations for H is adapted from Lee 2005, which states that it is applicable redward of Lyman alpha. See Colgan 2016 for details on He.

Korg.ContinuumAbsorption.total_continuum_absorption — Method

total_continuum_absorption(νs, T, nₑ, number_densities, partition_funcs; error_oobounds)

The total continuum linear absoprtion coefficient, α, at many frequencies, ν.

Arguments

νs are frequencies in Hz

T is temperature in K

nₑ is the electron number density in cm^-3

number_densities is a Dict mapping each Species to its number density

partition_funcs is a Dict mapping each Species to its partition function (e.g. Korg.partition_funcs)

error_oobounds::Bool specifies the behavior of most continnum absorption sources when passed frequencies or temperature values that are out of bounds for their implementation. When false (the default), those absorption sources are ignored at those values. Otherwise, an error is thrown.

Note

For efficiency reasons, νs must be sorted. While this function technically supports any sorted AbstractVector, it is most effient when passed an AbstractRange.

Korg.ContinuumAbsorption.Stancil1994 — Module

This module contains transcriptions of the tables from Stancil 1994, who calculated ff and bf absorption coefficients for H₂⁺ and He₂⁺.

It also contains pre-constructed interpolation objects.