This page contains info for people interested in contributing code to Korg. If you have questions, do not hesitate to ask.

Code guidelines

Try to be explicit about units throughout the code, particularly when not using CGS.

Whenever possible, calculations should be precise up to a factor of $10^{-3}$. When it's easy and inexpensive, they should be precise to $10^{-5}$ or better.

Ensure types are generic enough to support dual numbers and autodifferentiation.

Limit lines to 100 characters.

Use ASCII characters in the names of functions that are part of the public API.

Unless they will never be called elsewhere, provide docstrings describing the inputs, assumptions and outputs of any functions you write.

Continuum absorption

Steps for implementing new continuum sources of absorption:

Define a helper function that computes a single absorption coefficient (in units of cm⁻²). The function should accept ν (in Hz) and T (in K) as the first and second arguments, respectively. The convention is for it should share a name with the corresponding public function, but have an underscore pre-appended (e.g. we define _H_I_bf to help implement H_I_bf).

The public function is the function constructed and returned by Korg.ContinuumAbsorption.bounds_checked_absorption that wraps the above helper function. This wrapper function implements bounds-checking for ν and T and supports the keyword arguments described in Continuum Absorption Kwargs.

Add a docstring describing the new function. At the very least, please describe any non-standard arguments and include a reference in the docstring to the source where the function was taken from.

Add a line to doc/src/API.md under the Continuum absorption heading to render the docstring of your new function.

Add a line to total_continuum_absorption that calls the new public function for absorption.

The first two steps may not apply for sources that don't directly depend on ν and T (e.g. absorption from scattering).

Experimental support for bound-free metal absorption coefficients

Overview

Korg currently has experimental support for computing the absorption coefficient contribution from bound-free absorption by metals using data from the Opacity Project's online database, TOPbase.

For a given species (e.g. neutral C or singly-ionized C) TOPbase provides many tables of bound-free cross-sections (without stimulated emission correction), as functions of wavelength. The weighted_bf_cross_section_TOPBase can be used to compute the effective bound-free cross-section (including corrections from stimulated emission), assuming LTE, for the given species.

Our tentative plan for the future is to use weighted_bf_cross_section_TOPBase for each species to construct a 2D tables of the effective bound-free cross-section that can be interpolated over temperature and wavelength. However, in the short-term the absorption_coef_bf_TOPBase function (which calls weighted_bf_cross_section_TOPBase) can be used to directly compute the absorption coefficient from the TOPbase data.

Data Sources

Currently, to call weighted_bf_cross_section_TOPBase or absorption_coef_bf_TOPBase for a given species, a table is required from TOPbase that is the result of a photoionisation Cross Section query. The query should only request data for a single species.

Precise details about the requirements of the query is provided in the docstring of each_photo_ion_subtable. Note that leading/trailing white-space can cause issues for the parsers of these tables. For examples of these tables, see the TOPbase_cross_section_He_II.txt and TOPbase_cross_section_H_I.txt files in the test/data directory.

The path to this table should be directly passed to weighted_bf_cross_section_TOPBase and absorption_coef_bf_TOPBase via the cross_sec_file keyword argument. Alternatively, the code supports searching for the table at the path $KORG_OP_DIR/<symbol>_<ion-state>.txt, where

KORG_OP_DIR is the name of an environment variable

<symbol> is replaced with the species atomic symbol

<ion-state> is replaced with the capitalized Roman numerals describing the ionization state.

Examples of the base filename (under this environment variable approach) include H_I.txt, He_I.txt, He_II.txt, or Al_III.txt.

In the future, we also plan to add support for reading in photoionization tables from the NORAD-Atomic-Data webpage.

Caveats

Be aware that there are discrepancies between the theoretical energy levels computed by the Opacity Project and empirical measurements. This can be seen by looking at the plots produced by the misc/gen_Opacity_Project_comparison_plot.jl plot. To recover accurate UV opacities, Gustafsson+ 2008 needed to correct the precise photon energies in the tables for C I, Mg I, Al I, and Si I. Before providing full support for these opacities, we may want to consider making similar corrections.

Complete API

Here are all the documented methods in Korg.

Korg.Formula — Type

Represents an atom or molecule, irespective of its charge.

Korg.Line — Method

Line(wl::F, log_gf::F, species::Species, E_lower::F, 
     gamma_rad::Union{F, Missing}=missing, gamma_stark::Union{F, Missing}=missing, 
     vdw::Union{F, Tuple{F, F}, Missing}, missing) where F <: Real

Construct a Line. If any of gamma_rad, gamma_stark, or vdW are missing, guess them. vdW may be log(ΓvdW) (assumed if negative), ΓvdW (assumed if 0 < vdW < 1), or packed ABO parameters (assumed if vdW > 1). It may also be passed as a Tuple, (σ, α).

Korg.PlanarAtmosphere — Method

PlanarAtmosphere(atm::ShellAtmosphere)

Construct a planar atmosphere with the data from a shell atmosphere. Mostly useful for testing.

Korg.ShellAtmosphere — Method

ShellAtmosphere(atm::PlanarAtmosphere, R)

Construct a shell atmosphere with the data from a planar atmosphere and an outer radius. Mostly useful for testing.

Korg.Species — Type

Represents an atom or molecule (a Formula) with a particular number of electrons (regardless of their configuration).

Korg.Species — Method

Species(code::AbstractString)

Parse the "species code" as it is often specified in linelists and return a the "astronomy" notation. 01.00 → "HI", 02.01 → "HeII", 02.1000 → "HeII", 0608 → "COI", etc.

To parse at compile time, use the species string macro, i.e. species"H I".

Korg._plane_parallel_approximate_transfer_integral — Method

_plane_parallel_approximate_transfer_integral(τ, m, b)

The exact solution to $\int (m\tau + b) E_2(\tau)$ d\tau$.

The exponential integral function, expint, captures the integral over the disk of the star to get the emergent astrophysical flux. You can verify it by substituting the variable of integration in the exponential integal, t, with mu=1/t.

Korg.air_to_vacuum — Method

air_to_vacuum(λ; cgs=λ<1)

Convert λ from an air to vacuum. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

Korg.all_mu_transfer_integral — Method

all_mu_transfer_integral(τ, S)

Compute exactly the solution to the transfer integral obtained be linearly interpolating the source function, S across optical depths τ, without approximating the factor of E₂(τ).

Korg.approximate_gammas — Method

approximate_gammas(wl, species, E_lower; ionization_energies=Korg.ionization_energies)

A simplified form of the Unsoeld (1955) approximation for van der Waals broadening and the Cowley 1971 approximation for Stark broadening, evaluated at 10,000 K. Used for atomic lines with no vdW and stark broadening info in the linelist.

Returns (γ_stark, log10(γ_vdW)) in Hz, where these are the per-perturber quantities. For autoionizing lines (those for which Eupper > χ), Returns 0.0 for γvdW.

In the calculation of n*², uses the approximation that $\overbar{r^2} = 5/2 {n^*}^4 / Z^2$ which neglects the dependence on the angular momentum quantum number, l, in the the form given by Warner 1967 (the earliest english work reporting the Unsoeld result).

Korg.approximate_radiative_gamma — Method

approximate_radiative_gamma(wl, log_gf)

Approximate radiate broadening parameter.

Korg.blackbody — Method

blackbody(T, λ)

The value of the Planck blackbody function for temperature T at wavelength λ.

Korg.constant_R_LSF — Method

constant_R_LSF(flux, wls, R)

Applies a gaussian line spread function the the spectrum with flux vector flux and wavelength vector wls with constant spectral resolution, $R = \lambda/\Delta\lambda.$

For the best match to data, your wavelength range should extend a couple $\Delta\lambda$ outside the region you are going to compare.

Warning

This is a naive, slow implementation. Do not use it when performance matters.

constant_R_LSF will have weird behavior if your wavelength grid is not locally linearly-spaced. It is intended to be run on a fine wavelength grid, then downsampled to the observational (or otherwise desired) grid.

Korg.contained — Method

contained(value, interval)

Returns whether value is contained by interval.

Examples

julia> contained(0.5, Interval(1.0,10.0))
false
julia> contained(5.0, Interval(1.0,10.0))
true

Korg.contained_slice — Method

contained_slice(vals, interval)

Returns a range of indices denoting the elements of vals (which are assumed to be sorted in increasing order) that are contained by interval. When no entries are contained by interval, this returns (1,0) (which is a valid empty slice).

Korg.cumulative_trapezoid_rule! — Function

cumulative_trapezoid_rule!(out, xs, fs, [len=length(xs)])

Approximate the closed integral of f(x) from the first element of xs to each element of xs (up to len) with the trapezoid rule, given f values fs.

Assigns to the preallocated vector out.

Korg.doppler_width — Method

the width of the doppler-broadening profile

Korg.exponential_integral_2 — Method

exponential_integral_2(x)

Approximate second order exponential integral, E_2(x). This stiches together several series expansions to get an approximation which is accurate within 1% for all x.

Korg.get_absolute_abundances — Method

get_absolute_abundances(metallicity, A_X)

Calculate $n_X/n_\mathrm{total}$ for each element X given some specified abundances, $A(X)$. Use the metallicity [$X$/H] to calculate those remaining from the solar values (except He).

Korg.get_mass — Method

get_mass(f::Formula)

Returns the mass [g] of f.

Korg.hydrogen_line_absorption! — Method

hydrogen_line_absorption(λs, T, nₑ, nH_I, UH_I, hline_stark_profiles, ξ; 
                              stark_window_size=3e-7, self_window_size=1e-6)

Calculate contribution to the the absorption coefficient, α, from hydrogen lines in units of cm^-1, at wavelengths λs. For Halpha, Hbeta, and Hgamma, add the p-d approximated profiles from Barklem, Piskunovet, and O'Mara 2000 to the absortion coefficient. This "convolution by summation" is inexact, but true convolution is expensive.

Arguments:

T: temperature [K]

nₑ: electron number density [cm^-3]

nH_I: neutral hydrogen number density [cm^-3]

UH_I: the value of the neutral hydrogen partition function

hline_stark_profiles: (returned by setuphlinestark_profiles`)

ξ: microturbulent velocity [cm/s]. This is only applied to Hα-Hγ. Other hydrogen lines profiles are dominated by stark broadening, and the stark broadened profiles are pre-convolved with a doppler profile.

stark_window_size: the max distance from each line center [cm] at which to calculate the stark broadening profile

self_window_size: the max distance from each line center [cm] at which to calculate the line absorption for Hα-Hγ (those dominated by self-broadening).

Korg.inverse_gaussian_density — Method

inverse_gaussian_density(ρ, σ)

Calculate the inverse of a (0-centered) Gaussian PDF with standard deviation σ, i.e. the value of x for which ρ = exp(-0.5 x^2/σ^2}) / √[2π], which is given by σ √[-2 log (√[2π]σρ)]. Returns 0 when ρ is larger than any value taken on by the PDF.

See also: inverse_lorentz_density.

Korg.inverse_lorentz_density — Method

inverse_lorentz_density(ρ, γ)

Calculate the inverse of a (0-centered) Lorentz PDF with width γ, i.e. the value of x for which ρ = 1 / (π γ (1 + x^2/γ^2)), which is given by √[γ/(πρ) - γ^2]. Returns 0 when ρ is larger than any value taken on by the PDF.

See also: inverse_gaussian_density.

Korg.ismolecule — Method

ismolecule(f::Formula)

true when f is composed of more than one atom

Korg.line_absorption! — Method

line_absorption(linelist, λs, temp, nₑ, n_densities, partition_fns, ξ
               ; α_cntm=nothing, cutoff_threshold=1e-3, window_size=20.0*1e-8)

Calculate the opacity coefficient, α, in units of cm^-1 from all lines in linelist, at wavelengths λs [cm^-1].

other arguments:

temp the temerature in K

n_densities, a Dict mapping species to absolute number density [cm^-3].

partition_fns, a Dict containing the partition function of each species

ξ is the microturbulent velocity in cm/s (n.b. NOT km/s)

The three keyword arguments specify how the size of the window over which each line is calculated is chosen.

If α_cntm is passed as a callable returning the continuum opacity as a function of wavelength, the line window will extend to 4 doppler widths or the wavelength at which the Lorentz wings of the line are at cutoff_threshold * α_cntm[line.wl], whichever is greater. cutoff_threshold defaults to 1e-3.

if α_cntm is not passed, defaults to window_size, which is 2e-7 (in cm, i.e. 20 Å) unless otherwise specified

Korg.line_profile — Method

line_profile(λ₀, invΔλ_D, Δλ_L, line_amplitude, λ)

A voigt profile centered on λ₀ with Doppler width 1/invΔλ_D and Lorentz width Δλ_L evaluated at λ (cm). Note that this returns values in units of cm^-1.

Korg.molecular_equilibrium — Method

molecular_equilibrium(MEQS, T, nₜ, nₑ; x0)

Iteratively solve for the number density of each species. Returns a Dict mapping species to number densities.

arguments:

the system of molecular equilibrium equations MEQs (the thing returned by

molecular_equilibrium_equations)

the temperature T,

the number density of non-electron particles nₜ

the electron number density nₑ

optionally, x0, a starting point for the solver

Korg.molecular_equilibrium_equations — Method

molecular_equilibrium_equations(absolute_abundances, ironization_energies, partiation_fns, equilibrium_constants)

Returns a NamedTuple representing the system of equations specifying molecular equilibrium.

arguments:

A Dict of absolute_abundances, NX/Ntotal

a Dict of ionization energies, ionization_energies. The keys of act as a list of all atoms.

a Dict of partition functions, partition_fns

a Dict of log molecular equilibrium constants, equilibrium_constants, in partial pressure form. The keys of equilibrium_constants act as a list of all molecules.

The system of equations is specified with the number densities of the neutral atoms as free parameters. Each equation specifies the conservation of a particular species, e.g. (simplified)

n(O) = n(CO) + n(OH) + n(O I) + n(O II) + n(O III).

In this equation:

n(O), the number density of oxygen atoms in any form comes absolute_abundances and the total

number density (supplied later)

n(O I) is a free parameter. The numerical solver is varying this to satisfy the system of equations.

n(O II), and n(O III) come from the Saha (ionization) equation given n(O I)

n(CO) and n(OH) come from the molecular equilibrium constants K, which are precomputed over a range of temperatures.

Equilibrium constants are defined in terms of partial pressures, so e.g.

K(OH)  ==  (p(O) p(H)) / p(OH)  ==  (n(O) n(H)) / n(OH)) kT

This could also be computed via a Saha equation involving the molecular partition function and dissolution energy.

Korg.parse_fwf — Method

parse_fwf(source, rowspec; datarow, lastrow)

Parse fixed-width files, returning a vector of NamedTuples. source can be either an AbstractVector of Strings or a filename. rowspec should be a vector of tuples each containing, in order

a range object corresponding to the column postion

the type. Strings will be pulled directly & other types will be parsed with parse.

the column name, a Symbol.

optionally, a function to apply to the parsed value.

Korg.planar_transfer — Method

planar_transfer(α, S, τ_ref, α_ref)

Returns the astrophysical flux. See radiative_transfer for an explantion of the arguments.

Korg.radiative_transfer — Function

radiative_transfer(atm::ModelAtmosphere, α, S, α_ref, mu_grid)

Returns the astrophysical flux at each wavelength.

inputs:

atm: the model atmosphere.

α: a matrix (atmospheric layers × wavelengths) containing the absoprtion coefficient

S: the source fuction as a matrix of the same shape.

α_ref: the continuum absoprtion coefficient at the reference wavelength (5000 Å). Used to rescale the total absorption to match the model atmosphere. This value should be calculated by Korg.

mu_grid: (required if atm is a ShellAtmosphere) the values of μ at which to calculate the surface intensity, which is integrated to obtain the astrophysical flux.

Korg.ray_transfer_integral — Method

ray_transfer_integral(τ, S)

Compute exactly the solution to the transfer integral obtained be linearly interpolating the source function, S across optical depths τ, without approximating the factor of exp(-τ).

This breaks the integral into the sum of integrals of the form $\int (m\tau + b) \exp(-\tau)$ d\tau$ , which is equal to $ -\exp(-\tau) (m*\tau + b + m)$.

Korg.read_linelist — Method

read_linelist(filename; format="vald", isotopic_abundances=Korg.isotopic_abundances)

Parse a linelist file, returning a vector of Lines.

Pass format="kurucz" for a Kurucz linelist (format=kurucz_vac if it uses vacuum wavelengths; Be warned that Korg will not assume that wavelengths are vacuum below 2000 Å),format="vald" for a VALD linelist, and format="moog" for a MOOG linelist (doesn't support broadening parameters or dissociation energies).

VALD linelists (the default and preferred format) can be either "short" or "long" format, "extract all" or "extract stellar". Air wavelengths will automatically be converted into vacuum wavelengths, and energy levels will be automatically converted from cm$^{-1}$ to eV.

When they are not pre-scaled by isotopic abundace (which VALD does by default), Korg will automatically adjust the log_gf of each line according to isotopic_abundances, which defaults to the values from NIST. To use custom isotopic abundances, just pass isotopic_abundances as a dictionary mapping (atomic number, atomic weight) pairs to abundances between 0 and 1.

Korg.read_model_atmosphere — Method

read_model_atmosphere(filename; truncate_at_10000K=true)

Parse the provided model atmosphere file in MARCS ".mod" format. Returns either a PlanarAtmosphere or a ShellAtmosphere.

When truncate_at_10000K is true, layers with temperatures greater than 10,000 Kelvin will be elided. These are typically at the deepest level, where optical depth is very large, and have minimal impact on the surface spectrum. They are ignored by default because Korg's default partition and molecular equillibrium functions are tabulated only up to that tempurature.

Korg.read_partition_funcs — Method

functon read_atomic_partition_funcs([transform=identity])

Constructs a Dict holding tables containing partition function or equilibrium constant values across ln(temperature).

The optional argument, transform is applied to each value. It is used to perform unit conversions, when loading equilibrium constants

Korg.rectify — Method

rectify(flux, wls; bandwidth=50, q=0.95, wl_step=1.0)

Rectify the spectrum with flux vector flux and wavelengths wls by dividing out a moving q-quantile with window size bandwidth. wl_step controls the size of the grid that the moving quantile is calculated on and interpolated from. Setting wl_step to 0 results in the exact calculation with no interpolation, but note that this is very slow.

Experiments on real spectra show an agreement between the interpolated rectified spectrum and the "exact" one (with default values) at the 3 × 10^-4 level.

Korg.saha_ion_weights — Method

saha_ion_weights(T, nₑ, atom, ionization_energies, partition_functions)

Returns (wII, wIII), where wII is the ratio of singly ionized to neutral atoms of a given element, and wIII is the ration of doubly ionized to neutral atoms.

arguments:

temperature T [K]

electron number density nₑ [cm^-3]

atom, the atomic number of the element

ionization_energies is a collection indexed by integers (e.g. a Vector) mappping elements' atomic numbers to their first three ionization energies

partition_funcs is a Dict mapping species to their partition functions

Korg.scaled_stark — Method

the stark broadening gamma scaled acording to its temperature dependence

Korg.scaled_vdW — Function

the vdW broadening gamma scaled acording to its temperature dependence, using either simple scaling or ABO

Korg.setup_equilibrium_constants — Method

setup_equilibrium_constants()

Returns a Dict holding the default (Barklem & Collet 2016) log equilibrium constants, which are in pressure units.

In equation 7 in this paper, m is the reduced mass, m₁m₂/(m₁ + m₂). This can be verified by re-deriving the equilibrium constants from the partition functions and dissolution constants provided by the paper.

Korg.setup_hydrogen_stark_profiles — Function

load_hydrogen_stark_profiles()

Load the (doppler-convolved) Stark-broadening profiles from disk.

Korg.setup_ionization_energies — Function

setup_ionization_energies([filename])

Parses the table of ionization energies and returns it as a dictionary mapping elements to their ionization energies, [χ₁, χ₂, χ₃] in eV.

Korg.setup_partition_funcs — Function

setup_partition_funcs()

Returns a Dict holding the default (Barklem & Collet 2016) partition functions.

The partition function is linearly interpolated from values tabulated for the paper. The partition functions can only be evaluated at temperatures between 1e-5 K and 1e4 K. The tabulated values were NOT computed with a cutoff principle quantum number. This can lead to a couple percent error for some neutral species in solar type stars (I think the problem becomes worse in hotter stars). See the paper for more details.

Note that the paper seems to suggest that we could actually use cubic interpolation. That should be revisited in the future.

Korg.sigma_line — Method

sigma_line(wl)

The cross-section (divided by gf) at wavelength wl in Ångstroms of a transition for which the product of the degeneracy and oscillator strength is 10^log_gf.

Korg.spherical_transfer — Method

spherical_transfer(α, S, τ_ref, α_ref, radii, μ_surface_grid)

Perform radiative transfer along rays emerging at the μ values in μ_surface_grid in a spherically symmetric atmosphere resolved at radii radii [cm]. See radiative_transfer for an explantion of the arguments. Note that radii should be in decreasing order.

Returns (flux, intensity), where flux is the astrophysical flux, and intensity, a matrix of shape (mu values × wavelengths), is the surface intensity as a function of μ.

Korg.synthesize — Function

synthesize(atm, linelist, λ_start, λ_stop, [λ_step=0.01]; metallicity=0, abundances=Dict(), vmic=0, ... )

Solve the transfer equation in the model atmosphere atm with the transitions in linelist at the from λ_start to λ_stop in steps of λ_step (all Å) to get the resultant astrophysical flux at each wavelength. An AbstractRange can also be probvided directly in place of λ_start, λ_stop, and λ_step.

Returns a named tuple with keys:

flux: the output spectrum

alpha: the linear absorption coefficient at each wavelenth and atmospheric layer a Matrix of size (layers x wavelengths)

number_densities: A dictionary mapping Species to vectors of number densities at each atmospheric layer

wavelengths: The vacuum wavelenths (in Å) over which the synthesis was performed. If air_wavelengths=true this will not be the same as the input wavelenths.

Optional arguments:

metallicity, i.e. [metals/H] is log_10 solar relative

abundances is a Dict mapping atomic symbols to $A(X)$ format abundances, i.e. $A(x) = \log_{10}(n_X/n_\mathrm{H}) + 12$, where $n_X$ is the number density of $X$. These override metallicity.

vmic (default: 0) is the microturbulent velocity, $\xi$, in km/s.

air_wavelengths (default: false): Whether or not the input wavelengths are air wavelenths to be converted to vacuum wavelengths by Korg. The conversion will not be exact, so that the wavelenth range can internally be represented by an evenly-spaced range. If the approximation error is greater than wavelength_conversion_warn_threshold, an error will be thrown. (To do wavelength conversions yourself, see air_to_vacuum and vacuum_to_air.)

wavelength_conversion_warn_threshold (default: 1e-4): see air_wavelengths.

line_buffer (default: 10): the farthest (in Å) any line can be from the provided wavelenth range before it is discarded. If the edge of your window is near a strong line, you may have to turn this up.

cntm_step (default 1): the distance (in Å) between point at which the continuum opacity is calculated.

hydrogen_lines (default: true): whether or not to include H lines in the synthesis.

mu_grid: the range of (surface) μ values at which to calculate the surface flux when doing transfer in spherical geometry (when atm is a ShellAtmosphere).

line_cutoff_threshold (default: 1e-3): the fraction of the continuum absorption coefficient at which line profiles are truncated. This has major performance impacts, since line absorption calculations dominate more syntheses. Turn it down for more precision at the expense of runtime. The default value should effect final spectra below the 10^-3 level.

ionization_energies, a Dict mapping Species to their first three ionization energies, defaults to Korg.ionization_energies.

partition_funcs, a Dict mapping Species to partition functions (in terms of ln(T)). Defaults to data from Barklem & Collet 2016, Korg.partition_funcs.

equilibrium_constants, a Dict mapping Species representing diatomic molecules to their molecular equilbrium constants in partial pressure form. Defaults to data from Barklem and Collet 2016, Korg.equilibrium_constants.

Korg.translational_U — Method

translational_U(m, T)

The (inverse) contribution to the partition function from the free movement of a particle. Used in the Saha equation.

arguments

m is the particle mass

T is the temperature in K

Korg.trapezoid_rule — Method

trapezoid_rule(xs, fs)

Approximate the integral f(x) with the trapezoid rule over x-values xs given f(x) values fs.

Korg.vacuum_to_air — Method

vacuum_to_air(λ; cgs=λ<1)

convert λ from a vacuum to air. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

Korg.voigt — Method

voigt(α, v)

The voigt function, $H$. Approximation from Hunger 1965.

Korg.λ_to_ν_bound — Method

λ_to_ν_bound(λ_bound)

Converts a λ Inverval (in cm) to an equivalent ν Interval (in Hz), correctly accounting for tricky floating point details at the bounds.

Korg.ContinuumAbsorption.StateID — Type

StateID(iSLP, iLV)

StateID represents a given state (electron configuration) of an ion species. The arguments are taken from the TOPBase.

Arguments

iSLP::UInt16: An integer representing the quantum atomic numbers SLπ (total spin, orbital angular momentum, and parity). The value is given by 100*(2*S+1) + 10*L + π. Note that TOPBase documentation refers to the parity quantum number as P or π, but for consistency with how TOPBase presents the data, we prefer P in our implementation code.

iLV::UInt8: Distinguishes between configurations that have the same values of iSLP. TOPBase seems to index the configuration based on the energy level of the configuration (with iLV = 1 corresponding to the lowest energy).

Notes

The combination of this struct and an ion species (or neutral element) should be enough to fully specify a given electron configuration. For example, consider He I:

iSLP = 100, iLV = 1 has the electron configuration: 1s²

iSLP = 100, iLV = 2 has the electron configuration: 1s 2s

iSLP = 100, iLV = 3 has the electron configuration: 1s 3s

iSLP = 111, iLV = 1 has the electron configuration: 1s 2p

iSLP = 111, iLV = 2 has the electron configuration: 1s 3p

Korg.ContinuumAbsorption.StateProp — Type

StateProp(ion_energy_ryd, excitation_potential_ryd, statistical_weight, electron_config)

StateProp holds data associated with different electron configurations. It's used to hold data read from the TOPBase cross-section queries and to compute the statistical weight.

Korg.ContinuumAbsorption._Hminus_bf_cross_section — Method

_Hminus_bf_cross_section(λ, ion_energy_H⁻)

Compute the H⁻ bound-free cross-section at a given wavelength (specified in Å). The cross-section has units of megabarns per H⁻ particle and does NOT include a correction for stimulated emission.

This function linearly interpolates the table provided in Wishart (1979).

Korg.ContinuumAbsorption._hydrogenic_bf_high_n_absorption — Method

_hydrogenic_bf_high_n_absorption(ν, T, Z, nmin, nsdens_div_partition, ion_energy)

Approximates the sum of absorption coefficient contributions by all bound-free transitions in which the electron originates in an energy level of nmin or larger using an integral.

In full detail, the calculated quantity is described as ∑{n'=nmin}^∞ ndens(n=n') * αν(n=n') * (1 - exp(-hν/k/T))/ρ, where α(n=n') is the absorption coefficient for the bound-free atomic absorption coefficient (uncorrected for stimulated emission). This function approximates this sum with an integral.

Arguments

ν: frequency in Hz

T: temperature in K

Z::Integer: Z is the atomic number of the ion (1 for HI)

nmin::Integer: The lowest energy level (principle quantum number) included in the calculation

nsdens_div_partition is the number density of the current species divided by the partition function.

ion_energy: the ionization energy from the ground state (in eV).

Notes

This is based on equation (5.6) from Kurucz (1970) (which give the formula for opacity). I think ρ was simply omitted from that equation.

Korg.ContinuumAbsorption._load_gauntff_table — Function

_load_gauntff_table([fname])

Returns a table of thermally-averaged free-free Gaunt factors, and the values of log₁₀(γ²) and log₁₀(u) associated with each point.

This loads the non-relativistic free-free data published by van Hoof et al. (2014).

Note: This function code could trivially be adapted to load the relativistic free-free gaunt factors published by van Hoof et al (2015).

Korg.ContinuumAbsorption._ndens_Hminus — Function

_ndens_Hminus(nH_I_div_partition, ne, T, ion_energy = _H⁻_ion_energy)

Compute the number density of H⁻ (implements eqn 5.10 of Kurucz 1970). This is an application of the saha equation where the "ground state" is H⁻ and the "first ionization state" is H I. The partition function of H⁻ is 1 at all temperatures.

Korg.ContinuumAbsorption.absorption_coef_bf_TOPBase — Method

absorption_coef_bf_TOPBase(λ, T, ndens_species, species; kwargs...)

Computes the bound-free linear absorption coefficient, α, (in units of cm⁻¹) using data from the Opacity Project (downloaded from TOPBase) and assuming LTE. The result is stored in a 2D array with a size given by (length(λ), length(T)).

Arguments

λ::AbstractVector: wavelengths (in cm) to compute α at.

T::AbstractVector: temperatures (in K) to compute α at.

ndens_species::AbstractVector: number density (in cm⁻³) of the species at each temperature (this should have the same length as T).

species::Species: the starting atom or atomic ion of the bound-free reaction.

Keyword Arguments

This function accepts all the same keyword arguments as weighted_bf_cross_section_TOPBase.

Korg.ContinuumAbsorption.bounds_checked_absorption — Method

bounds_checked_absorption(func; ν_bound, temp_bound)

Constructs a wrapped function that implements bounds checking and extrapolation

Parameters

func: a function that has a signature f(ν::Real, T::Real, args...)::Real, where ν is frequency (in Hz) and T is temperature (in K)

ν_bound::Interval: Interval of frequencies (in Hz) over which func is valid.

temp_bound::Interval: Interval of temperatures (in K) over which func is valid.

The resulting function will have this signature:

wrapped_func(ν::AbstractVector{<:Real}, T::Real, args...; kwargs...)

Wrapped Function Parameters

ν::AbstractVector{<:Real}: sorted vector of frequencies (in Hz)

T::Real: temperature (in K)

args...: function-specific arguments

For a description for the kwargs..., see Continuum Absorption Kwargs.

Korg.ContinuumAbsorption.each_photo_ion_subtable — Method

each_photo_ion_subtable(filename::AbstractString)

Create an iterable EachPhotoIonSubtable object for parsing photoionization cross-sections.

The input file is organized as a series of subtables. Each subtable provides the photoionization energies (make sure this isn't the energy of the photoionized electron) in Ryd, and the associated cross section (in megabarns) for a different state.

Each entry in the EachPhotoIonSubtable iterable provides a named tuple with the keys:

name: holds the species name

state_id: holds a StateID instance

photonenergyryd: the photoionization values (check), in Ryd, stored in an array of Float32 vals

crosssectionMegaBarn: the associated photoionization cross-sections. This array should have the same shape and datatype as photonenergyryd

The underlying file is only closed after the EachPhotoIonSubtable is garbage collected.

#Notes This implementation is not robust enough to deal with blank lines at the start or end of the file.

The table should have been produced by a TOPbase Photoionisation Cross Section query that uses the following query options:

Only a single atomic number and a single number of electrons is included in the query (other choices may yield too many results, which causes TOPBase to truncate output).

There isn't any limitation on the number the quantum numbers, the energy or the levels

The query request levels ordered in the "Level order in each series"

Korg.ContinuumAbsorption.electron_scattering — Method

electron_scattering(nₑ)

Compute the linear absorption coefficient, α, from scattering off of free electrons. This has no wavelength dependence. It assumes isotropic scattering. (See, e.g. Gray p 160.)

Arguments

nₑ::F: number density of free electrons (in cgs)

Korg.ContinuumAbsorption.gaunt_ff_vanHoof — Method

gaunt_ff_vanHoof(log_u, log_γ2)

computes the thermally averaged, non-relativistic free-free gaunt factor by interpolating the table provided by van Hoof et al. (2014).

Arguments

log_u: Equal to log₁₀(u) = log₁₀(hν/(kTₑ))

log_γ2: Equal to log₁₀(γ²) = log₁₀(RydbergZ²/(kTₑ))

Rydberg is the "infinite mass unit of energy" and Tₑ is the temperature of free electrons (for our purposes, we assume that free electrons are in thermal equilibrium with ions and neutral species).

Notes

van Hoof et al. (2014) computed the associated data table with a non-relativistic approach, which is invalid at very high temperatures. They conclude (from comparisons with a different paper) that their "results should be accurate up to electron temperatures of roughly 100 MK". This is more than adequate for stellar atmospheres. In van Hoof et al. (2015), they find that relativistic effects introduce a ∼0.75% at 100MK, for Z = 1 (when Z > 1, the change is smaller).

This function currently uses linear interpolation. However, van Hoof et al. (2014) provides an implementation of a third-order Lagrange scheme, which "reaches a relative precision better than 1.5e-4 everywhere." The C and Fortran implementations of this scheme can be found here, and are copyrighted by a BSD-style license.

Earlier variants of this function used less-accurate data from section 5.1 of Kurucz (1970) that extended over a smaller interval of data. That table was originally derived from a figure in Karsas and Latter (1961) and it's now used for testing purposes.

Korg.ContinuumAbsorption.hydrogenic_bf_absorption — Function

hydrogenic_bf_absorption(ν, T, Z, nmax_explicit_sum, nsdens_div_partition,
                         ion_energy, [integrate_high_n])

Compute the bound-free linear absorption coefficient contributed by all energy states of a Hydrogenic species

The calculation is broken into 2 parts: (i) the contributions of the lowest energy states are explicitly summed and (ii) the contributions of the higher energy states are estimated with an integral.

Arguments

ν: frequency in Hz

T: temperature in K

Z::Integer: Z is the atomic number of the species (e.g. 1 for H I or 2 for He II)

nsdens_div_partition is the total number density of the species divided by the species's partition function.

ion_energy: the ionization energy from the ground state (in eV). This can be estimated as Z²*RydbergH (RydbergH is the ionization energy of Hydrogen)

nmax_explicit_sum::Integer: The highest energy level whose absorption contribution is included in the explicit sum. The contributions from higher levels are included in the integral.

integrate_high_n::bool: When this is false, bf absorption from higher energy states are not estimated at all. Default is true.

Notes

This follows the approach described in section 5.1 of Kurucz (1970).

Korg.ContinuumAbsorption.hydrogenic_ff_absorption — Method

hydrogenic_ff_absorption(ν, T, Z, ni, ne)

computes the free-free linear absorption coefficient for a hydrogenic species

The naming convention for free-free absorption is counter-intuitive. A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). In practice, this means that ni should refer to:

the number density of H II if computing the H I free-free absorption

the number density of He III if computing the He II free-free absorption

the number density of Li IV if computing the Li III free-free absorption

Arguments

Z::Integer: the charge of the ion. For example, this is 1 for ionized H.

ni: the number density of the ion species in cm⁻³.

ne: the number density of free electrons.

ν: frequency in Hz

T: temperature in K

Note

This approach was adopted from equation 5.8 from section 5.1 of Kurucz (1970). Comparison against equation 5.18b of Rybicki & Lightman (2004), reveals that the equation in Kurucz (1970) omits the dependence on ρ. According to Rybicki & Lightman (2004) the free-free absorption coefficient (corrected for stimulated emission) is:

    α = coef * Z² *ne * ni * (1 - exp(-hplanck*ν/(kboltz*T))) * g_ff / (sqrt(T) * ν³)

Note that the g_ff is the free-free gaunt factor and coef is ∼3.7e8 (a more exact coefficient can be computed from eqn 5.18a).

With this in mind, equation 5.8 of Kurucz (1970) should actually read

    κ = ne * n(H II) * F_ν(T) * (1 - exp(-hplanck*ν/(kboltz*T))) / ρ

where Fν(T) = coef * Z² * gff / (sqrt(T) * ν³).

See gaunt_ff_vanHoof for details about where our gaunt factor data comes from. For simplicity, we enforce temperature and ν bounds constraints that don't include all of the available gaunt factor data. In practice, this should never be a concern for stellar spectroscopy.

Korg.ContinuumAbsorption.ndens_state_hydrogenic — Method

ndens_state_hydrogenic(n, nsdens_div_partition, ion_energy, T)

Calculates the LTE number density (in cm^-3) of a hydrogenic species at a given energy level.

Arguments

n::Integer: The quantum number of the state

nsdens_div_partition: The total (including all states) number density of the current

ionization species (in cm^-3) divided by the species's partition function.

ion_energy: The minimum energy (in eV) required to ionize the species (from the ground

state). This can be estimated as Z²ion_energy of hydrogen or Z²Rydberg_H.

T: Temperature

Note

This assumes that all hydrogenic species have a statistical weight of gₙ = 2*n².

This was taken from equation (5.4) of Kurucz (1970) (although this comes directly from the boltzmann equation).

Korg.ContinuumAbsorption.rayleigh — Method

rayleigh(λs, nH_I, nHe_II, nH2)

Absorption coefficient from Rayleigh scattering by neutral H, He, and H2. Formulations for H and He are via Colgan+ 2016. Formulation for H2 from Dalgarno and Williams 1962.

The Dalgarno and Williams H2 is applicable redward of 1300 Å. Since Rayleigh scattering breaks down when the particle size to wavelength ratio gets large, we that all frequencies passed to this function be equivalent to 1300 Å or greater.

The formulations for H is adapted from Lee 2005, which states that it is applicable redward of Lyman alpha. See Colgan 2016 for details on He.

Korg.ContinuumAbsorption.total_continuum_absorption — Method

total_continuum_absorption(νs, T, nₑ, number_densities, partition_funcs; error_oobounds)

The total continuum linear absoprtion coefficient, α, at many frequencies, ν.

Arguments

νs are frequencies in Hz

T is temperature in K

nₑ is the electron number density in cm^-3

number_densities is a Dict mapping each Species to its number density

partition_funcs is a Dict mapping each Species to its partition function (e.g. Korg.partition_funcs)

error_oobounds::Bool specifies the behavior of most continnum absorption sources when passed frequencies or temperature values that are out of bounds for their implementation. When false (the default), those absorption sources are ignored at those values. Otherwise, an error is thrown.

Note

For efficiency reasons, νs must be sorted. While this function technically supports any sorted AbstractVector, it is most effient when passed an AbstractRange.

Korg.ContinuumAbsorption.weighted_bf_cross_section_TOPBase — Method

weighted_bf_cross_section_TOPBase(λ_vals, T_vals, species; extrapolation_bc = 0.0
                                  cross_sec_file = nothing,
                                  partition_func = partition_funcs[species])

Calculate the combined bound-free cross section (in cm²) of an ion species using data from the Opacity Project. The result is the mean of the the bound-free cross section of each electron state, weighted by their Boltzmann probabilities. It includes the LTE correction for stimulated emission. The result is stored in a 2D array with a size given by (length(λ_vals), length(T_vals)).

Arguments

λ_vals::AbstractVector: wavelengths (in cm) to compute the cross sections at

T_vals::AbstractVector: temperatures (in K) to compute the cross sections at

species::Species: the starting atom or atomic ion of the bound-free reaction.

Keyword Arguments

extrapolation_bc: Specifies handling of extrapolation during linear interpolation (this is passed to Interpolations.LinearInterpolation).

cross_sec_file: Optional keyword that specifies the path to the text file that holds the cross-section subtables. When nothing (the default), this function tries to load the file from $KORG_OP_DIR/<atomic-symbol>_<ion-state>.txt. <atomic-symbol> is replaced with the atomic symbol of the species (e.g. H, He, Al) and <ion-state> is replaced with the capital Roman numeral(s) specifying the ionization state.

partition_func: Specifies the partition function for the current species. This should be a callable that accepts ln(temperature) as an argument. Defaults to the built-in values.

Explanation

In more mathematical rigor, this function basically evaluates the following equation:

$\overline{\sigma_{\rm bf}}(\lambda,T) = \sum_{{\bf n}\in{\rm states}} \frac{ g_{\bf n}}{U(T)} e^{-E_{\bf n}/(k_B T)}\left(1 - e^{-h c/(\lambda k_B T)}\right) \sigma_{{\rm bf}, {\bf n}}(\lambda)$

In this equation:

$g_{\bf n}$ is the statistical weight of state ${\bf n}$.

$U$ is the partition function.

$E_{\bf n}$ is the excitation potential of state ${\bf n}$. In other words, its the energy difference between state ${\bf n}$ and the ground state.

$\sigma_{\lambda,{\rm bf}, {\bf n}}$ is the bound-free cross-section (a.k.a. photo-ionization cross section) of state ${\bf n}$.

The linear aborption coefficient for bound-free absorption is simply: $\alpha_{\lambda,{\rm bf}} = n_{\rm tot} \overline{\sigma_{\rm bf}}(\lambda,T)$.

Notes

In the future, it needs to be possible to pass an argument that adjusts the energy levels of different electron configurations.