Developer documentation

This page contains info for people interested in contributing code to Korg. If you have questions, do not hesitate to ask.

Julia development

If you've never developed a package in Julia, here are some tips.

Once you have a local copy of Korg, you can make it importable locally by deving it from the Julia environment in which you would like to run it: Pkg.dev("/path/to/Korg"). If you run Pkg.dev("Korg") instead, Julia will automatically clone the repo to ~/.julia/dev/Korg. (See the Pkg documentation for details.)

When working on your local copy of Korg, Revise is easiest way to make and test changes without constantly restarting your Julia session.

To run the test suite locally, start a Julia session in the Korg root directory and run ]activate . then test. This will automatically run test/runtests.jl in the test environment.

Documentation is generated with Documentor. Continuous integration on github will ensure that the documentation is generated without errors, but it won't catch all formatting problems. If you wish to generate documentation locally in order to check that everything is as expected, start a Julia session in Korg/docs, activate the test environment (]activate .), and run instantiate. This downloads and installs the docs dependences, and will only have to be run once. To generate documentation, run julia --project make.jl on the command line (the --project flag activates the local environment). The generated docs can be served from docs/build.

Code guidelines

Try to be explicit about units throughout the code, particularly when not using CGS.

Whenever possible, calculations should be precise up to a factor of $10^{-3}$. When it's easy and inexpensive, they should be precise to $10^{-5}$ or better. There are exceptions to this (e.g. the Voigt function), but ideally they will eventually be vanquished.

Ensure types are generic enough to support dual numbers and autodifferentiation.

Limit lines to 100 characters.

Use ASCII characters in the names of functions that are part of the public API.

Unless they will never be called elsewhere, provide docstrings describing the inputs, assumptions and outputs of any functions you write.

Continuum absorption

Steps for implementing new continuum sources of absorption:

Define a helper function that computes a single absorption coefficient (in units of cm⁻²). The function should accept ν (in Hz) and T (in K) as the first and second arguments, respectively. The convention is for it should share a name with the corresponding public function, but have an underscore pre-appended (e.g. we define _H_I_bf to help implement H_I_bf).

The public function is the function constructed and returned by Korg.ContinuumAbsorption.bounds_checked_absorption that wraps the above helper function. This wrapper function implements bounds-checking for ν and T and supports the keyword arguments described in Continuum Absorption Kwargs.

Add a docstring describing the new function. At the very least, please describe any non-standard arguments and include a reference in the docstring to the source where the function was taken from.

Add a line to doc/src/API.md under the Continuum absorption heading to render the docstring of your new function.

Add a line to total_continuum_absorption that calls the new public function for absorption.

The first two steps may not apply for sources that don't directly depend on ν and T (e.g. absorption from scattering).

Manifest.toml and Project.toml

Read more about these files here. These files are used by the julia package manager. Project.toml records dependencies, and you'll notice that the test and docs directories have their own Project.tomls for test and documentation-specific dependencies. Manifest.toml records exact package versions used when a package was run. It enables someone else to reproduce the exact environment and results later. There are a few directories containing scripts that generate Korg's data files which have there own Projects.tomls and Manifest.tomls, for example data/bf_cross-sections/.

Where to put data

If you are adding data to Korg, data/README provides an overview of the options and how to decide between them.

Complete API

Here are all the documented methods in Korg.

Korg.Formula — Type

Represents an atom or molecule, irespective of its charge.

Korg.Line — Method

Line(wl::F, log_gf::F, species::Species, E_lower::F, 
     gamma_rad::Union{F, Missing}=missing, gamma_stark::Union{F, Missing}=missing, 
     vdw::Union{F, Tuple{F, F}, Missing}, missing) where F <: Real

Arguments:

wl: wavelength, in cm

log_gf: (log base 10) oscillator strength (unitless)

species: the Species associated with the line

E_lower: The energy (excitiation potential) of the lower energy level (eV)

Optional Arguments (these override default recipes):

gamma_rad: Fundemental width

gamma_stark: Stark broadening width at 10,000 K (s⁻¹)

vdW: If this is present, it may may be log(Γ_vdW) (assumed if negative) or the ABO parameters as packed float or a Tuple, (σ, α).

Note the the "gamma" values here are FWHM, not HWHM, of the Lorenztian component of the line profile, and are in units of s⁻¹.

Korg.PlanarAtmosphere — Method

PlanarAtmosphere(atm::ShellAtmosphere)

Construct a planar atmosphere with the data from a shell atmosphere. Mostly useful for testing.

Korg.ShellAtmosphere — Method

ShellAtmosphere(atm::PlanarAtmosphere, R)

Construct a shell atmosphere with the data from a planar atmosphere and an outer radius. Mostly useful for testing.

Korg.Species — Type

Represents an atom or molecule (a Formula) with a particular number of electrons (regardless of their configuration).

Korg.Species — Method

Species(code::AbstractString)

Parse the "species code" in many of the forms in which it is often specifieds and return an object representing the sepcies.

Examples

"H I" -> H I

"H 1" -> H I

"H 1" -> H I

"H_1" -> H I

"H.I" -> H I

"H 2" -> H II

"H2" -> H₂

"H" -> H I

"01.00" → H I

"02.01" → He II

"02.1000" → He II

"0608" → CO I

Note

To parse at compile time, use the species string macro, i.e. species"H I". This is important in hot inner loops.

Korg._get_multi_X_H — Method

Given a vector of abundances, A_X, get [I+J+K/H], where Zs = [I,J,K] is a vector of atomic numbers. This is used to calculate, for example, [α/H] and [metals/H].

Korg._korg_data_dir — Method

Returns the local where Korg's large (too big for git) data files are stored. At present, this is only the model atmosphere archive used by interpolate_marcs.

Korg.air_to_vacuum — Method

air_to_vacuum(λ; cgs=λ<1)

Convert λ from an air to vacuum. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

See also: vacuum_to_air.

Korg.all_atomic_species — Method

all_atomic_species()

Returns an iterator that runs over all atomic species supported by Korg.

Korg.approximate_gammas — Method

approximate_gammas(wl, species, E_lower; ionization_energies=Korg.ionization_energies)

A simplified form of the Unsoeld (1955) approximation for van der Waals broadening and the Cowley 1971 approximation for Stark broadening, evaluated at 10,000 K. Used for atomic lines with no vdW and stark broadening info in the linelist.

Returns (γ_stark, log10(γ_vdW)) in Hz, where these are the per-perturber quantities. For autoionizing lines (those for which Eupper > χ), Returns 0.0 for γvdW. Note the the "gamma" values here are FWHM, not HWHM, of the Lorenztian component of the line profile.

In the calculation of n*², uses the approximation that $\overbar{r^2} = 5/2 {n^*}^4 / Z^2$ which neglects the dependence on the angular momentum quantum number, l, in the the form given by Warner 1967 (the earliest english work reporting the Unsoeld result).

Korg.approximate_radiative_gamma — Method

approximate_radiative_gamma(wl, log_gf)

Approximate radiate broadening parameter. When using this, make sure that log_gf is the true value (not adjusted for isotopic abundance).

Korg.blackbody — Method

blackbody(T, λ)

The value of the Planck blackbody function for temperature T at wavelength λ [cm].

Korg.chemical_equilibrium — Method

chemical_equilibrium(T, nₜ, nₑ, absolute_abundances, ionization_energies, 
                     partition_fns, log_equilibrium_constants; x0=nothing)

Iteratively solve for the number density of each species. Returns a Dict mapping species to number densities.

arguments:

the temperature, T, in K

the number density of non-electron particles nₜ

the electron number density nₑ

A Dict of absolute_abundances, NX/Ntotal

a Dict of ionization energies, ionization_energies. The keys of act as a list of all atoms.

a Dict of partition functions, partition_fns

a Dict of log molecular equilibrium constants, log_equilibrium_constants, in partial pressure form. The keys of equilibrium_constants act as a list of all molecules.

keyword arguments:

x0 (default: nothing) is an initial guess for the solution (in the format internal to chemical_equilibrium). If not supplied, a good guess is computed by neglecting molecules.

electron_number_density_warn_threshold (default: 0.25) is the fractional difference between the calculated electron number density and the model atmosphere electron number density at which a warning is issued.

The system of equations is specified with the number densities of the neutral atoms as free parameters. Each equation specifies the conservation of a particular species, e.g. (simplified)

n(O) = n(CO) + n(OH) + n(O I) + n(O II) + n(O III).

In this equation:

n(O), the number density of oxygen atoms in any form comes absolute_abundances and the total

number density (supplied later)

n(O I) is a free parameter. The numerical solver is varying this to satisfy the system of equations.

n(O II), and n(O III) come from the Saha (ionization) equation given n(O I)

n(CO) and n(OH) come from the molecular equilibrium constants K, which are precomputed over a range of temperatures.

Equilibrium constants are defined in terms of partial pressures, so e.g.

K(OH)  ==  (p(O) p(H)) / p(OH)  ==  (n(O) n(H)) / n(OH)) kT

Korg.constant_R_LSF — Method

constant_R_LSF(flux, wls, R; window_size=3)

Applies a gaussian line spread function the the spectrum with flux vector flux and wavelength vector wls with constant spectral resolution, $R = \lambda/\Delta\lambda$, where $\Delta\lambda$ is the LSF FWHM. The window_size argument specifies how far out to extend the convolution kernel in standard deviations.

For the best match to data, your wavelength range should extend a couple $\Delta\lambda$ outside the region you are going to compare.

Warning

This is a naive, slow implementation. Do not use it when performance matters.

constant_R_LSF will have weird behavior if your wavelength grid is not locally linearly-spaced. It is intended to be run on a fine wavelength grid, then downsampled to the observational (or otherwise desired) grid.

Korg.contained — Method

contained(value, interval)

Returns whether value is contained by interval.

Examples

julia> contained(0.5, Interval(1.0,10.0))
false
julia> contained(5.0, Interval(1.0,10.0))
true

Korg.contained_slice — Method

contained_slice(vals, interval)

Returns a range of indices denoting the elements of vals (which are assumed to be sorted in increasing order) that are contained by interval. When no entries are contained by interval, this returns (1,0) (which is a valid empty slice).

Korg.doppler_width — Method

doppler_width(λ₀ T, m, ξ)

The standard deviation of of the doppler-broadening profile. In standard spectroscopy texts, the Doppler width often refers to σ√2, but this is σ

Korg.download_atmosphere_archive — Function

download_atmosphere_archive()

Download the data used by interpolate_marcs, a repacked version of the MARCS SDSS grid. By default, the archive is stored at .korg/SDSS_MARCS_atmospheres.h5. This location can be set with the KORG_DATA_DIR environment variable.

Korg.format_A_X — Method

format_A_X(default_metals_H, default_alpha_H, abundances; kwargs... )

Returns a 92 element vector containing abundances in $A(X)$ ($\log_{10}(X/H) + 12$) format for elements from hydrogen to uranium.

Arguments

You can specify abundance with these positional arguments. All are optional, but if default_alpha_H is provided, default_metals_H must be as well.

default_metals_H (default: 0), i.e. [metals/H] is the $\log_{10}$ solar-relative abundance of elements heavier than He. It is overriden by default_alpha and abundances on a per-element basis.

default_alpha_H (default: same as default_metals_H), i.e. [alpha/H] is the $\log_{10}$ solar-relative abundance of the alpha elements (defined to be C, O, Ne, Mg, Si, S, Ar, Ca, and Ti). It is overriden by abundances on a per-element basis.

abundances is a Dict mapping atomic numbers or symbols to [$X$/H] abundances. (Set solar_relative=false to use $A(X)$ abundances instead.) These override default_metals_H. This is the only way to specify an abundance of He that is non-solar.

Keyword arguments

solar_relative (default: true): When true, interpret abundances as being in [$X$/H] ($\log_{10}$ solar-relative) format. When false, interpret them as $A(X)$ abundances, i.e. $A(x) = \log_{10}(n_X/n_\mathrm{H}) + 12$, where $n_X$ is the number density of $X$. Note that abundances not specified default to the solar value still depend on the solar value, as they are set according to default_metals_H and default_alpha_H.

solar_abundances (default: Korg.asplund_2020_solar_abundances) is the set of solar abundances to use, as a vector indexed by atomic number. Korg.asplund_2009_solar_abundances and Korg.grevesse_2007_solar_abundances are also provided for convienience.

Korg.get_alpha_H — Method

get_alpha_H(A_X)

Calculate [α/H] given a vector, A_X of absolute abundances, $A(X) = \log_{10}(n_α/n_\mathrm{H})$. Here, the alpha elements are defined to be O, Ne, Mg, Si, S, Ar, Ca, Ti. See also get_alpha_H.

Korg.get_atmosphere_archive — Method

Returns the model atmosphere grid, first loading it into memory if necessary. It's not loaded at import time because it's large, and model atmosphere interpolation isn't always needed. This is called automatically when running interpolate_marcs.

Korg.get_atoms — Method

get_atoms(x)

Returns an array view containing the atomic number of each atom that makes up the formula or species x. E.g. get_atoms(Korg.species"H2O") yields [1, 1, 8].

Korg.get_electron_number_densities — Method

electronnumberdensities(atm::ModelAtmosphere) = [l.electronnumberdensity for l in atm.layers]

This is a convienince functions for making plots, etc. Note that it doesn't access quantities in a memory-efficient order.

Korg.get_log_nK — Method

get_inv_nK(mol, T, log_equilibrium_constants)

Given a molecule, mol, a temperature, T, and a dictionary of log equilbrium constants in partial pressure form, return the inverse equilibrium constant in number density form, i.e. 1/nK where nK = n(A)n(B)/n(AB).

Korg.get_mass — Method

get_mass(f::Formula)

Returns the mass [g] of f.

Korg.get_metals_H — Method

get_metals_H(A_X)

Calculate [metals/H] given a vector, A_X of absolute abundances, $A(X) = \log_{10}(n_M/n_\mathrm{H})$. See also get_alpha_H.

Korg.get_number_densities — Method

number_densities(atm::ModelAtmosphere) = [l.number_density for l in atm.layers]

This is a convienince functions for making plots, etc. Note that it doesn't access quantities in a memory-efficient order.

Korg.get_tau_5000s — Method

tau_5000s(atm::ModelAtmosphere) = [l.tau_5000 for l in atm.layers]

This is a convienince functions for making plots, etc. Note that it doesn't access quantities in a memory-efficient order.

Korg.get_temps — Method

temps(atm::ModelAtmosphere) = [l.temp for l in atm.layers]

This is a convienince functions for making plots, etc. Note that it doesn't access quantities in a memory-efficient order.

Korg.get_zs — Method

zs(atm::ModelAtmosphere) = [l.z for l in atm.layers]

This is a convienince functions for making plots, etc. Note that it doesn't access quantities in a memory-efficient order.

Korg.hummer_mihalas_U_H — Method

hummer_mihalas_U_H(T, nH, nHe, ne)

!!!note This is experimental, and not used by Korg for spectral synthesis.

Calculate the partition function of neutral hydrogen using the occupation probability formalism from Hummer and Mihalas 1988. See hummer_mihalas_w for details.

Korg.hummer_mihalas_w — Method

hummer_mihalas_w(T, n, E_level, nH, nHe, ne)

Calculate the corection, w, to the occupation fraction of a hydrogen energy level using the occupation probability formalism from Hummer and Mihalas 1988, optionally with the generalization by Hubeny+ 1994. (Sometimes Daepen+ 1987 is cited instead, but H&M seems to be where the theory originated. Presumably it was delayed in publication.)

The expression for w is in equation 4.71 of H&M. K, the QM correction used in defined in equation 4.24. Note that H&M's "N"s are numbers (not number densities), and their "V" is volume. These quantities apear only in the form N/V, so we use the number densities instead.

This is based partially on Paul Barklem and Kjell Eriksson's WCALC fortran routine (part of HBOP.f), which is used by (at least) Turbospectrum and SME. As in that routine, we do consider hydrogen and helium as the relevant neutral species, and assume them to be in the ground state. All ions are assumed to have charge 1. Unlike that routine, the generalization to the formalism from Hubeny+ 1994 is turned off by default because I haven't closely checked it. The difference effects the chargedterm only, and temerature is only used when `usehubeny_generalizationis set totrue`.

Korg.hydrogen_line_absorption! — Method

hydrogen_line_absorption!(αs, λs, T, nₑ, nH_I, UH_I, ξ, window_size; kwargs...)

Calculate contribution to the the absorption coefficient, αs, from hydrogen lines in units of cm^-1, at wavelengths λs (a vector of ranges).

Uses profiles from Stehlé & Hutcheon (1999), which include Stark and Doppler broadening. For Halpha, Hbeta, and Hgamma, the p-d approximated profiles from Barklem, Piskunovet, and O'Mara 2000 are added to the absortion coefficient. This "convolution by summation" is inexact, but true convolution is expensive.

Arguments:

T: temperature [K]

nₑ: electron number density [cm^-3]

nH_I: neutral hydrogen number density [cm^-3]

UH_I: the value of the neutral hydrogen partition function

ξ: microturbulent velocity [cm/s]. This is only applied to Hα-Hγ. Other hydrogen lines profiles are dominated by stark broadening, and the stark broadened profiles are pre-convolved with a doppler profile.

window_size: the max distance from each line center [cm] at which to calculate the stark and self broadening profiles absorption for Hα-Hγ (those dominated by self-broadening).

Keyword arguments:

stark_profiles (default: Korg._hline_stark_profiles): tables from which to interpolate Stark profiles

use_MHD: whether or not to use the Mihalas-Daeppen-Hummer formalism to adjust the occupation probabilities of each hydrogen orbital for plasma effects. Default: true.

Korg.interpolate_marcs — Method

interpolate_marcs(Teff, logg, Fe_M=0, alpha_M=0, C_M=0; kwargs...)
interpolate_marcs(Teff, logg, A_X; kwargs...)

Returns a model atmosphere obtained by interpolating the MARCS SDSS atmosphere grid, which provides atmospheres for varying values of $T_\mathrm{eff}$, `\log g, [metals/H], [alpha/metals], and [C/metals]. If the A_X (a vector of abundances in the format returned by format_A_X and accepted by synthesize) is provided instead of M_H, alpha_M, and C_M, the solar-relative ratios will be reconstructed assuming Grevesse+ 2007 solar abundances.

The model atmosphere grid is a repacked version of the MARCS SDSS grid. Before you use interpolate_marcs for the first time, you will have to download the grid with by running download_atmosphere_archive(). By default, the archive is stored at .korg/SDSS_MARCS_atmospheres.h5. This location can be set with the KORG_DATA_DIR environment variable.

keyword arguments

spherical: whether or not to return a ShellAtmosphere (as opposed to a PlanarAtmosphere). By default true when logg < 3.5.

archive: The atmosphere archive to use. This is used to override the default grid for testing.

solar_abundances: (default: grevesse_2007_solar_abundances) The solar abundances to use when A_X is provided instead of M_H, alpha_M, and C_M. The default is chosen to match that of the atmosphere grid, and is probably no good reason to change it.

clamp_abundances: (default: false) allowed when specifying A_X direction. Whether or not to clamp the abundance paramerters to be within the range of the MARCS grid to avoid throwing an out of bounds error. Use with caution.

Warning

Atmosphere interpolation contributes non-negligeble error to synthesized spectra below Teff ≈ 4250 K. We do not endorse using it for science in that regime. See https://github.com/ajwheeler/Korg.jl/issues/164 for a discussion of the issue.

Korg.inverse_gaussian_density — Method

inverse_gaussian_density(ρ, σ)

Calculate the inverse of a (0-centered) Gaussian PDF with standard deviation σ, i.e. the value of x for which ρ = exp(-0.5 x^2/σ^2}) / √[2π], which is given by σ √[-2 log (√[2π]σρ)]. Returns 0 when ρ is larger than any value taken on by the PDF.

See also: inverse_lorentz_density.

Korg.inverse_lorentz_density — Method

inverse_lorentz_density(ρ, γ)

Calculate the inverse of a (0-centered) Lorentz PDF with width γ, i.e. the value of x for which ρ = 1 / (π γ (1 + x^2/γ^2)), which is given by √[γ/(πρ) - γ^2]. Returns 0 when ρ is larger than any value taken on by the PDF.

See also: inverse_gaussian_density.

Korg.ismolecule — Method

ismolecule(f::Formula)

true when f is composed of more than one atom

Korg.line_absorption! — Method

line_absorption(linelist, λs, temp, nₑ, n_densities, partition_fns, ξ
               ; α_cntm=nothing, cutoff_threshold=1e-3, window_size=20.0*1e-8)

Calculate the opacity coefficient, α, in units of cm^-1 from all lines in linelist, at wavelengths λs [cm^-1].

other arguments:

temp the temerature in K (as a vector, for multiple layers, if you like)

n_densities, a Dict mapping species to absolute number density in cm^-3 (as a vector, if temp is a vector).

partition_fns, a Dict containing the partition function of each species

ξ is the microturbulent velocity in cm/s (n.b. NOT km/s)

α_cntm is as a callable returning the continuum opacity as a function of wavelength. The window within which a line is calculated will extend to the wavelength at which the Lorentz wings or Doppler core of the line are at cutoff_threshold * α_cntm[line.wl], whichever is greater.

cuttoff_threshold (optional, default: 1e-3): see α_cntm

Korg.line_profile — Method

line_profile(λ₀, σ, γ, amplitude, λ)

A voigt profile centered on λ₀ with Doppler width σ (NOT √[2] σ, as the "Doppler width" is often defined) and Lorentz HWHM γ evaluated at λ (cm). Returns values in units of cm^-1.

Korg.load_atomic_partition_functions — Function

load_atomic_partition_functions()

Loads saved tabulated values for atomic partition functions from disk. Returns a dictionary mapping species to interpolators over log(T).

Korg.load_exomol_partition_functions — Method

load_exomol_partition_functions()

Loads the exomol partition funcstions for polyatomic molecules from the HDF5 archive. Returns a dictionary mapping species to interpolators over log(T).

Korg.move_bounds — Method

move_bounds(λs, lb0, ub0, λ₀, window_size)

Using lb0 and ub0 as initial guesses, return the indices of λs, (lb, ub) corresponding to λ₀±window_size.λs` can either be a

Vector, in which case lb and ub are used as first guesses

Range, in which case lb and ub are ignored and the bounds are computed directly

Vector of Ranges, in which case lb and ub are also ignored. In this case, the bounds returned are indices into the concatenation of the ranges.

Korg.n_atoms — Method

n_atoms(x)

The number of atoms in the Korg.Species or Korg.Formula x.

Korg.read_Barklem_Collet_table — Method

functon read_Barklem_Collet_table(;transform=identity)

Constructs a Dict holding tables containing partition function or equilibrium constant values across ln(temperature). Applies transform (which you can use to, e.g. change units) to each example.

Korg.read_linelist — Method

read_linelist(filename; format="vald", isotopic_abundances=Korg.isotopic_abundances)

Parse a linelist file, returning a vector of Lines.

The format keyword argument can be used to specify one of these linelist formats (default: "vald"):

"vald" for a VALD linelist. These can be either "short" or "long" format, "extract all" or "extract stellar". Air wavelengths will automatically be converted into vacuum wavelengths, and energy levels will be automatically converted from cm$^{-1}$ to eV.

"kurucz" for a Kurucz linelist (format=kurucz_vac if it uses vacuum wavelengths; Be warned that Korg will not assume that wavelengths are vacuum below 2000 Å),

"moog" for a MOOG linelist (doesn't support broadening parameters or dissociation energies).

"turbospectrum" for a Turbospectrum linelist in air wavelengths. Note that Korg doesn't make use of the (optional) orbital angular momentum quantum number, l, for the upper or lower levels, so it won't fall back on generic ABO recipes when the ABO parameters are not available. Korg's interpretation of the fdamp parameter is also slightly different from Turbospectrum's. See the documentation of the vdW parameter of Line for details. Korg will error if encounters an Unsoeld fudge factor, which it does not support.

"turbospectrum_vac" for a Turbospectrum linelist in vacuum wavelengths.

For VALD and Turbospectrum linelists with isotope information available, Korg will scale log gf values by isotopic abundance (unless VALD has already pre-scaled them), using isotopic abundances from NIST ([Korg.isotopicabundances]). To use custom isotopic abundances, just pass `isotopicabundances` with the same structure: a dict mapping atomic number to a dict mapping from atomic weight to abundance.

Be warned that for linelists which are pre-scaled for isotopic abundance, the estimation of radiative broadening from log(gf) is not accurate.

Korg.read_model_atmosphere — Method

read_model_atmosphere(filename)

Parse the provided model atmosphere file in MARCS ".mod" format. Returns either a PlanarAtmosphere or a ShellAtmosphere.

Warning

Korg does not yet support cool ($\lesssim$ 3500 K) stars. While it will happily parse their model atmospheres, it may crash when you feed them into synthesize.

Korg.rectify — Method

rectify(flux, wls; bandwidth=50, q=0.95, wl_step=1.0)

Rectify the spectrum with flux vector flux and wavelengths wls by dividing out a moving q-quantile with window size bandwidth. wl_step controls the size of the grid that the moving quantile is calculated on and interpolated from. Setting wl_step to 0 results in the exact calculation with no interpolation, but note that this is very slow when the wls is sampled for synthesis (~0.01 Å).

Experiments on real spectra show an agreement between the interpolated rectified spectrum and the "exact" one (with default values) at the 3 × 10^-4 level.

Warning

This function should not be applied to data with observational error, as taking a quantile will bias the rectification relative to the noiseless case. It is intended as a fast way to compute nice-looking rectified theoretical spectra.

Korg.saha_ion_weights — Method

saha_ion_weights(T, nₑ, atom, ionization_energies, partition_functions)

Returns (wII, wIII), where wII is the ratio of singly ionized to neutral atoms of a given element, and wIII is the ration of doubly ionized to neutral atoms.

arguments:

temperature T [K]

electron number density nₑ [cm^-3]

atom, the atomic number of the element

ionization_energies is a collection indexed by integers (e.g. a Vector) mappping elements' atomic numbers to their first three ionization energies

partition_funcs is a Dict mapping species to their partition functions

Korg.scaled_stark — Method

the stark broadening gamma scaled acording to its temperature dependence

Korg.scaled_vdW — Function

scaled_vdW(vdW, m, T)

The vdW broadening gamma scaled acording to its temperature dependence, using either simple scaling or ABO. See Anstee & O'Mara (1995) or https://www.astro.uu.se/~barklem/howto.html for the definition of the ABO γ.

vdW should be either γ_vdW evaluated at 10,000 K, or tuple containing the ABO params (σ, α). The species mass, m, is ignored in the former case.

Korg.setup_ionization_energies — Function

setup_ionization_energies([filename])

Parses the table of ionization energies and returns it as a dictionary mapping elements to their ionization energies, [χ₁, χ₂, χ₃] in eV.

Korg.setup_partition_funcs_and_equilibrium_constants — Method

setup_partition_funcs_and_equilibrium_constants()

Returns two dictionaries. One holding the default patition functions, and one holding the default log10 equilibrium constants.

Default partition functions

The partition functions are custom (calculated from NIST levels) for atoms, from Barklem & Collet 2016 for diatomic molecules, and from exomol for polyatomic molecules. For each molecule, we include only the most abundant isotopologue.

Note than none of these parition functions include plasma effects, e.g. via the Mihalas Hummer Daeppen occupation probability formalism. They are for isolated species. This can lead to a couple percent error for neutral alkalis and to greater errors for hydrogen in some atmospheres, particularly those of hot stars.

Default equilibrium constants

Molecules have equilibrium constants in addition to patition functions. For the diatomics, these are provided by Barklem and Collet, which extensively discusses the dissociation energies. For polyatomics, we calculate these ourselves, using atomization energies calculated from the enthalpies of formation at 0K from NIST's CCCDB.

Korg's equilbrum constants are in terms of partial pressures, since that's what Barklem and Collet provide.

Korg.sigma_line — Method

sigma_line(wl)

The cross-section (divided by gf) at wavelength wl in Ångstroms of a transition for which the product of the degeneracy and oscillator strength is 10^log_gf.

Korg.synthesize — Function

synthesize(atm, linelist, A_X, λ_start, λ_stop, [λ_step=0.01]; kwargs... )
synthesize(atm, linelist, A_X, wavelength_ranges; kwargs... )

Compute a synthetic spectrum.

Arguments

atm: the model atmosphere (see read_model_atmosphere)

linelist: A vector of [Line](@ref)s (see read_linelist)

A_X: a vector containing the A(X) abundances (log(X/H) + 12) for elements from hydrogen to uranium. (see format_A_X)

λ_start: the lower bound (in Å) of the region you wish to synthesize.

λ_stop: the upper bound (in Å) of the region you wish to synthesize.

λ_step (default: 0.01): the (approximate) step size to take (in Å).

If you provide a vector of wavelength ranges in place of λ_start and λ_stop, the spectrum will be synthesized over each range with minimal overhead. The ranges can be any Julia AbstractRange, for example: 5000:0.01:5010.

Returns

A named tuple with keys:

flux: the output spectrum

alpha: the linear absorption coefficient at each wavelenth and atmospheric layer a Matrix of size (layers x wavelengths)

number_densities: A dictionary mapping Species to vectors of number densities at each atmospheric layer

electron_number_density: the electron number density at each atmospheric layer

wavelengths: The vacuum wavelenths (in Å) over which the synthesis was performed. If air_wavelengths=true this will not be the same as the input wavelenths.

subspectra: A vector of ranges which can be used to index into flux to extract the spectrum for each range provided in wavelength_ranges. If you use the standard λ_start, λ_stop, λ_step arguments, this will be a vector containing only one range.

Example

to synthesize a spectrum between 5000 Å and 5100 Å, with all metal abundances set to 0.5 dex less than the solar value except carbon, except carbon, which we set to [C/H]=-0.25:

atm = read_model_atmosphere("path/to/atmosphere.mod")
linelist = read_linelist("path/to/linelist.vald")
A_X = format_A_X(-0.5, Dict("C" => -0.25))
solution = synthesize(atm, linelist, A_X, 5000, 5100)

Optional arguments:

vmic (default: 0) is the microturbulent velocity, $\xi$, in km/s.

air_wavelengths (default: false): Whether or not the input wavelengths are air wavelenths to be converted to vacuum wavelengths by Korg. The conversion will not be exact, so that the wavelenth range can internally be represented by an evenly-spaced range. If the approximation error is greater than wavelength_conversion_warn_threshold, an error will be thrown. (To do wavelength conversions yourself, see air_to_vacuum and vacuum_to_air.)

wavelength_conversion_warn_threshold (default: 1e-4): see air_wavelengths. (In Å.)

line_buffer (default: 10): the farthest (in Å) any line can be from the provided wavelenth range before it is discarded. If the edge of your window is near a strong line, you may have to turn this up.

cntm_step (default 1): the distance (in Å) between point at which the continuum opacity is calculated.

hydrogen_lines (default: true): whether or not to include H lines in the synthesis.

use_MHD_for_hydrogen_lines (default: true): whether or not to use the MHD occupation probability formalism for hydrogen lines. (MHD is always used for hydrogen bound-free absorption.)

hydrogen_line_window_size (default: 150): the mamximum distance (in Å) from each hydrogen line center at which to calculate its contribution to the total absorption coefficient.

n_mu_points (default: 20): the number of μ values at which to calculate the surface flux when doing transfer in spherical geometry (when atm is a ShellAtmosphere). 20 points is sufficient for accuracy at the 10^-3 level.

line_cutoff_threshold (default: 3e-4): the fraction of the continuum absorption coefficient at which line profiles are truncated. This has major performance impacts, since line absorption calculations dominate more syntheses. Turn it down for more precision at the expense of runtime. The default value should effect final spectra below the 10^-3 level.

electron_number_density_warn_threshold (default: 0.25): if the relative difference between the calculated electron number density and the input electron number density is greater than this value, a warning is printed.

ionization_energies, a Dict mapping Species to their first three ionization energies, defaults to Korg.ionization_energies.

partition_funcs, a Dict mapping Species to partition functions (in terms of ln(T)). Defaults to data from Barklem & Collet 2016, Korg.default_partition_funcs.

equilibrium_constants, a Dict mapping Species representing diatomic molecules to the base-10 log of their molecular equilbrium constants in partial pressure form. Defaults to data from Barklem and Collet 2016, Korg.default_log_equilibrium_constants.

bezier_radiative_transfer (default: false): Use the radiative transfer scheme. This is for testing purposes only.

Korg.translational_U — Method

translational_U(m, T)

The (inverse) contribution to the partition function from the free movement of a particle. Used in the Saha equation.

arguments

m is the particle mass

T is the temperature in K

Korg.vacuum_to_air — Method

vacuum_to_air(λ; cgs=λ<1)

convert λ from a vacuum to air. λ is assumed to be in Å if it is ⩾ 1, in cm otherwise. Formula from Birch and Downs (1994) via the VALD website.

See also: air_to_vacuum.

Korg.voigt_hjerting — Method

voigt_hjerting(α, v)

The Hjerting function, $H$, somtimes called the Voigt-Hjerting function. $H$ is defined as H(α, v) = ∫^∞_∞ exp(-y^2) / ((u-y)^2 + α^2) dy (see e.g. the unnumbered equation after Gray equation 11.47). It is equal to the ratio of the absorption coefficient to the value of the absorption coefficient obtained at the line center with only Doppler broadening.

If x = λ-λ₀, Δλ_D = σ√2 is the Doppler width, and Δλ_L = 4πγ is the Lorentz width,

voigt(x|Δλ_D, Δλ_L) = H(Δλ_L/(4πΔλ_D), x/Δλ_D) / (Δλ_D√π)
                    = H(γ/(σ√2), x/(σ√2)) / (σ√(2π))

Approximation from Hunger 1965.

Korg.λ_to_ν_bound — Method

λ_to_ν_bound(λ_bound)

Converts a λ Inverval (in cm) to an equivalent ν Interval (in Hz), correctly accounting for tricky floating point details at the bounds.

Korg.CubicSplines.CubicSpline — Method

CubicSpline(xs, ys; extrapolate=false)

Construct a interpolant using xs and ys as the knot coordinates. Assumes xs is sorted. Apply this object as a function to interpolate at any x value in the domain. If extrapolate is false, x values outside [xs[1], xs[end]] throw errors, if extrapolate is true, the interpolant uses flat extrapolation, i.e. it returns the extreme value.

Korg.CubicSplines.cumulative_integral! — Method

cumulative_integral!(out, A, t1, t2)

Given a curve described by the spine, A, Calculates the integral from t1 to t for all t = t1, t2, and all spline knots in between. So if t1 is A.t[1] and t2 is A.t[end], out should have the same length as A.t.

Korg.ContinuumAbsorption.H_I_bf — Method

H_I_bf(νs, T, nH, nHe, ne, invU_H; n_max_MHD=6, use_hubeny_generalization=false, 
       taper=false, use_MHD_for_Lyman=false)

The bound-free linear absorption coefficient contributed by all energy states of a neutral Hydrogen atom. Even though the Mihalas-Hummer-Daeppen (MHD) occupation probability formalism is not used in Korg when computing the hydrogen partition function, it is used here. That means that series limit jumps (e.g. the Balmer jump), are "rounded off", as photons with less than the "classical" ionization energy can ionize if the upper level is dissolved into the continuum.

Required Arguments

νs: sorted frequency vector in Hz

T: temperature in K

nH_I: the total number density of neutral Hydrogen (in cm⁻³)

nHe_I: the total number density of neutral Helium (in cm⁻³)

ne: the number density of electrons (in cm⁻³)

invU_H: The inverse of the neutral hydrogen partition function (neglecting contributions from the MHD formalism)

For n=1 through n=n_max_MHD (default: 6), the cross-sections are computed using Nahar 2021. These are modified using the MHD formalism to account for level dissolution. For larger n, the cross-sections are calculated with a simple analytic formula (see simple_hydrogen_bf_cross_section).

Because MHD level dissolution applied to the the Lyman series limit leads to inflated cross-sections in the visible, we don't use MHD for bf absorption from n=1. This can be overridden by setting use_MHD_for_Lyman=true, in which case you will also want to set taper=true, which the same tapering of the cross-section as HBOP to fix the problem.

The use_hubeny_generalization keyword argument enables the generalization of the MHD from Hubeny 1994. It is experimental and switched off by default.

Korg.ContinuumAbsorption._load_gauntff_table — Function

_load_gauntff_table([fname])

Returns a table of thermally-averaged free-free Gaunt factors, and the values of log₁₀(γ²) and log₁₀(u) associated with each point.

This loads the non-relativistic free-free data published by van Hoof et al. (2014).

Note: This function code could trivially be adapted to load the relativistic free-free gaunt factors published by van Hoof et al (2015).

Korg.ContinuumAbsorption._ndens_Hminus — Function

_ndens_Hminus(nH_I_div_partition, ne, T, ion_energy = _H⁻_ion_energy)

Compute the number density of H⁻ (implements eqn 5.10 of Kurucz 1970). This is an application of the saha equation where the "ground state" is H⁻ and the "first ionization state" is H I. The partition function of H⁻ is 1 at all temperatures.

Korg.ContinuumAbsorption.bounds_checked_absorption — Method

bounds_checked_absorption(func; ν_bound, temp_bound)

Constructs a wrapped function that implements bounds checking and extrapolation

Parameters

func: a function that has a signature f(ν::Real, T::Real, args...)::Real, where ν is frequency (in Hz) and T is temperature (in K)

ν_bound::Interval: Interval of frequencies (in Hz) over which func is valid.

temp_bound::Interval: Interval of temperatures (in K) over which func is valid.

The resulting function will have this signature:

wrapped_func(ν::AbstractVector{<:Real}, T::Real, args...; kwargs...)

Wrapped Function Parameters

ν::AbstractVector{<:Real}: sorted vector of frequencies (in Hz)

T::Real: temperature (in K)

args...: function-specific arguments

For a description for the kwargs..., see Continuum Absorption Kwargs.

Korg.ContinuumAbsorption.electron_scattering — Method

electron_scattering(nₑ)

Compute the linear absorption coefficient, α, from scattering off of free electrons. This has no wavelength dependence. It assumes isotropic scattering. (See, e.g. Gray p 160.)

Arguments

nₑ::F: number density of free electrons (in cgs)

Korg.ContinuumAbsorption.gaunt_ff_vanHoof — Method

gaunt_ff_vanHoof(log_u, log_γ2)

computes the thermally averaged, non-relativistic free-free gaunt factor by interpolating the table provided by van Hoof et al. (2014).

Arguments

log_u: Equal to log₁₀(u) = log₁₀(hν/(kTₑ))

log_γ2: Equal to log₁₀(γ²) = log₁₀(RydbergZ²/(kTₑ))

Rydberg is the "infinite mass unit of energy" and Tₑ is the temperature of free electrons (for our purposes, we assume that free electrons are in thermal equilibrium with ions and neutral species).

Notes

van Hoof et al. (2014) computed the associated data table with a non-relativistic approach, which is invalid at very high temperatures. They conclude (from comparisons with a different paper) that their "results should be accurate up to electron temperatures of roughly 100 MK". This is more than adequate for stellar atmospheres. In van Hoof et al. (2015), they find that relativistic effects introduce a ∼0.75% at 100MK, for Z = 1 (when Z > 1, the change is smaller).

This function currently uses linear interpolation. However, van Hoof et al. (2014) provides an implementation of a third-order Lagrange scheme, which "reaches a relative precision better than 1.5e-4 everywhere." The C and Fortran implementations of this scheme can be found here, and are copyrighted by a BSD-style license.

Earlier variants of this function used less-accurate data from section 5.1 of Kurucz (1970) that extended over a smaller interval of data. That table was originally derived from a figure in Karsas and Latter (1961) and it's now used for testing purposes.

Korg.ContinuumAbsorption.hydrogenic_ff_absorption — Method

hydrogenic_ff_absorption(ν, T, Z, ni, ne)

computes the free-free linear absorption coefficient for a hydrogenic species

The naming convention for free-free absorption is counter-intuitive. A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). In practice, this means that ni should refer to:

the number density of H II if computing the H I free-free absorption

the number density of He III if computing the He II free-free absorption

the number density of Li IV if computing the Li III free-free absorption

Arguments

Z::Integer: the charge of the ion. For example, this is 1 for ionized H.

ni: the number density of the ion species in cm⁻³.

ne: the number density of free electrons.

ν: frequency in Hz

T: temperature in K

Note

This approach was adopted from equation 5.8 from section 5.1 of Kurucz (1970). Comparison against equation 5.18b of Rybicki & Lightman (2004), reveals that the equation in Kurucz (1970) omits the dependence on ρ. According to Rybicki & Lightman (2004) the free-free absorption coefficient (corrected for stimulated emission) is:

    α = coef * Z² * ne * ni * (1 - exp(-hplanck*ν/(kboltz*T))) * g_ff / (sqrt(T) * ν³)

Note that the g_ff is the free-free gaunt factor and coef is ∼3.7e8 (a more exact coefficient can be computed from eqn 5.18a).

With this in mind, equation 5.8 of Kurucz (1970) should actually read

    κ = ne * n(H II) * F_ν(T) * (1 - exp(-hplanck*ν/(kboltz*T))) / ρ

where Fν(T) = coef * Z² * gff / (sqrt(T) * ν³).

See gaunt_ff_vanHoof for details about where our gaunt factor data comes from. For simplicity, we enforce temperature and ν bounds constraints that don't include all of the available gaunt factor data. In practice, this should never be a concern for stellar spectroscopy.

Korg.ContinuumAbsorption.metal_bf_absorption! — Method

metal_bf_absorption!(α, νs, T, number_densities)

Adds to α the contributions of bf metal opacities. Uses precomputed tables from TOPBase for Li, Be, B, C, N, O, F, Ne, Na, Mg, Al, Si, S, Ar, and Ca. Uses tables from NORAD for Fe. For these elements, tables have been precomputed for the neutral and singly ionized species assuming and LTE distribution of energy levels. See Korg/data/metal_bf_cross-sections/ for the scripts which generate the tables.

Cross sections were computed for 100 K < T < 100,000 K and frequencies corresponding to 500 Å < λ < 30,000 Å. Outside of either of those ranges, flat extrapolation is used (i.e. the extreme value is returned).

Korg.ContinuumAbsorption.positive_ion_ff_absorption! — Method

positive_ion_ff_absorption!(α_out::Vector{Real}, ν::Real, T::Real, number_densities::Dict, 
                            ne::Real, departure_coefficients=Peach1970.departure_coefficients)

Computes the linear absorption coefficient (in cm⁻¹) for all free-free interactions involving positively charged atomic species. Uses the provided departure coefficients when they are available, and the uncorrected hydrogenic approximation when they are not.

Arguments

ν: frequency in Hz

T: temperature in K

number_densities is a Dict mapping each Species to its number density.

ne: the number density of free electrons.

departure_coefficients (optional, defaults to ContinuumAbsorption.Peach1970.departure_coefficients: a dictionary mapping species to the departure coefficients for the ff process it participates in (e.g. species"C II" maps to the C III ff departure coefficients–see note below). Departure coefficients should be callables taking temperature and (photon energy / RydbergH / Zeff^2).

Note

A free-free interaction is named as though the species interacting with the free electron had one more bound electron (in other words it's named as though the free-electron and ion were bound together). To make matters more confusing, some sources present the free-free linear absorption coefficient's formula in terms of the number density of the species in the interaction's name (by including the Saha equation in the formula).

To compute the absorption for a given free-free interaction, this function accesses elements from the dictionary arguments that are associated with the species that participates in the interaction. For example:

Si I ff absorption: uses the number density of Si II (number_densities[species"Si II"]) and checks departure_coefficients[species"Si II"] for departure coefficients.

Si II ff absorption: uses the number density of Si III (number_densities[species"Si III"]) and checks departure_coefficients[species"Si III"] for departure coefficients.

Korg.ContinuumAbsorption.rayleigh — Method

rayleigh(λs, nH_I, nHe_II, nH2)

Absorption coefficient from Rayleigh scattering by neutral H, He, and H2. Formulations for H and He are via Colgan+ 2016. Formulation for H2 from Dalgarno and Williams 1962.

The Dalgarno and Williams H2 is applicable redward of 1300 Å. Since Rayleigh scattering breaks down when the particle size to wavelength ratio gets large, we that all frequencies passed to this function be equivalent to 1300 Å or greater.

The formulations for H is adapted from Lee 2005, which states that it is applicable redward of Lyman alpha. See Colgan 2016 for details on He.

Korg.ContinuumAbsorption.simple_hydrogen_bf_cross_section — Method

simple_hydrogen_bf_cross_section(n::Integer, ν::Real)

Calculate the H I bf cross section in megabarns using a very simple approximation. See, for example, Kurucz 1970 equation 5.5 (though see note below). This implementation is used to extrapolate the cross-section past the ionization energy of an unperturbed hydrogen atom, as is required to take level dissolution into account with the MHD formalism.

Equation 5.5 of Kurucz had a typo in it. In the numerator of the fraction that is multiplied by the entire polynomial, Z² should be Z⁴. This was discovered during comparisons with data from the Opacity Project, and can be confirmed by looking at eqn 5.6 of Kurucz (it uses Z⁴ instead of Z²) or by comparison against equation 10.54 of Rybicki & Lightman.

Korg.ContinuumAbsorption.total_continuum_absorption — Method

total_continuum_absorption(νs, T, nₑ, number_densities, partition_funcs; error_oobounds)

The total continuum linear absoprtion coefficient, α, at many frequencies, ν.

Arguments

νs are frequencies in Hz

T is temperature in K

nₑ is the electron number density in cm^-3

number_densities is a Dict mapping each Species to its number density

partition_funcs is a Dict mapping each Species to its partition function (e.g. Korg.partition_funcs)

error_oobounds::Bool specifies the behavior of most continnum absorption sources when passed frequencies or temperature values that are out of bounds for their implementation. When false (the default), those absorption sources are ignored at those values. Otherwise, an error is thrown.

Note

For efficiency reasons, νs must be sorted. While this function technically supports any sorted AbstractVector, it is most effient when passed an AbstractRange.

Korg.ContinuumAbsorption.Stancil1994 — Module

This module contains transcriptions of the tables from Stancil 1994, who calculated ff and bf absorption coefficients for H₂⁺ and He₂⁺.

It also contains pre-constructed interpolation objects.

Korg.ContinuumAbsorption.Peach1970.departure_coefficients — Method

Peach1970.departure_coefficients()

This module contains interpolators of the tabulated ff departure coeffients from Peach+ 1970, which we use to correct the hydrogenic ff absorption coefficient for H I ff, C I ff, Si I ff, and Mg I ff. It contains a dictionary (returned by departure_coefficients()), which maps Species to interpolator objects. Crucially, the dictionary is indexed by the species which actually participates in the interaction, not the one after which the interaction is named.

Outside the regime in which Peach 1970 provides data, the interpolators return 0, falling back to the hydreogenic approximation.

The species for which we use corrections are the same species which get corrected in MARCS/Turbospectrum (see Table 1 of Gustafsson+ 2008). The choices seem are largely motivated by which species have departure terms at normal stellar atmosphere conditions and which species are most abundant in the sun. For C II ff, we include only the contribution from the ¹S parent term, even though (in contrast to other speices) information is available for the ³Pᵒ term as well.

The free-free absorption coefficient (including stimulated emission) is given by:

$\alpha_{ m ff} = \alpha_{ m hydrogenic, ff}(\nu, T, n_i, n_e; Z) (1 + D(T, \sigma))$,

where

$\alpha_{ m hydrogenic, ff}(\nu, T, n_i, n_e; Z)$ should include the correction for stimulated emission.

$n_i$ is the number density fo the ion species that participates in the interation, not the species the interaction is named after.

$n_e$ is the number density of free electrons.

$D(T, \sigma)$ is specified as the departure arg, and is expected to interpolate over

the tabulated values specified in Table III of Peach (1970).

σ denotes the energy of the photon in units of RydbergH*Zeff²

It might not be immediately obvious how the above equation relates to the equations presented in Peach (1970). Peach describes the calculation for $k_ u^F$, the free-free absorption coefficient (uncorrected for stimulated emission) per particle of the species that the interaction is named after. In other words, he computes:

$k_ u^F = \alpha_{ m ff}/n_{i-1} \left(1 - e^\frac{-h\nu}{k T}\right)^{-1}$,

where $n_{i-1}$ is the number density of the species that the interaction is named after. $k_ u^F$ can directly be computed, under LTE, from just $u$, $T$, and $n_{i-1}$ (the Saha Equation relates $\alpha_{ m ff}$'s dependence on $n_e$ and $n_i$ to $n_{i-1}$ and $T$. Gray (2005) follows a similar convention when describing free-free absorption.

Warning

The tabulated data in this module was taken from Peach 1970 using OCR software, and may contain mis-read values, although they produce reasonable behavior and there are no obvious problems.

Korg.RadiativeTransfer.generate_mu_grid — Method

generate_mu_grid(n_points)

Used by both radiative transfer schemes to compute quadature over μ. Returns (μ_grid, μ_weights).

Korg.RadiativeTransfer.MoogStyleTransfer — Module

Korg's default radiative transfer implementation. See also: RadiativeTransfer.BezierTransfer

Korg.RadiativeTransfer.MoogStyleTransfer._plane_parallel_approximate_transfer_integral — Method

_plane_parallel_approximate_transfer_integral(τ, m, b)

The exact solution to $\int (m\tau + b) E_2(\tau)$ d\tau$.

The exponential integral function, expint, captures the integral over the disk of the star to get the emergent astrophysical flux. You can verify it by substituting the variable of integration in the exponential integal, t, with mu=1/t.

Korg.RadiativeTransfer.MoogStyleTransfer.all_mu_transfer_integral — Method

all_mu_transfer_integral(τ, S)

Compute exactly the solution to the transfer integral obtained be linearly interpolating the source function, S across optical depths τ, without approximating the factor of E₂(τ).

Korg.RadiativeTransfer.MoogStyleTransfer.cumulative_trapezoid_rule! — Function

cumulative_trapezoid_rule!(out, xs, fs, [len=length(xs)])

Approximate the closed integral of f(x) from the first element of xs to each element of xs (up to len) with the trapezoid rule, given f values fs.

Assigns to the preallocated vector out.

Korg.RadiativeTransfer.MoogStyleTransfer.exponential_integral_2 — Method

exponential_integral_2(x)

Approximate second order exponential integral, E_2(x). This stiches together several series expansions to get an approximation which is accurate within 1% for all x.

Korg.RadiativeTransfer.MoogStyleTransfer.planar_transfer — Method

planar_transfer(α, S, τ_ref, α_ref)

Returns the astrophysical flux. See radiative_transfer for an explantion of the arguments.

Korg.RadiativeTransfer.MoogStyleTransfer.radiative_transfer — Function

radiative_transfer(atm::ModelAtmosphere, α, S, α_ref, mu_grid)

Returns the astrophysical flux at each wavelength.

inputs:

atm: the model atmosphere.

α: a matrix (atmospheric layers × wavelengths) containing the absoprtion coefficient

S: the source fuction as a matrix of the same shape.

α_ref: the continuum absoprtion coefficient at the reference wavelength (5000 Å). Used to rescale the total absorption to match the model atmosphere. This value should be calculated by Korg.

mu_grid: (required if atm is a ShellAtmosphere) the values of μ at which to calculate the surface intensity, which is integrated to obtain the astrophysical flux.

Keyword arguments:

n_nu_points: the number of rays to use when integrating over I_surface(μ) to obtain the astrophysical flux. This doesn't do anything if atm is a PlanarAtmosphere.

Korg.RadiativeTransfer.MoogStyleTransfer.ray_transfer_integral — Method

ray_transfer_integral(τ, S)

Compute exactly the solution to the transfer integral obtained be linearly interpolating the source function, S across optical depths τ, without approximating the factor of exp(-τ).

This breaks the integral into the sum of integrals of the form $\int (m\tau + b) \exp(-\tau)$ d\tau$ , which is equal to $ -\exp(-\tau) (m*\tau + b + m)$.

Korg.RadiativeTransfer.MoogStyleTransfer.spherical_transfer — Method

spherical_transfer(α, S, τ_ref, α_ref, radii, μ_surface_grid)

Perform radiative transfer along rays emerging at the μ values in μ_surface_grid in a spherically symmetric atmosphere resolved at radii radii [cm]. See radiative_transfer for an explantion of the arguments. Note that radii should be in decreasing order.

Returns (flux, intensity), where flux is the astrophysical flux, and intensity, a matrix of shape (mu values × wavelengths), is the surface intensity as a function of μ.

Korg.RadiativeTransfer.BezierTransfer — Module

Functions to solve the radiative transfer formal solution using the de la Cruz Rodríguez and Piskunov method.

We would like to migrate to this transfer implementation as the default eventually, but it doens't seem to perform much better than RadiativeTransfer.MoogStyleTransfer` in practice. The treatment of both the path-length (ds) and what happens at the midplane of the star are more careful in this implementation, but it's not yet as fast and it doesn't produce results any closer to the MARCS flux files (even when adjusting the absorption coeffs to match the MARCS numbers).

Korg.RadiativeTransfer.BezierTransfer.compute_tau_bezier! — Method

compute_tau_bezier(τ, s, α)

Compute optical depth (write to τ) along a ray with coordinate s and absorption coefficient α. This is the method proposed in de la Cruz Rodríguez and Piskunov 2013, but the

Korg.RadiativeTransfer.BezierTransfer.compute_tau_spline_analytic! — Method

compute_tau_spline_analytic(τ, s, α)

Compute τ using a cubic spline to interpolate α. This is not used by radiative_transfer, but is included for completeness.

Korg.RadiativeTransfer.BezierTransfer.fritsch_butland_C — Method

fritsch_butland_C(x, y)

Given a set of x and y values, compute the bezier control points using the method of Fritch & Butland 1984, as suggested in de la Cruz Rodríguez and Piskunov 2013.

Korg.RadiativeTransfer.BezierTransfer.planar_transfer — Method

planar_transfer(α, S, z, n_μ_points)

Perform radiative transfer in a planar atmosphere. See radiative_transfer for an explantion of the arguments.

Returns (flux, intensity), where flux is the astrophysical flux, and intensity, a matrix of shape (mu values × wavelengths), is the surface intensity as a function of μ.

Korg.RadiativeTransfer.BezierTransfer.radiative_transfer — Method

radiative_transfer(atm::ModelAtmosphere, α, S, n_μ_points)

Returns the astrophysical flux at each wavelength.

inputs:

atm: the model atmosphere.

α: a matrix (atmospheric layers × wavelengths) containing the absoprtion coefficient

S: the source fuction as a matrix of the same shape. rescale the total absorption to match the model atmosphere. This value should be calculated by Korg.

n_μ_points: the number of quadrature points to use when integrating over I_surface(μ) to obtain the astrophysical flux.

Korg.RadiativeTransfer.BezierTransfer.ray_transfer_integral — Method

ray_transfer_integral(τ, S)

Given τ and S along a ray (at a particular wavelength), compute the intensity at the end of the ray (the surface of the star). This uses the method from de la Cruz Rodríguez and Piskunov 2013.

Korg.RadiativeTransfer.BezierTransfer.spherical_transfer — Method

spherical_transfer(α, S, τ_ref, α_ref, radii, μ_surface_grid)

Perform radiative transfer along rays emerging in a spherically symmetric atmosphere resolved at radii radii [cm]. See radiative_transfer for an explantion of the arguments. Note that radii should be in decreasing order.

Returns (flux, intensity), where flux is the astrophysical flux, and intensity, a matrix of shape (mu values × wavelengths), is the surface intensity as a function of μ.